### 2025-08-20 #8838 @danielskatz
👋 @rahildoshi97 - thanks for your submission. The editors will briefly discuss which track it should be in next.

While we do so, JOSS's recommended word count for papers is 250-1000 words. This paper is 1650 words. Please consider how you could reduce it, perhaps by moving some content to the repo/documentation and pointing to it there, and then make changes your .md file. You can use the command `@editorialbot check repository` to check a number of things, including the paper's word count.

editorialbot commands need to be the first entry in a new comment.

In addition, you could work on the possibly missing DOIs that editorialbot suggests, but note that some may be incorrect. Please feel free to make changes to your .bib file, then use the command `@editorialbot check references` to check again.

Finally, the command `@editorialbot generate pdf` will make a new PDF.
