### 2025-08-20 3acd508
Add dynamic VTK output naming
