### 2025-08-20 277fefb
Update paper to reduce word count and include missing DOIs as per JOSS recommendations
