### 2025-08-20 5ed3e94
Update paper to add missing references
