+++ paper/paper.bib
@@ -182,6 +182,15 @@
     isbn = {978-0913478080}
 }
 
+@book{callister2018materials,
+    title = {Materials Science and Engineering: An Introduction},
+    author = {Callister, William D. and Rethwisch, David G.},
+    edition = {10th},
+    year = 2018,
+    publisher = {John Wiley \& Sons},
+    isbn = {978-1119405498}
+}
+
 @book{hughes2012finite,
     title = {The Finite Element Method: Linear Static and Dynamic Finite Element Analysis},
     author = {Hughes, Thomas J. R.},
