### 2025-08-20 ea1e622
Fix invalid DOIs
