+++ paper/paper.bib
@@ -138,7 +138,7 @@
     pages = {i},
     year = {2013},
     isbn = {978-1-85617-633-0},
-    doi = {https://doi.org/10.1016/B978-1-85617-633-0.00019-8},
+    doi = {10.1016/B978-1-85617-633-0.00019-8},
     url = {https://www.sciencedirect.com/science/article/pii/B9781856176330000198}
 }
 
@@ -151,7 +151,7 @@
     pages = {i},
     year = {2014},
     isbn = {978-0-08-098205-2},
-    doi = {https://doi.org/10.1016/B978-0-08-098205-2.00011-1},
+    doi = {10.1016/B978-0-08-098205-2.00011-1},
     url = {https://www.sciencedirect.com/science/article/pii/B9780080982052000111}
 }
 
