### 2025-08-20 455916b
Add missing DOI and update paper
