+++ paper/paper.bib
@@ -196,7 +196,8 @@
     author = {Hughes, Thomas J. R.},
     year = 2012,
     publisher = {Dover Publications},
-    isbn = {978-0486411811}
+    isbn = {978-0486411811},
+    doi = {10.1016/0045-7825(87)90013-2},
 }
 
 @book{tadmor2011modeling,
