### 2025-08-21 3d5ecbc
Improve JOSS paper structure
