+++ paper/paper.md
@@ -87,16 +87,22 @@ MaterForge standardizes and simplifies the integration of realistic material beh
 
 - **Configurable Boundary Behavior**: Users can define how properties behave outside their specified ranges,
   choosing between constant-value or extrapolation to best match the physical behavior of the material.
-  The boundary behavior options work seamlessly with the regression capabilities to provide comprehensive data processing control
-  (\autoref{fig:regression_options_with_boundary_behavior_new}).
-```yaml
-    bounds: [constant, extrapolate] # Defines behavior for [lower, upper] bounds
-```
 
 - **Regression and Data Reduction**: The library integrates pwlf [@pwlf] to perform piecewise regression for large datasets.
   This simplifies complex property curves into efficient mathematical representations with configurable polynomial degrees and number of segments,
-  reducing computational overhead while maintaining physical accuracy (\autoref{fig:regression_options_with_boundary_behavior_new}).
+  reducing computational overhead while maintaining physical accuracy.
+
+- **Intelligent Simplification Timing**: MaterForge provides sophisticated control over when data simplification occurs
+  in the dependency chain via the `simplify` parameter.
+  `simplify: pre` optimizes performance by simplifying properties before they are used in dependent calculations,
+  while `simplify: post` defers simplification until all dependent properties have been computed, maximizing numerical accuracy.
+  This allows users to balance computational efficiency with numerical accuracy based on their specific simulation requirements.
+
+The boundary behavior options work seamlessly with the regression capabilities to provide comprehensive data processing control 
+(\autoref{fig:regression_options_with_boundary_behavior_new}).
+
 ```yaml
+    bounds: [constant, extrapolate] # Defines behavior for [lower, upper] bounds
     regression:      # Optional regression configuration
       simplify: pre  # 'pre' (before processing) or 'post' (after processing)
       degree: 2      # Polynomial degree for regression
@@ -104,12 +110,6 @@ MaterForge standardizes and simplifies the integration of realistic material beh
 ```
 
 ![MaterForge's data processing capabilities: regression and data reduction showing raw data (points) fitted with different polynomial degrees and segment configurations, and configurable boundary behavior options demonstrating constant versus extrapolate settings for the same density property, illustrating how MaterForge can reduce complexity while maintaining physical accuracy and providing flexible boundary control.\label{fig:regression_options_with_boundary_behavior_new}](figures/regression_options_with_boundary_behavior_new.png)
-
-- **Intelligent Simplification Timing**: MaterForge provides sophisticated control over when data simplification occurs
-  in the dependency chain via the `simplify` parameter.
-  `simplify: pre` optimizes performance by simplifying properties before they are used in dependent calculations,
-  while `simplify: post` defers simplification until all dependent properties have been computed, maximizing numerical accuracy.
-  This allows users to balance computational efficiency with numerical accuracy based on their specific simulation requirements.
 
 - **Inverse Property Computation**: The library can generate inverse piecewise-linear functions,
   enabling the determination of independent variables from known property values.
