### 2025-08-21 2241d61
Reduce word count
