+++ paper/paper.bib
@@ -15,44 +15,6 @@
     doi = {10.7717/peerj-cs.103}
 }
 
-@Article{         numpy,
-    title         = {Array programming with {NumPy}},
-    author        = {Charles R. Harris and K. Jarrod Millman and St{\'{e}}fan J.
-                 van der Walt and Ralf Gommers and Pauli Virtanen and David
-                 Cournapeau and Eric Wieser and Julian Taylor and Sebastian
-                 Berg and Nathaniel J. Smith and Robert Kern and Matti Picus
-                 and Stephan Hoyer and Marten H. van Kerkwijk and Matthew
-                 Brett and Allan Haldane and Jaime Fern{\'{a}}ndez del
-                 R{\'{i}}o and Mark Wiebe and Pearu Peterson and Pierre
-                 G{\'{e}}rard-Marchant and Kevin Sheppard and Tyler Reddy and
-                 Warren Weckesser and Hameer Abbasi and Christoph Gohlke and
-                 Travis E. Oliphant},
-    year          = {2020},
-    month         = sep,
-    journal       = {Nature},
-    volume        = {585},
-    number        = {7825},
-    pages         = {357--362},
-    doi           = {10.1038/s41586-020-2649-2},
-    publisher     = {Springer Science and Business Media {LLC}},
-    url           = {https://doi.org/10.1038/s41586-020-2649-2}
-}
-
-@Article{matplotlib,
-    Author    = {Hunter, J. D.},
-    Title     = {Matplotlib: A 2D graphics environment},
-    Journal   = {Computing in Science \& Engineering},
-    Volume    = {9},
-    Number    = {3},
-    Pages     = {90--95},
-    abstract  = {Matplotlib is a 2D graphics package used for Python for
-  application development, interactive scripting, and publication-quality
-  image generation across user interfaces and operating systems.},
-    publisher = {IEEE COMPUTER SOC},
-    doi       = {10.1109/MCSE.2007.55},
-    year      = 2007
-}
-
 @article{walberla,
     title = {waLBerla: A block-structured high-performance framework for multiphysics simulations},
     author = {Bauer, Martin and Köstler, Harald and Rüde, Ulrich},
@@ -75,13 +37,6 @@
     url = {https://doi.org/10.1145/3295500.3356186}
 }
 
-@Manual{pwlf,
-    author = {Jekel, Charles F. and Venter, Gerhard},
-    title = {{pwlf:} A Python Library for Fitting 1D Continuous Piecewise Linear Functions},
-    year = {2019},
-    url = {https://github.com/cjekel/piecewise_linear_fit_py}
-}
-
 @article{coolprop,
     author = {Bell, Ian H. and Wronski, Jorrit and Quoilin, Sylvain and Lemort, Vincent},
     title = {Pure and Pseudo-pure Fluid Thermophysical Property Evaluation and
@@ -96,7 +51,7 @@
     eprint = {http://pubs.acs.org/doi/pdf/10.1021/ie4033999}
 }
 
-@article{nist_webbook,
+@article{linstrom2001nist,
     author = {Linstrom, Peter J. and Mallard, William G.},
     title = {The NIST Chemistry WebBook:  A Chemical Data Resource on the Internet},
     journal = {Journal of Chemical \& Engineering Data},
@@ -106,18 +61,6 @@
     year = {2001},
     doi = {10.1021/je000236i},
     URL = {https://doi.org/10.1021/je000236i},
-    eprint = {https://doi.org/10.1021/je000236i},
[…]
