+++ paper/paper.md
@@ -44,7 +44,7 @@ and serves as a bridge between experimental data and numerical simulation.
 # Statement of Need
 
 Accurate numerical simulation requires accounting for material properties such as thermal conductivity, density, and viscosity
-that depend on variables like temperature, pressure, or strain rate [@lewis1996finite; @zienkiewicz2013finite].
+that depend on variables like temperature, pressure, or strain rate [@lewis1996finite].
 This challenge is compounded by the wide variation in data availability,
 from well-characterized models for established materials to sparse experimental points for novel materials.
 Consequently, property definitions can range from simple constants to complex tabular datasets or sophisticated equations,
@@ -58,42 +58,40 @@ Similarly, CALPHAD databases [@calphad] are powerful but often require proprieta
 and do not easily integrate with general-purpose simulation codes.
 
 This leads to ad hoc solutions, hindering workflow efficiency and FAIR data adoption [@wilkinson2016fair].
-MaterForge bridges this gap by providing a unified, open-source framework that leverages
-symbolic mathematics, automatic regression, and dependency resolution.
-By combining a user-friendly YAML configuration with powerful backend processing,
-MaterForge standardizes and simplifies the integration of realistic material behavior into scientific simulations.
+MaterForge bridges this gap by providing a unified, framework that leverages
+symbolic mathematics, automatic regression, and dependency resolution
+to standardizes and simplify the integration of realistic material behavior into scientific simulations.
 
 # Key Functionality
 
 - **Flexible Input Methods**: The library supports various property definition methods such as
   constant values, step functions, file-based data (.xlsx, .csv, .txt), tabular data, piecewise equations, and computed properties 
   (\autoref{fig:input_methods_new}).
-  This versatility allows users to leverage data from diverse sources with robust file processing capabilities.
+  This versatility allows users to leverage data from diverse sources.
 
 ![MaterForge's property definition methods with corresponding YAML examples and automatically generated validation plots.\label{fig:input_methods_new}](figures/input_methods_new.jpg)
 
 - **Extensible Material Support**: The framework supports any material type through its extensible architecture.
-  Currently implemented and thoroughly tested for pure metals and alloys through its unified interface,
-  its modular design allows for straightforward extension to other materials such as
+  Currently implemented for pure metals and alloys,
+  its modular design allows straightforward extension to materials such as
   ceramics, polymers, or composites.
 
 - **Automatic Dependency Resolution**: For dependent properties 
-  (e.g., thermal diffusivity calculated from thermal conductivity, density, and heat capacity),
+  (e.g., density calculated from thermal expansion coefficient),
   MaterForge automatically determines the correct processing order, resolves mathematical dependencies,
   and detects circular references.
 
 - **Configurable Boundary Behavior**: Users can define how properties behave outside their specified ranges,
   choosing between constant-value or extrapolation to best match the physical behavior of the material.
 
-- **Regression and Data Reduction**: The library performs piecewise regression for large datasets.
-  This simplifies complex property curves into efficient mathematical representations with configurable polynomial degrees and number of segments,
-  reducing computational overhead while maintaining physical accuracy.
+- **Regression and Data Reduction**: The library performs piecewise regression for large datasets, 
+  simplifying complex property curves into efficient mathematical representations with configurable polynomial degrees and segments,
+  reducing computational overhead while maintaining accuracy.
 
 - **Intelligent Simplification Timing**: MaterForge provides sophisticated control over when data simplification occurs
-  in the dependency chain via the `simplify` parameter.
[…]
