+++ paper/paper.bib
@@ -71,20 +71,6 @@
     isbn = {978-0471943617}
 }
 
-@incollection{zienkiewicz2013finite,
-    title = {The Finite Element Method: Its Basis and Fundamentals},
-    editor = {O.C. Zienkiewicz and R.L. Taylor and J.Z. Zhu},
-    booktitle = {The Finite Element Method: its Basis and Fundamentals (Seventh Edition)},
-    publisher = {Butterworth-Heinemann},
-    edition = {Seventh Edition},
-    address = {Oxford},
-    pages = {i},
-    year = {2013},
-    isbn = {978-1-85617-633-0},
-    doi = {10.1016/B978-1-85617-633-0.00019-8},
-    url = {https://www.sciencedirect.com/science/article/pii/B9781856176330000198}
-}
-
 @incollection{ashby2013materials,
     title = {Materials and Design},
     editor = {Michael Ashby and Kara Johnson},
