### 2025-08-21 533120a
Remove comments from YAML snippet
