### 2025-08-21 03e960a
Simplify python integration
