### 2025-08-21 1d45189
Reduce word count
