### 2025-08-21 bfc7840
Update comparison table
