### 2025-08-21 5a636a8
Update comparison table
