### 2025-08-21 cf351d1
Reduce word count
