### 2025-08-21 7e8574e
Fix YAML config composition
