### 2025-08-21 #8838 @danielskatz
@rahildoshi97 - We've decided the best fit for this is the Physics and Engineering track. I'm transferring it now. You should hear from the track editor in the future, but there is a little backlog there (and in the other track we considered), so please be patient.
