### 2025-08-21 e31eada
Update comparison table
