### 2025-08-21 2a084ba
Update comparison table
