### 2025-08-21 8b9a165
Update comparison table
