### 2025-08-21 b60b09f
Update comparison table
