### 2025-08-21 7640fd0
Update comparison table
