### 2025-08-21 #8838 @danielskatz
Yes, that was the other option that seemed possible, but we decided PE was slightly better for both topical and internal reasons.
