### 2025-08-21 d1dfedc
Update comparison table
