### 2025-08-21 fcf299d
Update comparison table
