### 2025-08-21 bd1cb0d
Update python integration and comparison table
