### 2025-08-25 07c47a8
Update CPU scaling code
