+++ apps/CodegenHeatEquationCPUScaling.cpp
@@ -27,6 +27,31 @@ void swapFields(StructuredBlockForest& blocks, BlockDataID uID, BlockDataID uTmp
     }
 }
 
+
+// Automatic process decomposition for any number of processes
+std::tuple<uint_t, uint_t, uint_t> calculateProcessDecomposition(uint_t numProcesses) {
+    // Find best factorization for 3D decomposition
+    uint_t procs_x = 1, procs_y = 1, procs_z = 1;
+    // Start with cube root as base
+    uint_t base = uint_c(std::round(std::cbrt(real_c(numProcesses))));
+    // Find factors close to cube root
+    for (uint_t z = base; z >= 1; --z) {
+        if (numProcesses % z == 0) {
+            uint_t remaining = numProcesses / z;
+            uint_t base_xy = uint_c(std::round(std::sqrt(real_c(remaining))));     
+            for (uint_t y = base_xy; y >= 1; --y) {
+                if (remaining % y == 0) {
+                    uint_t x = remaining / y;
+                    procs_x = x; procs_y = y; procs_z = z;
+                    goto found;
+                }
+            }
+        }
+    }
+    found:   
+    return std::make_tuple(procs_x, procs_y, procs_z);
+}
+
 void initDirichletBoundaries(const shared_ptr<StructuredBlockForest>& blocks,
                             BlockDataID uId, BlockDataID uTmpId) {
     // Simplified boundary initialization - hot north (3800K), cold elsewhere (300K)
@@ -100,79 +125,79 @@ int main(int argc, char** argv) {
 
     // Command line argument parsing
     bool weakScaling = false;
-    uint_t baseCells = 256;  // Default cells per block/rank
+    uint_t problemSize = 256;  // Default cells per block/rank
 
     if (argc > 1) {
         std::string scalingType(argv[1]);
         if (scalingType == "weak") weakScaling = true;
+        else if (scalingType == "strong") weakScaling = false;
+        else {WALBERLA_LOG_WARNING_ON_ROOT("Unknown scaling type: " << scalingType << ". Use 'weak' or 'strong'. Defaulting to strong.");}
     }
     if (argc > 2) {
-        baseCells = uint_c(std::stoi(argv[2]));
+        problemSize = uint_c(std::stoi(argv[2]));
     }
 
     const uint_t numProcesses = uint_c(MPIManager::instance()->numProcesses());
 
-    // Calculate domain decomposition
-    uint_t procs_x = 1, procs_y = 1, procs_z = 1;
-    if (numProcesses == 1) {
-        procs_x = 1; procs_y = 1; procs_z = 1;
-    } else if (numProcesses == 2) {
-        procs_x = 2; procs_y = 1; procs_z = 1;
-    } else if (numProcesses == 4) {
-        procs_x = 2; procs_y = 2; procs_z = 1;
-    } else if (numProcesses == 8) {
-        procs_x = 2; procs_y = 2; procs_z = 2;
-    } else if (numProcesses == 16) {
-        procs_x = 2; procs_y = 2; procs_z = 4;
-    } else if (numProcesses == 32) {
-        procs_x = 2; procs_y = 4; procs_z = 4;
-    } else if (numProcesses == 64) {
-        procs_x = 4; procs_y = 4; procs_z = 4;
-    } else {
-        WALBERLA_ABORT("Unsupported number of processes: " << numProcesses);
-    }
+    // Automatic process decomposition
+    uint_t procs_x, procs_y, procs_z;
[…]
