### 2025-08-25 7d79bbf
Add benchmarking scripts for LUMI-C performance testing
