### 2025-08-25 8c08363
Update gitignore to not ignore build script
