### 2025-08-25 4b4a0f7
Update CPU scaling code
