### 2025-08-26 fa87f94
Load GPU modules script
