### 2025-08-27 f4e4ab9
Update lumi setup script
