### 2025-08-27 db3f354
Update run scripts
