### 2025-08-27 1499652
Update and add single node run script
