### 2025-08-27 170e084
Update lumi setup script
