### 2025-08-27 bbadc98
Change pystencils target from CUDA to HIP for LUMI
