### 2025-08-27 f6fc35a
Update modules for GPU
