### 2025-08-27 2e6ea5f
Update CPU build script
