### 2025-08-27 91ec7ae
Update setup script
