### 2025-08-27 33bcd40
Update lumi setup
