### 2025-08-28 ee84492
Update setup for CPU and GPU on lumi
