+++ lumi.sh
@@ -1,4 +1,5 @@
 #!/bin/bash
+
 # LUMI setup for materforge project
 
 # CPU Setup - Default execution
@@ -8,12 +9,12 @@ lumi_cpu() {
     module load LUMI/24.03 partition/C
     module load PrgEnv-gnu
     module load buildtools/24.03 cray-python/3.11.7
-
+    
     # Activate virtual environment
     if [[ -f /project/project_465001284/venvs/materforge/bin/activate ]]; then
         source /project/project_465001284/venvs/materforge/bin/activate
     fi
-
+    
     echo "✅ LUMI CPU environment loaded!"
     echo "Partition: CPU (partition/C)"
     echo "Compiler: $(cc --version | head -1)"
@@ -23,7 +24,7 @@ lumi_cpu() {
 lumi_gpu() {
     module purge  
     module --force purge
-
+    
     # Clear ALL potentially problematic environment variables
     unset HIP_FLAGS
     unset HIPCXX_FLAGS
@@ -31,15 +32,17 @@ lumi_gpu() {
     unset CRAY_ROCM_INCLUDE_OPTS
     unset HIP_COMPILE_FLAGS
     unset HIP_LINK_FLAGS
-
+    
     module load LUMI/24.03 partition/G PrgEnv-cray buildtools/24.03 craype-accel-amd-gfx90a rocm/6.0.3
     module load cray-mpich
+    
+    # CRITICAL: Enable GPU support for MPI GTL
+    export MPICH_GPU_SUPPORT_ENABLED=1
 
     # Force override the problematic variable AFTER module loading
     export CRAY_ROCM_INCLUDE_OPTS="-I/opt/rocm-6.0.3/include -I/opt/rocm-6.0.3/include/rocprofiler -I/opt/rocm-6.0.3/include/roctracer -I/opt/rocm-6.0.3/include/hip"
-
     unset CRAY_ROCM_INCLUDE_OPTS
-
+    
     # Set ROCm environment variables
     export ROCM_PATH=/opt/rocm-6.0.3
     export HIP_PATH=/opt/rocm-6.0.3  
@@ -47,22 +50,31 @@ lumi_gpu() {
     export HIP_DEVICE_LIB_PATH=/opt/rocm-6.0.3/amdgcn/bitcode
     export CMAKE_PREFIX_PATH="/opt/rocm-6.0.3:/opt/rocm-6.0.3/lib/cmake"
     export HIP_PLATFORM=amd
-
+    
     # Additional safeguards - clear any cached CMake HIP flags
     export CMAKE_HIP_FLAGS=""
     
-    # Override the problematic CRAY_ROCM_INCLUDE_OPTS without the -D__HIP_PLATFORM_AMD__ 
-    # export CRAY_ROCM_INCLUDE_OPTS="-I/opt/rocm-6.0.3/include -I/opt/rocm-6.0.3/include/rocprofiler -I/opt/rocm-6.0.3/include/roctracer -I/opt/rocm-6.0.3/include/hip"
+    # CRITICAL: Set environment variables to make MPI headers visible to HIP compiler
+    MPI_INCLUDE_PATH="/opt/cray/pe/mpich/8.1.29/ofi/crayclang/17.0/include"
+    export CPATH="${MPI_INCLUDE_PATH}:$CPATH"
+    export C_INCLUDE_PATH="${MPI_INCLUDE_PATH}:$C_INCLUDE_PATH"  
+    export CPLUS_INCLUDE_PATH="${MPI_INCLUDE_PATH}:$CPLUS_INCLUDE_PATH"
+    
+    # Debug: Check if GTL libraries are available
+    echo "Checking GTL environment variables:"
[…]
