### 2025-08-28 f912085
Update CMakeLists for HIP and CUDA
