+++ apps/CMakeLists.txt
@@ -1,7 +1,5 @@
 cmake_minimum_required(VERSION 3.24)
-
 project(test_materforge LANGUAGES CXX)
-
 SET(CMAKE_CXX_STANDARD 20)
 SET(CMAKE_CXX_STANDARD_REQUIRED ON)
 
@@ -17,121 +15,160 @@ SET(WALBERLA_BUILD_WITH_CODEGEN OFF CACHE BOOL "")
 SET(WALBERLA_BUILD_WITH_OPENMP ON CACHE BOOL "")
 SET(WALBERLA_BUILD_WITH_MPI ON CACHE BOOL "")
 
-# Conditional GPU support
+# Build mode detection
+option(BUILD_CPU_ONLY "Build only CPU targets (ignore GPU)" OFF)
 option(ENABLE_GPU_SUPPORT "Enable GPU support" ON)
 
-# Check if CUDA was explicitly requested via command line
-if(DEFINED WALBERLA_BUILD_WITH_CUDA AND WALBERLA_BUILD_WITH_CUDA)
-    message(STATUS "CUDA explicitly requested via WALBERLA_BUILD_WITH_CUDA")
-    set(CUDA_REQUESTED TRUE)
-else()
-    set(CUDA_REQUESTED FALSE)
-endif()
-
-if(ENABLE_GPU_SUPPORT OR CUDA_REQUESTED)
-    # Check if CUDA is available
-    include(CheckLanguage)
-    check_language(CUDA)
-    if(CMAKE_CUDA_COMPILER)
-        enable_language(CUDA)
-        find_package(CUDAToolkit QUIET)
-        if (CUDAToolkit_FOUND)
-            SET(WALBERLA_BUILD_WITH_GPU_SUPPORT ON CACHE BOOL "" FORCE)
-            SET(WALBERLA_BUILD_WITH_CUDA ON CACHE BOOL "" FORCE)
-            message(STATUS "CUDA found - GPU support enabled")
-            set(GPU_AVAILABLE TRUE)
-        else()
-            message(WARNING "CUDA compiler found but toolkit missing - GPU support disabled")
-            SET(WALBERLA_BUILD_WITH_GPU_SUPPORT OFF CACHE BOOL "" FORCE)
-            SET(WALBERLA_BUILD_WITH_CUDA OFF CACHE BOOL "" FORCE)
-            set(GPU_AVAILABLE FALSE)
-        endif()
-    else()
-        if(CUDA_REQUESTED)
-            message(FATAL_ERROR "CUDA was explicitly requested but compiler not found - aborting configuration")
-        else()
-            message(STATUS "CUDA compiler not found - GPU support disabled")
-            SET(WALBERLA_BUILD_WITH_GPU_SUPPORT OFF CACHE BOOL "" FORCE)
-            SET(WALBERLA_BUILD_WITH_CUDA OFF CACHE BOOL "" FORCE)
-            set(GPU_AVAILABLE FALSE)
-        endif()
-    endif()
-else()
+# Configure waLBerla based on build mode
+if(BUILD_CPU_ONLY)
+    message(STATUS "CPU-only build - configuring waLBerla without GPU support")
     SET(WALBERLA_BUILD_WITH_GPU_SUPPORT OFF CACHE BOOL "" FORCE)
     SET(WALBERLA_BUILD_WITH_CUDA OFF CACHE BOOL "" FORCE)
+    SET(WALBERLA_BUILD_WITH_HIP OFF CACHE BOOL "" FORCE)
     set(GPU_AVAILABLE FALSE)
-    message(STATUS "GPU support manually disabled")
+else()
[…]
