### 2025-08-28 67ed62f
Add build script for lumi-g
