### 2025-08-28 c62b96c
Update build scripts for cpu and gpu
