### 2025-08-29 8532227
Fix GTL linking
