+++ apps/CMakeLists.txt
@@ -117,11 +117,25 @@ if(BUILD_CPU_ONLY)
 
 else()
     if(GPU_TYPE STREQUAL "HIP")
-        # Find MPI
         find_package(MPI REQUIRED)
+        message(STATUS "MPI_CXX_INCLUDE_DIRS: ${MPI_CXX_INCLUDE_DIRS}")
+        
+        # Get GTL library information from environment
+        set(PE_MPICH_GTL_DIR_amd_gfx90a "$ENV{PE_MPICH_GTL_DIR_amd_gfx90a}")
+        set(PE_MPICH_GTL_LIBS_amd_gfx90a "$ENV{PE_MPICH_GTL_LIBS_amd_gfx90a}")
+        
+        # Strip the -L and -l prefixes for CMake
+        if(PE_MPICH_GTL_DIR_amd_gfx90a)
+            string(REPLACE "-L" "" GTL_LIB_DIR "${PE_MPICH_GTL_DIR_amd_gfx90a}")
+        endif()
+        if(PE_MPICH_GTL_LIBS_amd_gfx90a)
+            string(REPLACE "-l" "" GTL_LIB_NAME "${PE_MPICH_GTL_LIBS_amd_gfx90a}")
+        endif()
         
         # Debug: Print MPI include directories
         message(STATUS "MPI_CXX_INCLUDE_DIRS: ${MPI_CXX_INCLUDE_DIRS}")
+        message(STATUS "GTL_LIB_DIR: ${GTL_LIB_DIR}")
+        message(STATUS "GTL_LIB_NAME: ${GTL_LIB_NAME}")
         
         # Create executables with full MPI::MPI_CXX target
         add_executable(HeatEquationWithMaterialGPU)
@@ -132,6 +146,12 @@ else()
                 hip::device
                 MPI::MPI_CXX)
         
+        # Link GTL library correctly
+        if(GTL_LIB_DIR AND GTL_LIB_NAME)
+            target_link_directories(HeatEquationWithMaterialGPU PRIVATE ${GTL_LIB_DIR})
+            target_link_libraries(HeatEquationWithMaterialGPU PRIVATE ${GTL_LIB_NAME})
+        endif()
+
         # Add compile options for HIP
         target_compile_options(HeatEquationWithMaterialGPU PRIVATE 
             $<$<COMPILE_LANGUAGE:HIP>:--offload-arch=gfx90a>)
@@ -147,7 +167,7 @@ else()
                 ${CMAKE_CURRENT_BINARY_DIR}/_gen/HeatEquationWithMaterialGPU/gen/HeatEquationKernelWithMaterialGPU.cpp
                 PROPERTIES LANGUAGE HIP)
 
-        # Same pattern for CodegenHeatEquationGPUScaling
+        # CodegenHeatEquationGPUScaling
         add_executable(CodegenHeatEquationGPUScaling)
         target_sources(CodegenHeatEquationGPUScaling PRIVATE CodegenHeatEquationGPUScaling.cpp)
         target_link_libraries(CodegenHeatEquationGPUScaling PRIVATE
@@ -156,6 +176,12 @@ else()
                 hip::device
                 MPI::MPI_CXX)
         
+        # Link GTL library correctly
+        if(GTL_LIB_DIR AND GTL_LIB_NAME)
+            target_link_directories(CodegenHeatEquationGPUScaling PRIVATE ${GTL_LIB_DIR})
+            target_link_libraries(CodegenHeatEquationGPUScaling PRIVATE ${GTL_LIB_NAME})
+        endif()
+
         target_compile_options(CodegenHeatEquationGPUScaling PRIVATE 
[…]
