### 2025-08-29 6e20e1b
Update CPU and GPU scaling code
