+++ apps/CodegenHeatEquationGPUScaling.cpp
@@ -2,7 +2,6 @@
 //! \file CodegenHeatEquationGPUScaling.cpp
 //! \author Rahil Doshi
 //======================================================================================================================
-
 #include "blockforest/Initialization.h"
 #include "core/Environment.h"
 #include "field/AddToStorage.h"
@@ -31,10 +30,37 @@ void swapFields(StructuredBlockForest& blocks, BlockDataID uID, BlockDataID uTmp
     }
 }
 
+// Automatic process decomposition for any number of processes (same as CPU)
+std::tuple<uint_t, uint_t, uint_t> calculateProcessDecomposition(uint_t numProcesses) {
+    // Find best factorization for 3D decomposition
+    uint_t procs_x = 1, procs_y = 1, procs_z = 1;
+    
+    // Start with cube root as base
+    uint_t base = uint_c(std::round(std::cbrt(real_c(numProcesses))));
+    
+    // Find factors close to cube root
+    for (uint_t z = base; z >= 1; --z) {
+        if (numProcesses % z == 0) {
+            uint_t remaining = numProcesses / z;
+            uint_t base_xy = uint_c(std::round(std::sqrt(real_c(remaining))));
+            
+            for (uint_t y = base_xy; y >= 1; --y) {
+                if (remaining % y == 0) {
+                    uint_t x = remaining / y;
+                    procs_x = x; procs_y = y; procs_z = z;
+                    goto found;
+                }
+            }
+        }
+    }
+    found:
+    return std::make_tuple(procs_x, procs_y, procs_z);
+}
+
 void initDirichletBoundaries(const shared_ptr<StructuredBlockForest>& blocks,
                             BlockDataID uId, BlockDataID uTmpId,
                             BlockDataID uCpuId, BlockDataID uTmpCpuId) {
-    // Initialize on CPU, then copy to GPU
+    // Simplified boundary initialization - hot north (3800K), cold elsewhere (300K)
     for (auto block = blocks->begin(); block != blocks->end(); ++block) {
         ScalarField* u = block->getData<ScalarField>(uCpuId);
         ScalarField* u_tmp = block->getData<ScalarField>(uTmpCpuId);
@@ -44,7 +70,7 @@ void initDirichletBoundaries(const shared_ptr<StructuredBlockForest>& blocks,
             CellInterval north = u->xyzSizeWithGhostLayer();
             north.yMin() = north.yMax();
             for (auto cell = north.begin(); cell != north.end(); ++cell) {
-                u->get(*cell) = real_c(3800.0);
+                u->get(*cell) = real_c(3800.0);  // Hot boundary
                 u_tmp->get(*cell) = real_c(3800.0);
             }
         }
@@ -53,9 +79,8 @@ void initDirichletBoundaries(const shared_ptr<StructuredBlockForest>& blocks,
             CellInterval south = u->xyzSizeWithGhostLayer();
             south.yMax() = south.yMin();
             for (auto cell = south.begin(); cell != south.end(); ++cell) {
-                real_t v = real_c(300.0); // Cold boundary
-                u->get(*cell) = v;
-                u_tmp->get(*cell) = v;
+                u->get(*cell) = real_c(300.0); // Cold boundary
[…]
