### 2025-08-29 8572e31
Update CPU and GPU scaling code
