### 2025-08-29 d17dc7a
Add single node run script for lumi-g
