### 2025-08-29 2d7dfcc
Remove depricated run script
