### 2025-09-01 9d8ae9c
Update build script for lumi-g
