### 2025-09-01 f1a3b51
Update single node run script for lumi-g
