### 2025-09-01 3fa2680
Update build script for lumi-g
