+++ benchmarking-scripts/materforge/HeatEquation/lumi-g/build_materforge.sh
@@ -6,29 +6,18 @@
 # ---
 
 timestamp=$(date +%s)
+
 SCRIPT_PATH="$(realpath $0)"
 NAME="materforge"
+
 MATERFORGE_REPO="https://i10git.cs.fau.de/rahil.doshi/materforge.git"
 MATERFORGE_BRANCH="bm"
 MATERFORGE_COMMIT=""
 
-# Updated paths
-SRC_DIR="/project/project_465001284/repos/materforge"
+SRC_DIR="${MATERFORGE_SRC_DIR:-$(pwd)/materforge_src}"
 BUILD_DIR="$(pwd)/build_${NAME}_${timestamp}"
 LOGFILE="${BUILD_DIR}/build_${timestamp}.log"
-
-# Print the values of the variables
-echo
-echo -e "${PROGRESS}--- Inside the materforge build script ---${NC}"
-echo
-echo -e "${INFO}BENCHMARK_NAME = $NAME"
-echo -e "${INFO}TIME_LABEL = $TIME_LABEL"
-echo -e "${INFO}BASE_DIR = $BASE_DIR"
-echo -e "${INFO}SRC_DIR = $SRC_DIR"
-echo -e "${INFO}BUILD_DIR = $BUILD_DIR"
-echo -e "${INFO}BUILD_LOG = $LOGFILE"
-echo -e "${INFO}SCRIPT_DIR = $(dirname "$SCRIPT_PATH")${NC}"
-echo
+VENV_PATH="${MATERFORGE_VENV:-/project/project_465001284/venvs/materforge}"
 
 mkdir -p "${BUILD_DIR}" || { echo -e "${ERROR}Failed to create build directory: ${BUILD_DIR}${NC}"; exit 1; }
 
@@ -38,7 +27,7 @@ mkdir -p "${BUILD_DIR}" || { echo -e "${ERROR}Failed to create build directory: 
     cat "${SCRIPT_PATH}"
     echo "---"
     
-    # Clone or update the materforge source code
+    # Clone or use existing repository
     if [ ! -d "${SRC_DIR}" ]; then
         echo "Cloning into ${SRC_DIR}"
         git clone -b ${MATERFORGE_BRANCH} ${MATERFORGE_REPO} ${SRC_DIR}
@@ -46,18 +35,20 @@ mkdir -p "${BUILD_DIR}" || { echo -e "${ERROR}Failed to create build directory: 
         echo "Using existing repository at ${SRC_DIR}"
     fi
 
-	echo "Git Info:"
-	echo "Git url = $(git remote get-url origin 2>/dev/null || echo 'No remote')"
-	echo "Git branch = $(git branch --show-current 2>/dev/null || echo 'No branch')"
-	echo "Git commit = $(git rev-parse HEAD 2>/dev/null || echo 'No commit')"
-	echo
-    
     cd "${SRC_DIR}" || exit 1
+
+    echo "Current directory: $(pwd)"
+    echo "Git Info:"
+    echo "Git url = $(git remote get-url origin 2>/dev/null || echo 'No remote')"
+    echo "Git branch = $(git branch --show-current 2>/dev/null || echo 'No branch')"
+    echo "Git commit = $(git rev-parse HEAD 2>/dev/null || echo 'No commit')"
+    echo
+
     git checkout .
-	git fetch -a
-	git checkout $MATERFORGE_BRANCH
-	[[ -n "$MATERFORGE_COMMIT" ]] && git checkout $MATERFORGE_COMMIT
-	git pull
+    git fetch -a
+    git checkout $MATERFORGE_BRANCH
+    [[ -n "$MATERFORGE_COMMIT" ]] && git checkout $MATERFORGE_COMMIT
+    git pull
 
     # Load LUMI-G modules
[…]
