### 2025-09-01 2dbad48
Update single node run script for lumi-g
