### 2025-09-01 c92fe8f
Update build and single nore run script for lumi-g
