### 2025-09-01 cdf7d00
Increase cells per GPU to 512
