### 2025-09-01 0791878
Add strong scaling scripts for lumi-g
