### 2025-09-01 a10cf8d
Update account for strong scaling scripts for lumi-g
