+++ benchmarking-scripts/materforge/HeatEquation/lumi-g/run_materforge_strong_scaling_medium.sh
@@ -7,7 +7,7 @@
 #SBATCH --ntasks-per-node=8
 #SBATCH --gpus-per-node=8
 #SBATCH --time=0-01:30:00
-#SBATCH --account=project_465001134
+#SBATCH --account=project_465001980
 
 unset SLURM_EXPORT_ENV
 
