### 2025-09-01 2db5cf0
Add weak scaling script for lumi-g
