### 2025-09-01 24b0637
Update weak scaling script for lumi-g
