### 2025-09-02 7a6926f
Update build scripts for lumi-c and lumi-g
