### 2025-09-02 213e982
Update build scripts for lumi-c and lumi-g
