+++ benchmarking-scripts/materforge/HeatEquation/lumi-c/build_materforge.sh
@@ -1,12 +1,12 @@
 #!/bin/bash -l
+
 # ---
 # type: buildscript
 # cluster: lumi-c
-# application: apps/materforge
+# application: materforge
 # ---
 
 timestamp=$(date +%s)
-
 SCRIPT_PATH="$(realpath $0)"
 NAME="materforge"
 
@@ -14,6 +14,7 @@ MATERFORGE_REPO="https://i10git.cs.fau.de/rahil.doshi/materforge.git"
 MATERFORGE_BRANCH="bm"
 MATERFORGE_COMMIT=""
 
+# Configurable paths
 SRC_DIR="${MATERFORGE_SRC_DIR:-$(pwd)/materforge_src}"
 BUILD_DIR="$(pwd)/build_${NAME}_${timestamp}"
 LOGFILE="${BUILD_DIR}/build_${timestamp}.log"
@@ -26,81 +27,84 @@ SUCCESS='\033[1;32m'
 INFO='\033[1;36m'
 PROGRESS='\033[1;35m'
 
+{
 mkdir -p "${BUILD_DIR}" || { echo -e "${ERROR}Failed to create build directory: ${BUILD_DIR}${NC}"; exit 1; }
 
-{
-	date -d @${timestamp}
-	echo "--- currently executed script: $(basename ${SCRIPT_PATH})"
-	cat "${SCRIPT_PATH}"
-	echo "---"
-	
-	# Clone or use existing repository with submodules
-	if [ ! -d "${SRC_DIR}" ]; then
-		echo -e "${PROGRESS}Cloning into ${SRC_DIR}${NC}"
-		git clone --recursive -b ${MATERFORGE_BRANCH} ${MATERFORGE_REPO} ${SRC_DIR}
-	else
-		echo -e "${PROGRESS}Using existing repository at ${SRC_DIR}${NC}"
-	fi
+date -d @${timestamp}
+echo "--- currently executed script: $(basename ${SCRIPT_PATH})"
+cat "${SCRIPT_PATH}"
+echo "---"
 
-	cd "${SRC_DIR}" || exit 1
+# Clone or use existing repository with submodules
+if [ ! -d "${SRC_DIR}" ]; then
+    echo -e "${PROGRESS}Cloning into ${SRC_DIR}${NC}"
+    git clone --recursive -b ${MATERFORGE_BRANCH} ${MATERFORGE_REPO} ${SRC_DIR}
+else
+    echo -e "${PROGRESS}Using existing repository at ${SRC_DIR}${NC}"
+fi
 
-	echo "Current directory: $(pwd)"
-	echo "Git Info:"
-	echo "Git url = $(git remote get-url origin 2>/dev/null || echo 'No remote')"
-	echo "Git branch = $(git branch --show-current 2>/dev/null || echo 'No branch')"
-	echo "Git commit = $(git rev-parse HEAD 2>/dev/null || echo 'No commit')"
-	echo
-	
-	git checkout .
-	git fetch -a
-	git checkout $MATERFORGE_BRANCH
-	[[ -n "$MATERFORGE_COMMIT" ]] && git checkout $MATERFORGE_COMMIT
-	git pull
-	
-	# Initialize and update submodules
-	echo -e "${PROGRESS}Updating submodules${NC}"
-	git submodule update --init --recursive || exit 1
+cd "${SRC_DIR}" || exit 1
 
-	# Load LUMI-C modules
-	module load LUMI/24.03 partition/C
-	module load PrgEnv-gnu
-	module load buildtools/24.03 cray-python/3.11.7
-	module list
-	
-	# Activate virtual environment
-	source /project/project_465001284/venvs/materforge/bin/activate
-	
[…]
