### 2025-09-02 dad4a47
Update build scripts for lumi-c and lumi-g
