### 2025-09-02 368b452
Add single node run script for lumi-c
