### 2025-09-02 3317f63
Update strong scaling run scripts for lumi-c
