### 2025-09-02 8712bbc
Update weak scaling run script for lumi-c
