### 2025-09-02 c604e53
Update default cells per CPU core to 384 for lumi-c
