### 2025-09-02 4123654
Update default cells per CPU core to 384 for lumi-c
