### 2025-09-03 f86da8a
Build app with hardcoaded values, without materforge
