### 2025-09-03 fe944e1
Update default cells per GPU to 768
