### 2025-09-03 28185e6
Extent heat_pde to 3D
