### 2025-09-03 467e892
Correct code for single node
