+++ apps/CodegenHeatEquationCPUScaling.cpp
@@ -124,13 +124,15 @@ int main(int argc, char** argv) {
 
     // Command line argument parsing
     bool weakScaling = false;
+    bool singleNode = false;
     uint_t problemSize = 256;  // Default cells per block/rank
 
     if (argc > 1) {
-        std::string scalingType(argv[1]);
-        if (scalingType == "weak") weakScaling = true;
-        else if (scalingType == "strong") weakScaling = false;
-        else {WALBERLA_LOG_WARNING_ON_ROOT("Unknown scaling type: " << scalingType << ". Use 'weak' or 'strong'. Defaulting to strong.");}
+        std::string testType(argv[1]);
+        if (testType == "weak") weakScaling = true;
+        else if (testType == "strong") weakScaling = false;
+        else if (testType == "single") singleNode = true;
+        else {WALBERLA_LOG_WARNING_ON_ROOT("Unknown test type: " << testType << ". Use 'weak', 'strong', or 'single'. Defaulting to strong.");}
     }
     if (argc > 2) {
         problemSize = uint_c(std::stoi(argv[2]));
@@ -145,7 +147,19 @@ int main(int argc, char** argv) {
     uint_t xCells, yCells, zCells;
     real_t xSize, ySize, zSize;
 
-    if (weakScaling) {
+    if (singleNode) {
+        // Single node test: all processes in one node
+        const uint_t totalCellsPerDim = problemSize;
+        xCells = totalCellsPerDim / procs_x;
+        yCells = totalCellsPerDim / procs_y;
+        zCells = totalCellsPerDim / procs_z;
+        if (totalCellsPerDim % procs_x != 0 || totalCellsPerDim % procs_y != 0 || 
+            totalCellsPerDim % procs_z != 0) {
+            WALBERLA_ABORT("Problem size " << problemSize << " not evenly divisible by process grid " << procs_x << "x" << procs_y << "x" << procs_z);
+        }
+        xSize = ySize = zSize = real_c(1.0);
+        WALBERLA_LOG_INFO_ON_ROOT("Single Node Test: " << numProcesses << " processes on one node with " << problemSize << "^3 cells each");
+    } else if (weakScaling) {
         // Weak scaling: constant cells per process
         xCells = problemSize;
         yCells = problemSize;
@@ -154,7 +168,7 @@ int main(int argc, char** argv) {
         xSize = real_c(procs_x * 1.0);
         ySize = real_c(procs_y * 1.0);
         zSize = real_c(procs_z * 1.0);
-        WALBERLA_LOG_INFO_ON_ROOT("Weak Scaling: " << problemSize << "^3 cells per process");
+        WALBERLA_LOG_INFO_ON_ROOT("Weak Scaling: " << numProcesses << " processes with " << problemSize << "^3 cells per process");
     } else {
         // Strong scaling: fixed total domain size
         const uint_t totalCellsPerDim = problemSize;
@@ -163,14 +177,13 @@ int main(int argc, char** argv) {
         zCells = totalCellsPerDim / procs_z;
         // Ensure cells are evenly divisible
         if (totalCellsPerDim % procs_x != 0 || totalCellsPerDim % procs_y != 0 || totalCellsPerDim % procs_z != 0) {
-            WALBERLA_ABORT("Problem size " << problemSize << " not evenly divisible by process grid " 
-                          << procs_x << "x" << procs_y << "x" << procs_z);
[…]
