### 2025-09-04 a753ec8
Update python API call to evaluate material properties
