+++ apps/CouetteFlow/CouetteFlowMaterial.yaml
@@ -1,0 +1,171 @@
+# ====================================================================================================
+# MATERFORGE MATERIAL CONFIGURATION FILE
+# ====================================================================================================
+# Material properties for Couette flow simulation with temperature-dependent viscosity
+# This file defines a steel-like material for temperature-dependent viscous flow analysis
+# ====================================================================================================
+
+name: "CouetteFlowSteel"
+material_type: alloy  # Steel is an alloy
+
+# Composition fractions must sum to 1.0 (simplified steel composition)
+composition:
+  Fe: 0.70   # Iron (primary component)
+  C: 0.01    # Carbon
+  Cr: 0.18   # Chromium (for corrosion resistance)
+  Ni: 0.08   # Nickel
+  Mn: 0.02   # Manganese
+  Si: 0.01   # Silicon
+
+# Required temperature properties for alloys
+solidus_temperature: 1673.0    # Temperature where melting begins (K) - 1400C
+liquidus_temperature: 1723.0   # Temperature where material is completely melted (K) - 1450C
+initial_boiling_temperature: 3100.0  # Temperature where boiling begins (K)
+final_boiling_temperature: 3200.0    # Temperature where material is completely vaporized (K)
+
+properties:
+  # ====================================================================================================
+  # CONSTANT PROPERTIES
+  # ====================================================================================================
+  
+  latent_heat_of_fusion: 270000.0      # J/kg - Heat of fusion for steel
+  latent_heat_of_vaporization: 6300000.0  # J/kg - Heat of vaporization for steel
+
+  # ====================================================================================================
+  # DENSITY - Temperature-dependent (thermal expansion)
+  # ====================================================================================================
+  
+  density:
+    dependency: (300, 1200, 91)  # Temperature range 300-1200K, 91 points (10K intervals)
+    equation: 7850.0 / (1.0 + 3.0 * thermal_expansion_coefficient * (T - 293.15))  # Thermal expansion model
+    bounds: [constant, constant]
+    regression:
+      simplify: post
+      degree: 2
+      segments: 2
+
+  # ====================================================================================================
+  # THERMAL EXPANSION COEFFICIENT - Temperature-dependent
+  # ====================================================================================================
+  
+  thermal_expansion_coefficient:
+    dependency: [300, 400, 500, 600, 700, 800, 900, 1000, 1100, 1200]
+    value: [11.7e-6, 12.8e-6, 13.5e-6, 14.0e-6, 14.4e-6, 14.8e-6, 15.1e-6, 15.4e-6, 15.7e-6, 16.0e-6]  # 1/K
+    bounds: [constant, constant]
+    regression:
+      simplify: post
+      degree: 1
+      segments: 1
+
+  # ====================================================================================================
+  # HEAT CAPACITY - Temperature-dependent
+  # ====================================================================================================
+  
+  heat_capacity:
+    dependency: [300, 400, 500, 600, 700, 800, 900, 1000, 1100, 1200]
+    value: [460, 500, 520, 540, 570, 600, 650, 720, 800, 900]  # J/(kg·K)
+    bounds: [constant, constant]
+    regression:
+      simplify: post
+      degree: 2
+      segments: 2
+
+  # ====================================================================================================
+  # THERMAL CONDUCTIVITY - Temperature-dependent 
+  # ====================================================================================================
+  
+  heat_conductivity:
+    dependency: [300, 400, 500, 600, 700, 800, 900, 1000, 1100, 1200]
[…]
