+++ apps/CouetteFlowSimulation.cpp
@@ -1,0 +1,274 @@
+//======================================================================================================================
+//! \file CouetteFlowSimulation.cpp
+//! \author Your Name
+//! \brief Couette Flow Simulation with temperature-dependent viscosity using waLBerla and materforge
+//======================================================================================================================
+
+#include "blockforest/Initialization.h"
+#include "blockforest/communication/UniformBufferedScheme.h"
+#include "core/Environment.h"
+#include "core/math/Constants.h"
+#include "field/AddToStorage.h"
+#include "field/communication/PackInfo.h"
+#include "stencil/D3Q19.h"
+#include "field/vtk/VTKWriter.h"
+#include "timeloop/SweepTimeloop.h"
+#include "core/timing/TimingPool.h"
+
+// Include generated headers (these should be generated by your Python scripts)
+#include "gen/CouetteFlowKernel.hpp"
+
+namespace walberla {
+namespace couette_flow {
+
+using namespace walberla;
+
+// Type definitions
+typedef GhostLayerField<real_t, 1> ScalarField;
+typedef GhostLayerField<Vector3<real_t>, 1> VectorField;
+typedef GhostLayerField<real_t, 19> PdfField;
+
+// Material properties structure
+struct MaterialProperties {
+    real_t density;
+    real_t reference_viscosity;
+    real_t temperature_coefficient; // α in μ(T) = μ₀ * exp(-α * (T - T_ref))
+    real_t reference_temperature;
+    real_t thermal_conductivity;
+    real_t specific_heat;
+    
+    MaterialProperties() :
+        density(7850.0),
+        reference_viscosity(0.001),
+        temperature_coefficient(0.002),
+        reference_temperature(300.0),
+        thermal_conductivity(50.0),
+        specific_heat(500.0)
+    {}
+    
+    // Temperature-dependent viscosity: μ(T) = μ₀ * exp(-α * (T - T_ref))
+    real_t getDynamicViscosity(real_t temperature) const {
+        return reference_viscosity * std::exp(-temperature_coefficient * (temperature - reference_temperature));
+    }
+    
+    // Kinematic viscosity: ν = μ/ρ
+    real_t getKinematicViscosity(real_t temperature) const {
+        return getDynamicViscosity(temperature) / density;
+    }
+    
+    // Thermal diffusivity: α = k/(ρ*cp)
+    real_t getThermalDiffusivity() const {
+        return thermal_conductivity / (density * specific_heat);
+    }
+};
+
+// Boundary condition application
+void applyBoundaryConditions(const shared_ptr<StructuredBlockForest>& blocks,
+                           BlockDataID velocityId, BlockDataID temperatureId,
+                           real_t wallVelocity, real_t hotTemperature, real_t coldTemperature) {
+    
+    for (auto block = blocks->begin(); block != blocks->end(); ++block) {
+        VectorField* velocity = block->getData<VectorField>(velocityId);
+        ScalarField* temperature = block->getData<ScalarField>(temperatureId);
+        
+        // Get domain boundaries
+        CellInterval domain = velocity->xyzSizeWithGhostLayer();
+        
+        // Top wall (moving wall with prescribed velocity and temperature)
+        if (blocks->atDomainYMaxBorder(*block)) {
+            CellInterval topWall = domain;
+            topWall.yMin() = topWall.yMax();
+            
+            for (auto cell = topWall.begin(); cell != topWall.end(); ++cell) {
+                velocity->get(*cell) = Vector3<real_t>(wallVelocity, 0.0, 0.0);
[…]
