+++ apps/CouetteFlowKernel.py
@@ -1,0 +1,213 @@
+#!/usr/bin/env python3
+"""
+CouetteFlowKernel.py
+Code generation script for Couette flow simulation with temperature-dependent viscosity
+"""
+
+import logging
+import pystencils as ps
+from pathlib import Path
+from pystencilssfg import SourceFileGenerator
+from pystencils import SymbolCreator
+from walberla.codegen import Sweep
+from lbmpy import LBMConfig, Stencil, Method, create_lb_method, create_lb_update_rule, LBMOptimisation
+
+# Configure logging
+logging.basicConfig(
+    level=logging.WARNING,
+    format="%(asctime)s %(levelname)s %(name)s -> %(message)s"
+)
+
+# Silence noisy libraries
+logging.getLogger('matplotlib').setLevel(logging.WARNING)
+logging.getLogger('PIL').setLevel(logging.WARNING)
+logging.getLogger('fontTools').setLevel(logging.WARNING)
+
+# Material properties integration flag
+use_material_library = False  # Set to True to use materforge integration
+
+with SourceFileGenerator() as sfg:
+    data_type = "float64"
+    
+    # Define fields
+    f_pdfs, f_pdfs_tmp = ps.fields(f"f_pdfs({19}), f_pdfs_tmp({19}): {data_type}[3D]", layout='fzyx')
+    f_velocity, f_temperature, f_viscosity, f_density, f_force = ps.fields(
+        f"f_velocity(3), f_temperature, f_viscosity, f_density, f_force(3): {data_type}[3D]", layout='fzyx'
+    )
+    
+    # Create symbol creator for parameters
+    s = SymbolCreator()
+    s_dx, s_dt, s_wall_velocity, s_viscosity_ref = s.s_dx, s.s_dt, s.s_wall_velocity, s.s_viscosity_ref
+    s_temp_coeff, s_temp_ref = s.s_temp_coeff, s.s_temp_ref
+    s_omega = s.s_omega  # LBM relaxation parameter
+    
+    # LBM configuration
+    stencil = Stencil.D3Q19
+    
+    # Temperature-dependent viscosity model
+    # μ(T) = μ_ref * exp(-α * (T - T_ref))
+    constant_viscosity = 0.001
+    
+    # Relaxation time from viscosity: τ = 3ν/c² + 0.5, where ν = μ/ρ, c² = 1/3
+    # τ = 3μ/(ρc²) + 0.5 = 9μ/ρ + 0.5
+    relaxation_time = 9.0 * constant_viscosity / f_density.center() + 0.5
+    omega = 1.0 / relaxation_time
+    
+    # Create LBM configuration
+    lbm_config = LBMConfig(
+        stencil=stencil,
+        method=Method.SRT,  # Single relaxation time
+        relaxation_rate=omega,
+        compressible=False,
+        zero_centered=False,
+        force=f_force.center_vector,
+        output={'density': f_density, 'velocity': f_velocity}
+    )
+    
+    # Create LBM method
+    lb_method = create_lb_method(lbm_config)
+    
+    # Create LBM optimization
+    lbm_opt = LBMOptimisation(
+        symbolic_field=f_pdfs,
+        symbolic_temporary_field=f_pdfs_tmp
+    )
+    
+    # Create main LBM update rule
+    main_assignments = []
+    subexpressions = []
+    
+    # Add temperature-dependent viscosity calculation
+    subexpressions.extend([
+        ps.Assignment(s.viscosity_local, constant_viscosity),
+        ps.Assignment(s.tau_local, relaxation_time),
+        ps.Assignment(s.omega_local, omega),
+    ])
+    
+    # Create LBM collision and streaming
+    collision_rule = create_lb_update_rule(
[…]
