+++ apps/CouetteFlowKernel.py
@@ -1,104 +1,124 @@
 #!/usr/bin/env python3
 """
-CouetteFlowKernel.py
+CouetteFlowKernel.py - CORRECTED WITH SIMPLE VELOCITY BOUNDARY CONDITIONS
 Code generation script for Couette flow simulation with temperature-dependent viscosity
 """
 
 import logging
+import sympy as sp
 import pystencils as ps
 from pathlib import Path
 from pystencilssfg import SourceFileGenerator
 from pystencils import SymbolCreator
 from walberla.codegen import Sweep
-from lbmpy import LBMConfig, Stencil, Method, create_lb_method, create_lb_update_rule, LBMOptimisation
+from lbmpy import LBMConfig, Stencil, Method, create_lb_method, create_lb_update_rule, LBMOptimisation, ForceModel
 
 # Configure logging
-logging.basicConfig(
-    level=logging.WARNING,
-    format="%(asctime)s %(levelname)s %(name)s -> %(message)s"
-)
-
-# Silence noisy libraries
-logging.getLogger('matplotlib').setLevel(logging.WARNING)
-logging.getLogger('PIL').setLevel(logging.WARNING)
-logging.getLogger('fontTools').setLevel(logging.WARNING)
+logging.basicConfig(level=logging.WARNING, format="%(asctime)s %(levelname)s %(name)s -> %(message)s")
 
 # Material properties integration flag
-use_material_library = False  # Set to True to use materforge integration
+use_material_library = True  # Enable materforge integration
 
 with SourceFileGenerator() as sfg:
     data_type = "float64"
     
-    # Define fields
-    f_pdfs, f_pdfs_tmp = ps.fields(f"f_pdfs({19}), f_pdfs_tmp({19}): {data_type}[3D]", layout='fzyx')
-    f_velocity, f_temperature, f_viscosity, f_density, f_force = ps.fields(
-        f"f_velocity(3), f_temperature, f_viscosity, f_density, f_force(3): {data_type}[3D]", layout='fzyx'
+    # Field definitions
+    stencil = Stencil.D3Q19
+    dim = 3
+    Q = 19
+    
+    # Define all fields
+    f_pdfs, f_pdfs_tmp, f_density, f_temperature, f_viscosity, f_force, f_velocity = ps.fields(
+        f"f_pdfs({Q}), f_pdfs_tmp({Q}), f_density, f_temperature, f_viscosity, f_force({dim}), f_velocity({dim}): {data_type}[{dim}D]", 
+        layout='fzyx'
     )
     
     # Create symbol creator for parameters
     s = SymbolCreator()
-    s_dx, s_dt, s_wall_velocity, s_viscosity_ref = s.s_dx, s.s_dt, s.s_wall_velocity, s.s_viscosity_ref
-    s_temp_coeff, s_temp_ref = s.s_temp_coeff, s.s_temp_ref
-    s_omega = s.s_omega  # LBM relaxation parameter
     
-    # LBM configuration
-    stencil = Stencil.D3Q19
+    # ===== MATERFORGE INTEGRATION =====
+    material_assignments = []
+    viscosity_expression = None
     
-    # Temperature-dependent viscosity model
-    # μ(T) = μ_ref * exp(-α * (T - T_ref))
-    constant_viscosity = 0.001
+    if use_material_library:
+        try:
+            from materforge.parsing.api import create_material
+            
+            yaml_path = Path(__file__).parent / 'CouetteFlowMaterial_Updated.yaml'
+            if yaml_path.exists():
+                print(f"Loading material from: {yaml_path}")
+                
+                mat = create_material(yaml_path=yaml_path, dependency=f_temperature.center(), enable_plotting=False)
+                
+                material_assignments.extend([
[…]
