### 2025-09-21 4e55a5f
Update input_methods.jpg
