### 2025-09-22 dd04b63
Update yaml config for couette flow
