+++ apps/CouetteFlowMaterial_Updated.yaml
@@ -2,42 +2,41 @@
 # MATERFORGE MATERIAL CONFIGURATION FILE
 # ====================================================================================================
 # Material properties for Couette flow simulation with temperature-dependent viscosity
-# This file defines a steel-like material for temperature-dependent viscous flow analysis
+# Higher viscosity values for numerical stability
 # ====================================================================================================
 
 name: "CouetteFlowSteel"
-material_type: alloy  # Steel is an alloy
+material_type: alloy
 
-# Composition fractions must sum to 1.0 (simplified steel composition)
 composition:
-  Fe: 0.70   # Iron (primary component)
-  C: 0.01    # Carbon
-  Cr: 0.18   # Chromium (for corrosion resistance)
-  Ni: 0.08   # Nickel
-  Mn: 0.02   # Manganese
-  Si: 0.01   # Silicon
+  Fe: 0.70
+  C: 0.01
+  Cr: 0.18
+  Ni: 0.08
+  Mn: 0.02
+  Si: 0.01
 
-# Required temperature properties for alloys
-solidus_temperature: 1673.0    # Temperature where melting begins (K) - 1400C
-liquidus_temperature: 1723.0   # Temperature where material is completely melted (K) - 1450C
-initial_boiling_temperature: 3100.0  # Temperature where boiling begins (K)
-final_boiling_temperature: 3200.0    # Temperature where material is completely vaporized (K)
+# Temperature properties
+solidus_temperature: 1673.0
+liquidus_temperature: 1723.0
+initial_boiling_temperature: 3100.0
+final_boiling_temperature: 3200.0
 
 properties:
   # ====================================================================================================
   # CONSTANT PROPERTIES
   # ====================================================================================================
   
-  latent_heat_of_fusion: 270000.0      # J/kg - Heat of fusion for steel
-  latent_heat_of_vaporization: 6300000.0  # J/kg - Heat of vaporization for steel
+  latent_heat_of_fusion: 270000.0
+  latent_heat_of_vaporization: 6300000.0
 
   # ====================================================================================================
   # DENSITY - Temperature-dependent (thermal expansion)
   # ====================================================================================================
   
   density:
-    dependency: (300, 1200, 91)  # Temperature range 300-1200K, 91 points (10K intervals)
-    equation: 7850.0 / (1.0 + 3.0 * thermal_expansion_coefficient * (T - 293.15))  # Thermal expansion model
+    dependency: (300, 600, 31)
+    equation: 7850.0 / (1.0 + 3.0 * thermal_expansion_coefficient * (T - 293.15))
     bounds: [constant, constant]
     regression:
       simplify: post
@@ -49,8 +48,8 @@ properties:
   # ====================================================================================================
   
   thermal_expansion_coefficient:
-    dependency: [300, 400, 500, 600, 700, 800, 900, 1000, 1100, 1200]
-    value: [11.7e-6, 12.8e-6, 13.5e-6, 14.0e-6, 14.4e-6, 14.8e-6, 15.1e-6, 15.4e-6, 15.7e-6, 16.0e-6]  # 1/K
+    dependency: [300, 400, 500, 600]
+    value: [11.7e-6, 12.8e-6, 13.5e-6, 14.0e-6]
     bounds: [constant, constant]
     regression:
       simplify: post
@@ -62,8 +61,8 @@ properties:
   # ====================================================================================================
   
   heat_capacity:
[…]
