### 2025-09-23 9964d87
Update CouetteFlowKernel.py
