+++ apps/CouetteFlowSimulation.cpp
@@ -1,7 +1,7 @@
 //======================================================================================================================
-//! \file CouetteFlowSimulation.cpp - OPTIMIZED WITH EFFICIENT BOUNDARY CONDITIONS
-//! \author Rahil Doshi  
-//! \brief Couette Flow Simulation with temperature-dependent viscosity using waLBerla and materforge
+//! \file CouetteFlowThermalSimulation.cpp - USING GENERATED BOUNDARIES
+//! \author Rahil Doshi
+//! \brief Thermal Couette flow using generated boundary sweeps
 //======================================================================================================================
 
 #include "blockforest/Initialization.h"
@@ -19,7 +19,7 @@
 #include "gen/CouetteFlowKernel.hpp"
 
 namespace walberla {
-namespace couette_flow {
+namespace couette_flow_thermal {
 
 using namespace walberla;
 
@@ -28,117 +28,144 @@ typedef GhostLayerField<real_t, 1> ScalarField;
 typedef GhostLayerField<real_t, 3> VectorField;
 typedef GhostLayerField<real_t, 19> PdfField;
 
-// Initialize fields
+// Initialize fields with thermal Couette profile
 void initializeFields(const shared_ptr<StructuredBlockForest>& blocks,
                      BlockDataID densityId, BlockDataID forceId, BlockDataID pdfFieldId,
                      BlockDataID temperatureId, BlockDataID velocityId, BlockDataID viscosityId,
                      real_t hotTemperature, real_t coldTemperature, real_t wallVelocity,
-                     real_t refTemp, real_t refViscosity) {
+                     uint_t yCells, real_t refViscosity) {
     
     for (auto block = blocks->begin(); block != blocks->end(); ++block) {
-        // Get field pointers
         VectorField* velocity = block->getData<VectorField>(velocityId);
         ScalarField* temperature = block->getData<ScalarField>(temperatureId);
         ScalarField* density = block->getData<ScalarField>(densityId);
         VectorField* force = block->getData<VectorField>(forceId);
         ScalarField* viscosity = block->getData<ScalarField>(viscosityId);
-
-        // Get domain size
-        const auto& domain = blocks->getDomain();
-        const real_t domainHeight = domain.ySize();
         
-        // Initialize fields
         for (auto cell = velocity->beginXYZ(); cell != velocity->end(); ++cell) {
             Cell localCell = cell.cell();
             Cell globalCell = localCell;
             blocks->transformBlockLocalToGlobalCell(globalCell, *block);
             
-            // Get normalized position
-            Vector3<real_t> cellCenter = blocks->getCellCenter(globalCell);
-            real_t y_normalized = std::max(0.0, std::min(1.0, 
-                (cellCenter[1] - domain.yMin()) / domainHeight));
+            real_t y_normalized = real_t(globalCell.y()) / real_t(yCells - 1);
+            y_normalized = std::max(0.0, std::min(1.0, y_normalized));
             
             // Linear temperature profile
             real_t temp = coldTemperature + (hotTemperature - coldTemperature) * y_normalized;
             temperature->get(localCell) = temp;
             
-            // INITIAL COUETTE PROFILE for faster convergence
-            real_t u_x = wallVelocity * y_normalized;
-            velocity->get(localCell, 0) = u_x;
+            // SIMPLE initial velocity - let LBM develop the profile
+            velocity->get(localCell, 0) = 0.0;  // Start from rest
[…]
