### 2025-09-23 a5b2653
Apply pressure gradient method for driving force
