### 2025-09-24 273a7de
Add grid independent pressure gradient force
