+++ apps/CouetteFlowSimulation.cpp
@@ -74,7 +74,7 @@ void initializeFields(const shared_ptr<StructuredBlockForest>& blocks,
     }
 }
 
-// Function to apply boundary conditions at proper locations
+// Function to apply boundary conditions
 void applyBoundaryConditions(const shared_ptr<StructuredBlockForest>& blocks,
                             const shared_ptr<TopWallBC>& topWallBC,
                             const shared_ptr<BottomWallBC>& bottomWallBC) {
@@ -106,21 +106,11 @@ void applyBoundaryConditions(const shared_ptr<StructuredBlockForest>& blocks,
     }
 }
 
-// Apply constant force to drive Couette flow
-void applyDrivingForce(const shared_ptr<StructuredBlockForest>& blocks,
-                      BlockDataID forceId, real_t forceValue) {
-    for (auto block = blocks->begin(); block != blocks->end(); ++block) {
-        VectorField* force = block->getData<VectorField>(forceId);
-        
-        for (auto cell = force->beginXYZ(); cell != force->end(); ++cell) {
-            Cell localCell = cell.cell();
-            force->get(localCell, 0) = forceValue;  // Constant driving force
-            force->get(localCell, 1) = 0.0;
-            force->get(localCell, 2) = 0.0;
-        }
-    }
-}
-
+/* ============================================================================
+ * This function applies the external driving force required for LBM Couette
+ * flow simulation. Without this force, direct velocity boundary conditions
+ * create unphysical negative velocities in the interior domain.
+ * ============================================================================ */
 // Grid-independent pressure gradient force
 void applyPressureGradientForce(const shared_ptr<StructuredBlockForest>& blocks,
                                BlockDataID forceId, real_t pressureGradient) {
@@ -130,10 +120,12 @@ void applyPressureGradientForce(const shared_ptr<StructuredBlockForest>& blocks,
         for (auto cell = force->beginXYZ(); cell != force->end(); ++cell) {
             Cell localCell = cell.cell();
             
-            // Constant pressure gradient force (grid-independent)
-            force->get(localCell, 0) = pressureGradient;  // Force per unit mass
-            force->get(localCell, 1) = 0.0;               // No Y-force
-            force->get(localCell, 2) = 0.0;               // No Z-force
+            // Constant pressure gradient force in flow direction (x) (grid-independent)
+            // This provides the energy needed to sustain Couette flow
+            // against LBM's natural tendency toward zero-velocity equilibrium
+            force->get(localCell, 0) = pressureGradient;  // Force per unit mass F_x = -dP/dx / rho
+            force->get(localCell, 1) = 0.0;               // No Y-force (wall normal)
+            force->get(localCell, 2) = 0.0;               // No Z-force (spanwise)
         }
     }
 }
@@ -143,18 +135,13 @@ int main(int argc, char** argv) {
     
     // Parameters
     const uint_t timesteps = 10000;
-    const uint_t vtkWriteFrequency = -50;
+    const uint_t vtkWriteFrequency = 50;
     const real_t wallVelocity = 0.02;
     const real_t hotTemperature = 600.0;
     const real_t coldTemperature = 300.0;
     const uint_t xBlocks = 1, yBlocks = 1, zBlocks = 1;
     const uint_t xCells = 128, yCells = 64, zCells = 32;
     const real_t refViscosity = 0.05;  // REDUCED viscosity for stability
[…]
