### 2025-09-24 40fd9b7
Check convergence and analytical solution
