+++ apps/CouetteFlowSimulation.cpp
@@ -1,7 +1,7 @@
 //======================================================================================================================
 //! \file CouetteFlowThermalSimulation.cpp - USING GENERATED BOUNDARIES
 //! \author Rahil Doshi
-//! \brief Thermal Couette flow using generated boundary sweeps
+//! \brief Thermal Couette flow using generated boundary sweeps with Reynolds number calculation and convergence monitoring
 //======================================================================================================================
 
 #include "blockforest/Initialization.h"
@@ -27,6 +27,13 @@ using namespace walberla;
 typedef GhostLayerField<real_t, 1> ScalarField;
 typedef GhostLayerField<real_t, 3> VectorField;
 typedef GhostLayerField<real_t, 19> PdfField;
+
+// Global variables for convergence monitoring and early termination
+static real_t previousMiddleVelocity = 0.0;
+static uint_t convergenceCheckInterval = 1000;
+static int convergedCount = 0;
+static bool earlyTerminationEnabled = true;
+static bool terminateEarly = false;
 
 // Initialize fields with thermal Couette profile
 void initializeFields(const shared_ptr<StructuredBlockForest>& blocks,
@@ -130,11 +137,66 @@ void applyPressureGradientForce(const shared_ptr<StructuredBlockForest>& blocks,
     }
 }
 
+// Enhanced function to monitor convergence with early termination
+void checkConvergence(const shared_ptr<StructuredBlockForest>& blocks, 
+                     BlockDataID velocityId, uint_t timestep, uint_t xCells, uint_t yCells, uint_t zCells) {
+    
+    if (timestep % convergenceCheckInterval != 0) return;
+    
+    WALBERLA_ROOT_SECTION() {
+        for (auto block = blocks->begin(); block != blocks->end(); ++block) {
+            VectorField* velocity = block->getData<VectorField>(velocityId);
+            Cell middleCell(static_cast<cell_idx_t>(xCells/2), static_cast<cell_idx_t>(yCells/2), static_cast<cell_idx_t>(zCells/2));
+            
+            real_t currentMiddleVelocity = velocity->get(middleCell, 0);
+            real_t velocityChange = std::abs(currentMiddleVelocity - previousMiddleVelocity);
+            real_t relativeChange = previousMiddleVelocity != 0.0 ? (velocityChange / std::abs(previousMiddleVelocity)) * 100.0 : 100.0;
+            
+            WALBERLA_LOG_INFO_ON_ROOT("Timestep " << timestep << ": Middle velocity = " << currentMiddleVelocity 
+                                    << ", Change = " << velocityChange 
+                                    << " (" << relativeChange << "%)");
+            
+            // Check if converged (relative change < 0.01%)
+            if (relativeChange < 0.01 && timestep > 2000) {
+                WALBERLA_LOG_INFO_ON_ROOT("*** CONVERGENCE ACHIEVED at timestep " << timestep << " ***");
+                WALBERLA_LOG_INFO_ON_ROOT("Relative velocity change: " << relativeChange << "% < 0.01%");
+                
+                // Early termination logic - ONLY increment if not already terminated
+                if (earlyTerminationEnabled && !terminateEarly) {
+                    convergedCount++;
+                    if (convergedCount >= 3) {  // Converged for 3 consecutive checks
[…]
