### 2025-09-25 acab98e
Update boundary conditions
