+++ apps/CouetteFlowSimulation.cpp
@@ -1,7 +1,13 @@
 //======================================================================================================================
-//! \file CouetteFlowThermalSimulation.cpp - USING GENERATED BOUNDARIES
+//! \file CouetteFlowThermalSimulation.cpp
 //! \author Rahil Doshi
-//! \brief Thermal Couette flow using generated boundary sweeps with Reynolds number calculation and convergence monitoring
+//! \brief Thermal Couette flow simulation using LBM with direct velocity BCs
+//! \details This simulation models thermal Couette flow between two parallel plates
+//!          using the Lattice Boltzmann Method (LBM). The top plate moves at
+//!          a constant velocity while the bottom plate is stationary. A linear
+//!          temperature gradient is imposed between the plates. The simulation
+//!          uses direct velocity boundary conditions at the walls and a small
+//!          pressure gradient force to sustain the flow.
 //======================================================================================================================
 
 #include "blockforest/Initialization.h"
@@ -27,13 +33,6 @@ using namespace walberla;
 typedef GhostLayerField<real_t, 1> ScalarField;
 typedef GhostLayerField<real_t, 3> VectorField;
 typedef GhostLayerField<real_t, 19> PdfField;
-
-// Global variables for convergence monitoring and early termination
-static real_t previousMiddleVelocity = 0.0;
-static uint_t convergenceCheckInterval = 1000;
-static int convergedCount = 0;
-static bool earlyTerminationEnabled = true;
-static bool terminateEarly = false;
 
 // Initialize fields with thermal Couette profile
 void initializeFields(const shared_ptr<StructuredBlockForest>& blocks,
@@ -61,13 +60,14 @@ void initializeFields(const shared_ptr<StructuredBlockForest>& blocks,
             real_t temp = coldTemperature + (hotTemperature - coldTemperature) * y_normalized;
             temperature->get(localCell) = temp;
             
-            // SIMPLE initial velocity - let LBM develop the profile
-            velocity->get(localCell, 0) = 0.0;  // Start from rest
+            // Initialize linear velocity profile
+            velocity->get(localCell, 0) = wallVelocity * y_normalized;
             velocity->get(localCell, 1) = 0.0;
             velocity->get(localCell, 2) = 0.0;
 
             density->get(localCell) = 1.0;
             
+            // Initialize force
             force->get(localCell, 0) = 0.0;
             force->get(localCell, 1) = 0.0;
             force->get(localCell, 2) = 0.0;
@@ -79,36 +79,44 @@ void initializeFields(const shared_ptr<StructuredBlockForest>& blocks,
         CouetteFlowInit init(densityId, forceId, pdfFieldId, velocityId);
         init(&*block);
     }
+    
+    WALBERLA_LOG_INFO_ON_ROOT("Field initialization complete");
 }
 
-// Function to apply boundary conditions
-void applyBoundaryConditions(const shared_ptr<StructuredBlockForest>& blocks,
-                            const shared_ptr<TopWallBC>& topWallBC,
-                            const shared_ptr<BottomWallBC>& bottomWallBC) {
+// Boundary condition application
+void applySimpleCouetteBoundaries(const shared_ptr<StructuredBlockForest>& blocks,
+                                  BlockDataID velocityId, real_t wallVelocity) {
     
[…]
