+++ apps/CouetteFlowSimulation.cpp
@@ -41,12 +41,39 @@ void initializeFields(const shared_ptr<StructuredBlockForest>& blocks,
                      real_t hotTemperature, real_t coldTemperature, real_t wallVelocity,
                      uint_t yCells, real_t refViscosity) {
     
+    WALBERLA_LOG_INFO_ON_ROOT("=== FIELD INITIALIZATION DEBUG ===");
+    WALBERLA_LOG_INFO_ON_ROOT("Parameters received:");
+    WALBERLA_LOG_INFO_ON_ROOT("  - yCells: " << yCells);
+    WALBERLA_LOG_INFO_ON_ROOT("  - Hot temperature: " << hotTemperature);
+    WALBERLA_LOG_INFO_ON_ROOT("  - Cold temperature: " << coldTemperature);
+    WALBERLA_LOG_INFO_ON_ROOT("  - Wall velocity: " << wallVelocity);
+    WALBERLA_LOG_INFO_ON_ROOT("  - Reference viscosity: " << refViscosity);
+    
+    uint_t totalCellsInitialized = 0;
+    uint_t blocksProcessed = 0;
+    
     for (auto block = blocks->begin(); block != blocks->end(); ++block) {
+        blocksProcessed++;
+        WALBERLA_LOG_INFO_ON_ROOT("Processing block " << blocksProcessed);
+        
+        // Get field pointers
         VectorField* velocity = block->getData<VectorField>(velocityId);
         ScalarField* temperature = block->getData<ScalarField>(temperatureId);
         ScalarField* density = block->getData<ScalarField>(densityId);
         VectorField* force = block->getData<VectorField>(forceId);
         ScalarField* viscosity = block->getData<ScalarField>(viscosityId);
+        
+        WALBERLA_LOG_INFO_ON_ROOT("Field pointers obtained successfully");
+        WALBERLA_LOG_INFO_ON_ROOT("  - Velocity field: " << (velocity != nullptr ? "OK" : "NULL"));
+        WALBERLA_LOG_INFO_ON_ROOT("  - Temperature field: " << (temperature != nullptr ? "OK" : "NULL"));
+        WALBERLA_LOG_INFO_ON_ROOT("  - Density field: " << (density != nullptr ? "OK" : "NULL"));
+        WALBERLA_LOG_INFO_ON_ROOT("  - Force field: " << (force != nullptr ? "OK" : "NULL"));
+        WALBERLA_LOG_INFO_ON_ROOT("  - Viscosity field: " << (viscosity != nullptr ? "OK" : "NULL"));
+        
+        // Sample some cells for debugging
+        uint_t cellsInBlock = 0;
+        real_t minTemp = 1e10, maxTemp = -1e10;
+        real_t minVel = 1e10, maxVel = -1e10;
         
         for (auto cell = velocity->beginXYZ(); cell != velocity->end(); ++cell) {
             Cell localCell = cell.cell();
@@ -54,14 +81,15 @@ void initializeFields(const shared_ptr<StructuredBlockForest>& blocks,
             blocks->transformBlockLocalToGlobalCell(globalCell, *block);
             
             real_t y_normalized = real_t(globalCell.y()) / real_t(yCells - 1);
-            y_normalized = std::max(0.0, std::min(1.0, y_normalized));
+            //y_normalized = std::max(0.0, std::min(1.0, y_normalized));
             
             // Linear temperature profile
             real_t temp = coldTemperature + (hotTemperature - coldTemperature) * y_normalized;
             temperature->get(localCell) = temp;
             
             // Initialize linear velocity profile
-            velocity->get(localCell, 0) = wallVelocity * y_normalized;
[…]
