### 2025-10-09 645a797
Replace sfg-walberla (walberla_codegen) with sweepgen
