### 2025-10-09 518f526
Update code for sweepgen, use code generated boundaries
