+++ apps/CouetteFlowKernel.py
@@ -1,246 +1,0 @@
-#!/usr/bin/env python3
-"""
-CouetteFlowKernel.py - 3D Thermal Couette Flow
-Code generation script for Couette flow simulation with temperature-dependent viscosity
-"""
-
-import logging
-import sympy as sp
-import pystencils as ps
-from pathlib import Path
-from pystencilssfg import SourceFileGenerator
-from pystencils import SymbolCreator
-from walberla.codegen import Sweep
-from lbmpy import LBMConfig, Stencil, Method, create_lb_method, create_lb_update_rule, LBMOptimisation, ForceModel
-
-# Configure logging
-logging.basicConfig(level=logging.WARNING, format="%(asctime)s %(levelname)s %(name)s -> %(message)s")
-
-# Material properties integration flag
-use_material_library = False  # Enable materforge integration
-
-print("=== COUETTE FLOW CODE GENERATION DEBUG LOG ===")
-print(f"Starting code generation at {Path(__file__).resolve()}")
-print(f"Material library integration: {'ENABLED' if use_material_library else 'DISABLED'}")
-
-with SourceFileGenerator() as sfg:
-    data_type = "float64"
-    
-    # Field definitions
-    stencil = Stencil.D3Q19
-    dim = 3
-    Q = 19
-    
-    print(f"LBM Configuration: Stencil={stencil}, Dimensions={dim}, Q={Q}, Data type={data_type}")
-    
-    # Define all fields
-    f_pdfs, f_pdfs_tmp, f_density, f_temperature, f_viscosity, f_force, f_velocity = ps.fields(
-        f"f_pdfs({Q}), f_pdfs_tmp({Q}), f_density, f_temperature, f_viscosity, f_force({dim}), f_velocity({dim}): {data_type}[{dim}D]", 
-        layout='fzyx'
-    )
-    
-    print("DEBUG: Field definitions created successfully")
-    print(f"  - PDFs: {f_pdfs.name} ({Q} components)")
-    print(f"  - Velocity: {f_velocity.name} ({dim} components)")
-    print(f"  - Force: {f_force.name} ({dim} components)")
-    print(f"  - Scalars: density, temperature, viscosity (1 component each)")
-    
-    # Create symbol creator for parameters
-    s = SymbolCreator()
-    
-    # ===== MATERFORGE INTEGRATION =====
-    material_assignments = []
-    viscosity_expression = None
-    
-    print("\n=== MATERFORGE INTEGRATION DEBUG ===")
-    
-    if use_material_library:
-        try:
-            from materforge.parsing.api import create_material
-            
-            yaml_path = Path(__file__).parent / 'CouetteFlowMaterial_Updated.yaml'
-            print(f"DEBUG: Looking for YAML at: {yaml_path.resolve()}")
-            
-            if yaml_path.exists():
-                print(f"Loading material from: {yaml_path}")
-                
-                mat = create_material(yaml_path=yaml_path, dependency=f_temperature.center(), enable_plotting=False)
-                print(f"DEBUG: Material has dynamic_viscosity: {hasattr(mat, 'dynamic_viscosity')}")
-                
-                if hasattr(mat, 'dynamic_viscosity'):
-                    print(f"DEBUG: Dynamic viscosity expression: {mat.dynamic_viscosity}")
-                
-                material_assignments.extend([
-                    ps.Assignment(s.density_mat, 1.0),  # Material density
-                    ps.Assignment(s.viscosity_mat, mat.dynamic_viscosity), 
-                ])
-                
-                viscosity_expression = s.viscosity_mat
[…]
