### 2025-10-10 eb4248e
Replace sfg-walberla (walberla_codegen) with sweepgen
