### 2025-10-10 25dc682
docs: separate BSD and GPL licenses for core library and apps
