+++ apps/README.md
@@ -1,0 +1,10 @@
+# MaterForge Applications
+
+This directory contains demonstration applications showing MaterForge integration with waLBerla and pystencils.
+
+## License
+All files in this directory (except the `walberla/` submodule) are licensed under GPL-3.0-or-later.
+See [LICENSE](LICENSE) for details.
+
+The `walberla/` subdirectory is a Git submodule containing waLBerla (GPLv3), maintained separately at:
+https://i10git.cs.fau.de/walberla/walberla
