+++ pyproject.toml
@@ -95,7 +95,7 @@ materforge-validate = "materforge.parsing.api:validate_yaml_file"
 [tool.setuptools.packages.find]
 where = ["src"]
 include = ["materforge*"]
-exclude = ["tests*", "apps*", "examples*", "*materforge_plots*"]
+exclude = ["tests*", "apps*", "examples*", "*materforge_plots*", "benchmarking-scripts*"]
 
 [tool.setuptools.package-data]
 materforge = [
