+++ CITATION.cff
@@ -5,7 +5,7 @@ authors:
     given-names: "Rahil Miten"
     email: "rahil.doshi@fau.de"
     affiliation: "Friedrich-Alexander-Universität Erlangen-Nürnberg"
-    orcid: "https://orcid.org/0009-0008-3570-9841"
+    orcid: "0009-0008-3570-9841"
 title: "MaterForge: Materials Formulation Engine with Python"
 version: "0.6.0"
 date-released: "2025-08-20"
