+++ src/materforge/core/materials.py
@@ -1,3 +1,6 @@
+# SPDX-FileCopyrightText: 2025 Rahil Miten Doshi, Friedrich-Alexander-Universität Erlangen-Nürnberg
+# SPDX-License-Identifier: BSD-3-Clause
+
 import logging
 from dataclasses import dataclass, field
 from typing import Dict, List, Optional, Tuple, Union
