+++ src/materforge/visualization/__init__.py
@@ -1,3 +1,6 @@
+# SPDX-FileCopyrightText: 2025 Rahil Miten Doshi, Friedrich-Alexander-Universität Erlangen-Nürnberg
+# SPDX-License-Identifier: BSD-3-Clause
+
 """Visualization tools for material properties."""
 
 from .plotters import PropertyVisualizer
