+++ src/materforge/parsing/io/__init__.py
@@ -1,3 +1,6 @@
+# SPDX-FileCopyrightText: 2025 Rahil Miten Doshi, Friedrich-Alexander-Universität Erlangen-Nürnberg
+# SPDX-License-Identifier: BSD-3-Clause
+
 """Input/output operations for data files."""
 
 from .data_handler import load_property_data
