### 2025-10-10 c2e4989
Add analytical solution script for temperature-dependent viscosity Couette flow
