+++ apps/compute_analytical_couette.py
@@ -1,0 +1,133 @@
+#!/usr/bin/env python3
+"""
+compute_analytical_couette.py - Compute analytical solution for variable viscosity Couette flow
+"""
+
+import numpy as np
+import matplotlib.pyplot as plt
+from scipy.integrate import cumulative_trapezoid
+import sympy as sp
+from materforge.parsing.api import create_material
+from pathlib import Path
+
+# Parameters
+T_bottom = 300.0  # K
+T_top = 600.0     # K
+u_max = 0.025
+H = 32  # Number of cells in z-direction
+
+# Temperature profile (linear): Create temperature array
+z = np.linspace(0, H, 1000)  # 1000 points from bottom to top
+T = T_bottom + (T_top - T_bottom) * z / H  # Linear temperature
+
+# Load viscosity model from MaterForge
+print("Loading viscosity model from MaterForge...")
+yaml_path = Path(__file__).parent / 'CouetteFlowMaterial_Updated.yaml'
+
+if yaml_path.exists():
+    # Create symbolic variable for temperature
+    T_sym = sp.Symbol('T')
+    
+    # Load material properties
+    mat = create_material(yaml_path=yaml_path, dependency=T_sym, enable_plotting=False)
+    
+    if hasattr(mat, 'dynamic_viscosity'):
+        nu_expr = mat.dynamic_viscosity
+        print(f"Symbolic viscosity expression:\n{nu_expr}\n")
+        
+        # Convert symbolic expression to numerical function using lambdify
+        # This allows evaluation on NumPy arrays
+        nu_func = sp.lambdify(T_sym, nu_expr, modules='numpy')
+    else:
+        print("Warning: No dynamic_viscosity attribute found, using constant viscosity")
+        nu_func = lambda T: 1.0/6.0
+else:
+    print(f"Warning: Material file {yaml_path} not found, using constant viscosity")
+    nu_func = lambda T: 1.0/6.0
+
+# Evaluate viscosity at all temperature points
+nu_z = nu_func(T)
+
+print(f"Viscosity evaluation:")
+print(f"  At T={T_bottom}K: nu = {nu_func(T_bottom):.8f}")
+print(f"  At T={T_top}K: nu = {nu_func(T_top):.8f}")
+print(f"  Viscosity variation: {(nu_func(T_bottom)/nu_func(T_top) - 1)*100:.2f}%\n")
+
+# Compute velocity profile: u(z) = U_wall * int_0^z (1/nu(z')) dz' / int_0^H (1/nu(z')) dz'
+integrand = 1.0 / nu_z  # Compute integrand (1/viscosity)
+# Numerical integration (cumulative trapezoid)
+integral_z = cumulative_trapezoid(integrand, z, initial=0)
+integral_H = integral_z[-1]  # Total integral from 0 to H
+
+# Velocity profile
+u_analytical = u_max * integral_z / integral_H
+
+# Linear profile (constant viscosity for comparison)
+u_linear = u_max * z / H
+
+# Compute deviation
+max_deviation = np.max(np.abs(u_analytical - u_linear))
+relative_deviation = max_deviation / u_max * 100
+
+print(f"Analytical solution statistics:")
+print(f"  Maximum deviation from linear: {max_deviation:.6e}")
+print(f"  Relative deviation: {relative_deviation:.2f}%")
+
+# Plot comparison
+plt.figure(figsize=(12, 5))
+
+# Subplot 1: Velocity profiles
+plt.subplot(1, 2, 1)
[…]
