### 2025-10-10 343e2e4
Integrate analytical solution into Couette flow codegen
