+++ apps/CouetteFlowSweeps.py
@@ -1,15 +1,12 @@
 #!/usr/bin/env python3
 
 """
-CouetteFlowKernel.py - 3D Thermal Couette Flow
-Code generation script for Couette flow simulation with temperature-dependent viscosity
+CouetteFlowSweeps.py - 3D Thermal Couette Flow with integrated analytical solution
 """
 
 import logging
 from pathlib import Path
-
 from argparse import ArgumentParser
-from dataclasses import replace
 
 from lbmpy import (
     LBStencil,
@@ -26,6 +23,8 @@ from lbmpy.macroscopic_value_kernels import macroscopic_values_setter
 
 import sympy as sp
 import pystencils as ps
+import numpy as np
+from scipy.integrate import cumulative_trapezoid
 
 from pystencilssfg import SourceFileGenerator
 from sweepgen import Sweep, get_build_config
@@ -43,6 +42,7 @@ use_materforge = False
 
 with SourceFileGenerator(keep_unknown_argv=True) as sfg:
     sfg.namespace("CouetteFlow::gen")
+    
     parser = ArgumentParser()
     parser.add_argument("-t", "--target", choices=["cpu", "gpu"], default="cpu")
     args = parser.parse_args(sfg.context.argv)
@@ -53,12 +53,13 @@ with SourceFileGenerator(keep_unknown_argv=True) as sfg:
     data_type = "float64"
     stencil = LBStencil(Stencil.D3Q19)
     D, Q = stencil.D, stencil.Q
-    # print(f"LBM Configuration: Stencil={stencil}, Dimensions={D}, Q={Q}, Data type={data_type}")
 
     # Symbolic parameters
-    nu, u_max = sp.symbols("nu, u_max")
-    T_top, T_bottom = sp.symbols("T_top, T_bottom")  # Temperature bounds
+    nu = sp.Symbol("nu")
+    u_max = sp.Symbol("u_max")
+    T_top, T_bottom = sp.symbols("T_top, T_bottom")
 
+    # Define fields
     f_pdfs, f_pdfs_tmp, f_density, f_temperature, f_viscosity, f_velocity = ps.fields(
         f"f_pdfs({Q}), f_pdfs_tmp({Q}), f_density, f_temperature, f_viscosity, f_velocity({D}): {data_type}[{D}D]",
         layout='fzyx'
@@ -66,29 +67,93 @@ with SourceFileGenerator(keep_unknown_argv=True) as sfg:
     
     const_nu = 1./6.  # Constant viscosity if MaterForge is not used
 
-    # Viscosity handling
+    analytical_velocity_expr = None
+    
     if use_materforge:
         print("Code generation: Using temperature-dependent viscosity from MaterForge")
         try:
             from materforge.parsing.api import create_material
-            from pathlib import Path
-
+            
             yaml_path = Path(__file__).parent / 'CouetteFlowMaterial_Updated.yaml'
             
             if yaml_path.exists():
+                # Create MaterForge material with temperature field dependency
                 mat = create_material(yaml_path=yaml_path, dependency=f_temperature.center(), enable_plotting=False)
-                nu_expr = mat.dynamic_viscosity if hasattr(mat, 'dynamic_viscosity') else const_nu
-                print(f"MaterForge viscosity expression loaded: {nu_expr}")
+                
+                if hasattr(mat, 'dynamic_viscosity'):
+                    nu_expr = mat.dynamic_viscosity
+                    print(f"MaterForge viscosity expression loaded: {nu_expr}")
+                    # ========================================================
+                    # COMPUTE ANALYTICAL SOLUTION
+                    # ========================================================
+                    print("Computing analytical solution for error calculation...")
+                    
+                    # Extract the symbolic temperature variable from nu_expr
[…]
