### 2025-10-13 01b1bf2
Add benchmarking to Couette flow simulation
