+++ apps/CouetteFlow.cpp
@@ -1,8 +1,6 @@
 //======================================================================================================================
-//
 //! \file CouetteFlow.cpp
 //! \author Rahil Doshi
-//
 //======================================================================================================================
 
 #include <iostream>
@@ -14,8 +12,10 @@
 #include "timeloop/all.h"
 #include "blockforest/communication/UniformBufferedScheme.h"
 #include "field/communication/StencilRestrictedPackInfo.h"
+//#include "field/communication/PackInfo.h"
 #include "domain_decomposition/SharedSweep.h"
 #include "walberla/experimental/Sweep.hpp"
+#include "core/timing/TimingPool.h"
 
 #include "gen/CouetteFlowSweeps.hpp"
 
@@ -40,6 +40,18 @@ void run(int argc, char **argv)
    auto config = env.config();
    auto blocks = blockforest::createUniformBlockGridFromConfig(config);
 
+   // Read domain configuration for performance metrics
+   Config::BlockHandle domainSetup = config->getBlock("DomainSetup");
+   Vector3<uint_t> numBlocks = domainSetup.getParameter<Vector3<uint_t>>("blocks");
+   Vector3<uint_t> cellsPerBlock = domainSetup.getParameter<Vector3<uint_t>>("cellsPerBlock");
+   
+   // Calculate total domain size
+   const uint_t xCells = numBlocks[0] * cellsPerBlock[0];
+   const uint_t yCells = numBlocks[1] * cellsPerBlock[1];
+   const uint_t zCells = numBlocks[2] * cellsPerBlock[2];
+   const uint_t totalCells = xCells * yCells * zCells;
+   const uint_t numProcesses = uint_c(MPIManager::instance()->numProcesses());
+
    // Field setup: PDFs, density, velocity, temperature and viscosity
    BlockDataID pdfsId = field::addToStorage<PdfField_T>(blocks, "pdfs", real_c(0.0), field::fzyx, 1);
    BlockDataID rhoId = field::addToStorage<ScalarField_T>(blocks, "rho", real_c(1.0), field::fzyx, 0);
@@ -55,6 +67,13 @@ void run(int argc, char **argv)
    // const bool useMaterForge = simParams.getParameter<bool>("useMaterForge", false);
    const real_t T_bottom = simParams.getParameter<real_t>("T_bottom");
    const real_t T_top = simParams.getParameter<real_t>("T_top");
+
+   WALBERLA_LOG_INFO_ON_ROOT("=== Domain Configuration ===")
+   WALBERLA_LOG_INFO_ON_ROOT("Blocks: " << numBlocks[0] << " x " << numBlocks[1] << " x " << numBlocks[2])
+   WALBERLA_LOG_INFO_ON_ROOT("Cells per block: " << cellsPerBlock[0] << " x " << cellsPerBlock[1] << " x " << cellsPerBlock[2])
+   WALBERLA_LOG_INFO_ON_ROOT("Total domain: " << xCells << " x " << yCells << " x " << zCells << " = " << totalCells << " cells")
+   WALBERLA_LOG_INFO_ON_ROOT("Number of processes: " << numProcesses)
+   WALBERLA_LOG_INFO_ON_ROOT("Cells per process: " << totalCells / numProcesses)
 
    // Prepare sweep functors using code-generated classes
    auto setAnalytical = gen::Couette::SetAnalyticalSolution{blocks, rhoId, uId, channelVelocity};
@@ -80,6 +99,8 @@ void run(int argc, char **argv)
    // Set up ghost layer communication
[…]
