### 2025-10-13 6d00b71
Add CPU scaling benchmark for Couette Flow
