+++ apps/CouetteFlowCPUScaling.cpp
@@ -1,0 +1,335 @@
+//======================================================================================================================
+//! \file CouetteFlowCPUScaling.cpp
+//! \author Rahil Doshi
+//! \brief CPU scaling benchmark for Couette flow using .prm configuration
+//======================================================================================================================
+
+#include <iostream>
+#include <cmath>
+
+#include "core/all.h"
+#include "blockforest/all.h"
+#include "stencil/all.h"
+#include "field/all.h"
+#include "timeloop/all.h"
+#include "blockforest/communication/UniformBufferedScheme.h"
+#include "field/communication/StencilRestrictedPackInfo.h"
+#include "domain_decomposition/SharedSweep.h"
+#include "walberla/experimental/Sweep.hpp"
+#include "core/timing/TimingPool.h"
+
+#include "gen/CouetteFlowSweeps.hpp"
+
+namespace CouetteFlow
+{
+
+using namespace walberla;
+
+// Type definitions
+using ScalarField_T = field::GhostLayerField<real_t, 1>;
+using VectorField_T = field::GhostLayerField<real_t, 3>;
+
+using LbStencil = stencil::D3Q19;
+using PdfField_T = field::GhostLayerField<real_t, LbStencil::Q>;
+
+using CommScheme = blockforest::communication::UniformBufferedScheme<LbStencil>;
+using PdfsPackInfo = field::communication::StencilRestrictedPackInfo<PdfField_T, LbStencil>;
+
+// Automatic process decomposition
+std::tuple<uint_t, uint_t, uint_t> calculateProcessDecomposition(uint_t numProcesses) {
+    uint_t procs_x = 1, procs_y = 1, procs_z = 1;
+    uint_t base = uint_c(std::round(std::cbrt(real_c(numProcesses))));
+    
+    for (uint_t z = base; z >= 1; --z) {
+        if (numProcesses % z == 0) {
+            uint_t remaining = numProcesses / z;
+            uint_t base_xy = uint_c(std::round(std::sqrt(real_c(remaining))));
+            for (uint_t y = base_xy; y >= 1; --y) {
+                if (remaining % y == 0) {
+                    procs_x = remaining / y;
+                    procs_y = y;
+                    procs_z = z;
+                    goto found;
+                }
+            }
+        }
+    }
+    found:
+    return std::make_tuple(procs_x, procs_y, procs_z);
+}
+
+void run(int argc, char **argv)
+{
+   Environment env{argc, argv};
+   auto config = env.config();
+   
+   const uint_t numProcesses = uint_c(MPIManager::instance()->numProcesses());
+   
+   // Read parameters from config
+   Config::BlockHandle simParams = config->getBlock("Parameters");
+   const std::string scalingTest = simParams.getParameter<std::string>("scalingTest", "none");
+   const uint_t problemSize = simParams.getParameter<uint_t>("problemSize", 64);
+   const real_t channelVelocity = simParams.getParameter<real_t>("u_max");
+   const real_t errorThreshold = simParams.getParameter<real_t>("errorThreshold");
+   const uint_t numTimesteps = simParams.getParameter<uint_t>("timesteps");
+   const real_t T_bottom = simParams.getParameter<real_t>("T_bottom");
[…]
