### 2025-10-13 c44584f
Add GPU scaling benchmark for Couette Flow
