+++ apps/CMakeLists.txt
@@ -90,7 +90,9 @@ sweepgen_venv_populate()
 
 # Build targets based on configuration
 if(BUILD_CPU_ONLY)
-    # CPU-only executables
+    # ============== CPU-ONLY EXECUTABLES ==============
+    
+    # Heat Equation executables
     add_executable(HeatEquationWithMaterial)
     target_sources(HeatEquationWithMaterial PRIVATE CodegenHeatEquationWithMaterial.cpp)
     target_link_libraries(HeatEquationWithMaterial PRIVATE
@@ -121,11 +123,13 @@ if(BUILD_CPU_ONLY)
             SCRIPTS HeatEquationKernelWithMaterialCPU.py
             FILE_EXTENSIONS .cpp .hpp)
 
+    # Couette Flow executables
     add_executable(CouetteFlowCodegen)
     target_sources(CouetteFlowCodegen PRIVATE CouetteFlow.cpp)
     target_link_libraries(CouetteFlowCodegen PRIVATE
-            walberla::blockforest walberla::boundary walberla::core walberla::experimental walberla::field 
-            walberla::lbm walberla::postprocessing walberla::stencil 
+            walberla::blockforest walberla::boundary walberla::core 
+            walberla::experimental walberla::field walberla::lbm 
+            walberla::postprocessing walberla::stencil 
             walberla::timeloop walberla::vtk walberla::sqlite
             MPI::MPI_CXX)
     walberla_generate_sources(CouetteFlowCodegen
@@ -143,13 +147,17 @@ if(BUILD_CPU_ONLY)
             SCRIPTS CouetteFlowSweeps.py
             FILE_EXTENSIONS .cpp .hpp)
 
-    message(STATUS "CPU executables will be built: HeatEquationWithMaterial, HeatEquationWithMaterialCPU, CodegenHeatEquationCPUScaling, CouetteFlowCodegen, CouetteFlowCPUScaling")
+    message(STATUS "CPU executables configured:")
+    message(STATUS "  - HeatEquationWithMaterial")
+    message(STATUS "  - HeatEquationWithMaterialCPU")
+    message(STATUS "  - CodegenHeatEquationCPUScaling")
+    message(STATUS "  - CouetteFlowCodegen")
+    message(STATUS "  - CouetteFlowCPUScaling")
 
 else()
+    # ============== GPU EXECUTABLES ==============
+    
     if(GPU_TYPE STREQUAL "HIP")
-        # find_package(MPI REQUIRED)
-        message(STATUS "MPI_CXX_INCLUDE_DIRS: ${MPI_CXX_INCLUDE_DIRS}")
-        
         # Get GTL library information from environment
         set(PE_MPICH_GTL_DIR_amd_gfx90a "$ENV{PE_MPICH_GTL_DIR_amd_gfx90a}")
         set(PE_MPICH_GTL_LIBS_amd_gfx90a "$ENV{PE_MPICH_GTL_LIBS_amd_gfx90a}")
@@ -167,7 +175,8 @@ else()
         message(STATUS "GTL_LIB_DIR: ${GTL_LIB_DIR}")
         message(STATUS "GTL_LIB_NAME: ${GTL_LIB_NAME}")
         
-        # Create executables with full MPI::MPI_CXX target
+        # ========== Heat Equation GPU Executables ==========
+        
         add_executable(HeatEquationWithMaterialGPU)
         target_sources(HeatEquationWithMaterialGPU PRIVATE CodegenHeatEquationWithMaterialGPU.cpp)
         target_link_libraries(HeatEquationWithMaterialGPU PRIVATE
@@ -225,16 +234,43 @@ else()
[…]
