### 2025-10-15 b5f0f8d
Fix script arguments for CouetteFlowGPUScaling
