### 2025-10-15 17f82e6
Replace sfg-walberla (walberla_codegen) with sweepgen
