### 2025-10-17 af6831b
WALBERLA_BUILD_EXAMPLES: OFF
