+++ apps/CMakeLists.txt
@@ -10,6 +10,7 @@ SET(WALBERLA_BUILD_TOOLS OFF CACHE BOOL "")
 SET(WALBERLA_BUILD_TUTORIALS OFF CACHE BOOL "")
 SET(WALBERLA_BUILD_SHOWCASES OFF CACHE BOOL "")
 SET(WALBERLA_BUILD_DOC OFF CACHE BOOL "")
+SET(WALBERLA_BUILD_EXAMPLES OFF CACHE BOOL "")
 SET(WALBERLA_BUILD_WITH_PYTHON ON CACHE BOOL "")
 # SET(WALBERLA_BUILD_WITH_CODEGEN OFF CACHE BOOL "")
 # SET(WALBERLA_BUILD_WITH_OPENMP ON CACHE BOOL "")
