### 2025-10-17 0da5b38
Add GPU communication as before function in timeloop
