### 2025-10-21 ea88948
Add GPU specific generator script
