+++ apps/CMakeLists.txt
@@ -257,12 +257,12 @@ else()
         set_property(TARGET CouetteFlowGPUScaling PROPERTY HIP_ARCHITECTURES "gfx90a")
         
         walberla_generate_sources(CouetteFlowGPUScaling
-                SCRIPTS CouetteFlowSweeps.py
+                SCRIPTS CouetteFlowSweepsGPU.py
                 SCRIPT_ARGS --target=gpu
                 FILE_EXTENSIONS .cpp .hpp)
                 
         set_source_files_properties(
-                ${CMAKE_CURRENT_BINARY_DIR}/_gen/CouetteFlowGPUScaling/gen/CouetteFlowSweeps.cpp
+                ${CMAKE_CURRENT_BINARY_DIR}/_gen/CouetteFlowGPUScaling/gen/CouetteFlowSweepsGPU.cpp
                 PROPERTIES LANGUAGE HIP)
 
         message(STATUS "HIP GPU executables configured:")
@@ -314,11 +314,11 @@ else()
                 MPI::MPI_CXX)
         set_property(TARGET CouetteFlowGPUScaling PROPERTY CUDA_SEPARABLE_COMPILATION ON)
         walberla_generate_sources(CouetteFlowGPUScaling
-                SCRIPTS CouetteFlowSweeps.py 
+                SCRIPTS CouetteFlowSweepsGPU.py 
                 SCRIPT_ARGS --target=gpu
                 FILE_EXTENSIONS .cu .hpp)
         set_source_files_properties(
-                ${CMAKE_CURRENT_BINARY_DIR}/_gen/CouetteFlowGPUScaling/gen/CouetteFlowSweeps.cu
+                ${CMAKE_CURRENT_BINARY_DIR}/_gen/CouetteFlowGPUScaling/gen/CouetteFlowSweepsGPU.cu
                 PROPERTIES LANGUAGE CUDA)
 
         message(STATUS "CUDA GPU executables configured:")
