### 2025-10-21 a6f14b5
Add GPU specific generator script
