+++ apps/CMakeLists.txt
@@ -257,7 +257,7 @@ else()
         set_property(TARGET CouetteFlowGPUScaling PROPERTY HIP_ARCHITECTURES "gfx90a")
         
         walberla_generate_sources(CouetteFlowGPUScaling
-                SCRIPTS CouetteFlowSweeps.py
+                SCRIPTS CouetteFlowSweepsGPU.py
                 SCRIPT_ARGS --target=gpu
                 FILE_EXTENSIONS .cpp .hpp)
                 
@@ -314,7 +314,7 @@ else()
                 MPI::MPI_CXX)
         set_property(TARGET CouetteFlowGPUScaling PROPERTY CUDA_SEPARABLE_COMPILATION ON)
         walberla_generate_sources(CouetteFlowGPUScaling
-                SCRIPTS CouetteFlowSweeps.py 
+                SCRIPTS CouetteFlowSweepsGPU.py 
                 SCRIPT_ARGS --target=gpu
                 FILE_EXTENSIONS .cu .hpp)
         set_source_files_properties(
