### 2025-10-22 75f644b
Add automated build script for CouetteFlow GPU on Lumi-G
