### 2025-10-22 b31b86f
Add automated single node run script for CouetteFlow GPU on Lumi-G
