### 2025-10-22 7c40d75
Add command-line parameter override support
