### 2025-10-22 2aa3028
Update single node run script with command-line parameter override
