+++ benchmarking-scripts/materforge/CouetteFlow/lumi-g/run_single_node.sh
@@ -11,26 +11,27 @@
 unset SLURM_EXPORT_ENV
 
 BUILD_DIR=""
-PRM_FILE="CouetteScaling.prm"
+SCALING_TYPE="weak"
+PROBLEM_SIZE=128
 
 usage() {
-    echo "Usage: $0 build_dir [prm_file]"
-    echo "  build_dir  The build directory (required)"
-    echo "  prm_file   Configuration file (optional, default: CouetteScaling.prm)"
+    echo "Usage: $0 build_dir [scaling_type] [problem_size]"
+    echo "  scaling_type: strong or weak (default: weak)"
+    echo "  problem_size: cells per GPU or total domain (default: 128)"
     exit 1
 }
 
 [ $# -eq 0 ] && usage
 BUILD_DIR=$(realpath $1)
-[ $# -ge 2 ] && PRM_FILE=$2
+[ $# -ge 2 ] && SCALING_TYPE=$2
+[ $# -ge 3 ] && PROBLEM_SIZE=$3
 
-SCRIPT_PATH="$(realpath $0)"
 JOB_DIR="$HOME/lss-rdm/jobs/$SLURM_JOBID/"
 mkdir -p ${JOB_DIR}
 cd ${JOB_DIR}
 
 cp ${BUILD_DIR}/build_*.log . 2>/dev/null || true
-cp ${BUILD_DIR}/${PRM_FILE} . 2>/dev/null || true
+cp ${BUILD_DIR}/*.prm . 2>/dev/null || true
 
 module load LUMI/24.03 partition/G buildtools/24.03 rocm/6.0.3 craype-accel-amd-gfx90a PrgEnv-cray
 module load cray-mpich
@@ -42,14 +43,16 @@ export MPICH_GPU_SUPPORT_ENABLED=1
 export ROCR_VISIBLE_DEVICES=0,1,2,3,4,5,6,7
 
 CMD="${BUILD_DIR}/CouetteFlowGPUScaling"
+PRM_FILE="${BUILD_DIR}/CouetteScaling.prm"
 
 [ ! -f "${CMD}" ] && { echo "Error: Executable not found"; exit 1; }
-[ ! -f "${BUILD_DIR}/${PRM_FILE}" ] && { echo "Error: .prm file not found"; exit 1; }
+[ ! -f "${PRM_FILE}" ] && { echo "Error: .prm file not found"; exit 1; }
 
 echo "========================================="
 echo "Single Node Couette Flow Benchmark"
-echo "Configuration: ${PRM_FILE}"
+echo "Scaling type: ${SCALING_TYPE}"
+echo "Problem size: ${PROBLEM_SIZE}"
 echo "Total GPUs: 8"
 echo "========================================="
 
-srun -n 8 --cpu-bind=${CPU_BIND} ${CMD} ${PRM_FILE}
+srun -n 8 --cpu-bind=${CPU_BIND} ${CMD} ${PRM_FILE} ${SCALING_TYPE} ${PROBLEM_SIZE}
