### 2025-10-24 0a8831c
Bump version to v0.6.1
