### 2025-10-24 a6bf519
Update CITATION.cff for v0.6.1
