+++ CITATION.cff
@@ -7,8 +7,8 @@ authors:
     affiliation: "Friedrich-Alexander-Universität Erlangen-Nürnberg"
     orcid: "0009-0008-3570-9841"
 title: "MaterForge: Materials Formulation Engine with Python"
-version: "0.6.0"
-date-released: "2025-08-20"
+version: "0.6.1"
+date-released: "2025-10-24"
 url: "https://i10git.cs.fau.de/rahil.doshi/materforge"
 repository-code: "https://i10git.cs.fau.de/rahil.doshi/materforge"
 license: "BSD-3-Clause"
