### 2025-10-24 952692d
Update setuptools_scm config to recognize v-prefixed tags
