+++ pyproject.toml
@@ -117,3 +117,4 @@ write_to = "src/materforge/_version.py"
 version_scheme = "python-simplified-semver"
 local_scheme = "node-and-date"
 fallback_version = "0.6.1"
+tag_regex = "^(?:v)?(?P<version>\\d+\\.\\d+\\.\\d+)$"
