### 2025-10-24 0298a90
Update CHANGELOG.md for v0.6.1 release
