+++ CHANGELOG.md
@@ -5,6 +5,41 @@ All notable changes to MaterForge will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.6.1] - 2025-10-24
+
+### Added
+- SPDX license headers to all Python source files in `src/materforge/`
+- SPDX (GPL-3.0-or-later) headers for application files in `apps/`
+- `apps/LICENSE` and `apps/README.md` describing separate GPL licensing
+- Tag regex support in `pyproject.toml` to handle `v`-prefixed tags
+- Expanded license section and summary table in `README.md`
+
+### Changed
+- Split BSD‑3‑Clause core license and GPLv3 examples licensing
+- Updated installation guide with submodule instructions
+- Excluded GPL apps from PyPI distribution
+
+### Fixed
+- `setuptools_scm` version detection for `v0.x`-style tags now returns the correct release version (0.6.1)
+
+## [0.6.0] - 2025-08-20
+
+### Added
+- Release automation improvements and documentation badge updates
+- Improved type annotations for dependency parameters (`Union[float, sp.Symbol]`)
+- Excluded `materforge_plots` from package distributions to reduce build size
+
+### Changed
+- Updated documentation for stable version release
+- Enhanced installation and usage documentation in `README.md`
+- Updated `CITATION.cff` metadata for improved scholarly citation support
+- Adjusted `pyproject.toml` for refined dependency management
+- Improved docstrings and code comments for maintainability
+
+### Fixed
+- Resolved redundant file inclusion issues in packaging
+- Minor fixes across documentation files and CI pipeline
+
 ## [0.5.6] - 2025-08-06
 
 ### Changed
