### 2025-11-10 6643277
Correct piecewise inverter boundary handling and slope detection
