+++ src/materforge/algorithms/piecewise_inverter.py
@@ -80,28 +80,100 @@ class PiecewiseInverter:
         self._validate_linear_only(piecewise_func, input_symbol)
         # Process each piece
         inverse_conditions = []
+        piece_boundaries = []  # Track (energy_value, temperature_boundary, expr, condition)
+        # First pass: collect all piece information
         for i, (expr, condition) in enumerate(piecewise_func.args):
             logger.debug("Processing piece %d: expr=%s, condition=%s", i + 1, expr, condition)
-            if condition == True:  # Final piece
-                inverse_expr = self._invert_linear_expression(expr, input_symbol, output_symbol)
-                inverse_conditions.append((inverse_expr, True))
-                logger.debug("Added final piece with universal condition")
+            if condition == True:  # Final piece (no condition)
+                # For final piece, we need to handle it specially
+                piece_boundaries.append({
+                    'index': i,
+                    'expr': expr,
+                    'condition': condition,
+                    'boundary_temp': float('inf'),
+                    'boundary_energy': float('inf'),
+                    'is_final': True,
+                    'slope': None
+                })
             else:
-                # Extract boundary and create inverse
+                # Extract temperature boundary
                 try:
-                    boundary = self._extract_boundary(condition, input_symbol)
-                    logger.debug("Extracted boundary: %.3f", boundary)
-                    inverse_expr = self._invert_linear_expression(expr, input_symbol, output_symbol)
+                    boundary_temp = self._extract_boundary(condition, input_symbol)
+                    logger.debug("Extracted boundary: %.3f", boundary_temp)
                     # Calculate energy at boundary for domain condition
-                    boundary_energy = float(expr.subs(input_symbol, boundary))
-                    inverse_conditions.append((inverse_expr, output_symbol < boundary_energy))
-                    logger.debug("Added piece %d: boundary_energy=%.3f", i + 1, boundary_energy)
+                    boundary_energy = float(expr.subs(input_symbol, boundary_temp))
+                    # Get slope to determine condition direction
+                    degree = sp.degree(expr, input_symbol)
+                    slope = None
+                    if degree == 1:
+                        coeffs = sp.Poly(expr, input_symbol).all_coeffs()
+                        slope = float(coeffs[0])
+                    piece_boundaries.append({
+                        'index': i,
+                        'expr': expr,
+                        'condition': condition,
+                        'boundary_temp': boundary_temp,
+                        'boundary_energy': boundary_energy,
+                        'is_final': False,
+                        'slope': slope
+                    })
+                    logger.debug("Boundary: T=%.3f -> E=%.3e (slope=%.6f)", boundary_temp, boundary_energy, slope if slope else 0)
                 except Exception as e:
                     logger.error("Error processing piece %d: %s", i + 1, e)
                     raise ValueError(f"Error processing piece {i + 1}: {str(e)}") from e
+        # Validate monotonicity
+        self._validate_monotonicity(piece_boundaries, input_symbol)
+        # Build inverse conditions
+        inverse_conditions = []
+        # Second pass: build inverse conditions with correct domain mapping
+        for i, piece_info in enumerate(piece_boundaries):
+            expr = piece_info['expr']
+            boundary_temp = piece_info['boundary_temp'] if not piece_info['is_final'] else None
+            boundary_energy = piece_info['boundary_energy']
+            is_final = piece_info['is_final']
+            slope = piece_info['slope']
+            inverse_expr = self._invert_linear_expression(expr, input_symbol, output_symbol, boundary_temp)
+            if is_final:
+                # Final piece: no upper bound
+                inverse_conditions.append((inverse_expr, True))
[…]
