### 2025-11-12 ef41a8a
Fix piecewise inverter final constant piece handling
