+++ tests/unit/test_algorithms/test_inversion.py
@@ -33,6 +33,7 @@ class TestPiecewiseInverter:
         )
         # Should not raise any exception
         PiecewiseInverter._validate_linear_only(piecewise_func, T)
+    
     def test_validate_linear_only_invalid(self):
         """Test validation fails for non-linear functions."""
         T = sp.Symbol('T')
@@ -66,10 +67,13 @@ class TestPiecewiseInverter:
         T = sp.Symbol('T')
         E = sp.Symbol('E')
         inverter = PiecewiseInverter()
-        # Test constant expression
+        # Test constant expression without boundary
         expr = sp.Float(1000)
-        inverse_expr = inverter._invert_linear_expression(expr, T, E)
+        inverse_expr = inverter._invert_linear_expression(expr, T, E, boundary_temp=None)
         assert inverse_expr == 1000.0
+        # Test constant expression with boundary
+        inverse_expr_with_boundary = inverter._invert_linear_expression(expr, T, E, boundary_temp=500.0)
+        assert inverse_expr_with_boundary == 500.0
 
     def test_create_energy_density_inverse(self, sample_aluminum_element):
         """Test creating energy density inverse for a material."""
@@ -98,7 +102,7 @@ class TestPiecewiseInverter:
         assert abs(recovered_temp - test_temp) < 1e-10
 
     def test_create_inverse_single_segment(self):
-        """Test creating inverse of single segment function - skip problematic case."""
+        """Test creating inverse of single segment function."""
         T = sp.Symbol('T')
         E = sp.Symbol('E')
         # Test with a simple linear expression directly
@@ -183,3 +187,224 @@ class TestPiecewiseInverter:
             # Use simplify to check mathematical equivalence, not structural equality
             difference = sp.simplify(inverse_expr - expected)
             assert difference == 0, f"Expected {expected}, got {inverse_expr}"
+
+    # ============================================================================
+    # Material-Specific Test Cases (Aluminum specific_enthalpy)
+    # ============================================================================
+
+    def test_material_specific_enthalpy_bounds_constant_constant(self):
+        """Test inverse of specific_enthalpy with bounds=[constant, constant].
+        
+        This represents a material property where:
+        - DOMAIN: T E [300K, 4500K]
+        - Lower bound: constant piece for T < 300K
+        - Upper bound: constant piece for T >= 4500K
+        """
+        T_C = sp.Symbol('T_C')
+        h = sp.Symbol('h')
+        
+        # Original piecewise function (Aluminum specific enthalpy)
+        specific_enthalpy = sp.Piecewise(
+            (119597.562064681, T_C < 300.0),
+            (661.839974761036*T_C - 78954.4303636301, T_C < 1672.410186589),
+            (3503.34591068727*T_C - 4831117.90285977, T_C < 1741.14311373293),
+            (847.6065746183*T_C - 207095.645993611, T_C < 4500.0),
+            (3607133.93978874, True)
+        )
+        
+        # Create inverse
+        inverse_h = PiecewiseInverter.create_inverse(specific_enthalpy, T_C, h)
+        assert isinstance(inverse_h, sp.Piecewise)
+        
+        # Test that the final piece returns T=4500 (boundary), not the constant
+        final_piece_expr = inverse_h.args[-1][0]  # Last piece expression
+        assert float(final_piece_expr) == 4500.0, \
+            f"Final piece should return T=4500, got {final_piece_expr}"
+        
+        # Test round-trip accuracy for representative temperatures ONLY within valid domain [300, 4500]
+        test_temps = [300, 500, 1000, 1500, 1700, 2000, 3000, 4500]
[…]
