### 2025-11-12 0675358
Update CITATION.cff for v0.6.3
