+++ CITATION.cff
@@ -7,8 +7,8 @@ authors:
     affiliation: "Friedrich-Alexander-Universität Erlangen-Nürnberg"
     orcid: "0009-0008-3570-9841"
 title: "MaterForge: Materials Formulation Engine with Python"
-version: "0.6.2"
-date-released: "2025-11-10"
+version: "0.6.3"
+date-released: "2025-11-12"
 url: "https://i10git.cs.fau.de/rahil.doshi/materforge"
 repository-code: "https://i10git.cs.fau.de/rahil.doshi/materforge"
 license: "BSD-3-Clause"
