### 2025-11-12 1b8c9d1
Update CHANGELOG.md for v0.6.3 release
