### 2025-11-13 8de0d16
Initialize velocity field to zero instead of analytical profile
