### 2025-11-13 1e34c65
Initialize velocity field to zero instead of analytical profile for CouetteFlowCPUScaling
