### 2025-11-18 c534777
Update yaml and prm configs for dynamic_viscosity for temperature ranges from 300-3000K
