+++ apps/CouetteFlowMaterial_Updated.yaml
@@ -87,7 +87,7 @@ properties:
   # ====================================================================================================
   
   dynamic_viscosity:
-    dependency: (300, 600, 31)  # Temperature range 300-600K, 31 points
+    dependency: (300, 3000, 31)  # Temperature range 300-600K, 31 points
     equation: 0.16667 * exp(-0.0005 * (T - 300.0))  # LBM-scaled viscosity
     bounds: [constant, constant]
     regression:
