### 2025-11-18 2657833
Switch between materforge property and constant, dynamic vtk output
