+++ apps/CouetteFlowCPUScaling.cpp
@@ -23,6 +23,8 @@
 namespace CouetteFlow
 {
 
+constexpr bool use_materForge = true;
+
 using namespace walberla;
 
 // Type definitions
@@ -178,13 +180,20 @@ void run(int argc, char **argv)
       initializePdfs(&b);
    }
    
-   // Stream-collide sweep
-   auto streamCollide = makeSharedSweep(
-      std::make_shared<gen::Couette::StreamCollide>(
-         //rhoId, pdfsId, uId, viscId  // constant viscosity
-         rhoId, pdfsId, tempId, uId, viscId  // temperature-dependent viscosity
-      )
-   );
+   // Create streamCollide based on compile-time constant
+   auto streamCollide = [&]() {
+      if constexpr (use_materForge) {
+         WALBERLA_LOG_INFO_ON_ROOT("Using temperature-dependent viscosity (MaterForge)")
+         return makeSharedSweep(std::make_shared<gen::Couette::StreamCollide>(
+               rhoId, pdfsId, tempId, uId, viscId  // temperature-dependent viscosity
+         ));
+      } else {
+         WALBERLA_LOG_INFO_ON_ROOT("Using constant viscosity")
+         return makeSharedSweep(std::make_shared<gen::Couette::StreamCollide>(
+            rhoId, pdfsId, uId, viscId  // constant viscosity
+         ));
+      }
+   }();
    
    // Communication
    CommScheme comm{blocks};
@@ -220,10 +229,16 @@ void run(int argc, char **argv)
    const uint_t vtkWriteFrequency = outputParams.getParameter<uint_t>("vtkWriteFrequency", 0);
    
    if (vtkWriteFrequency > 0) {
-      auto vtkOutput = vtk::createVTKOutput_BlockData(*blocks, "vtk", vtkWriteFrequency, 0, false, 
-                                                       "vtk_out_couette_cpu", "simulation_step", 
-                                                       false, true, true, false, 0);
-      
+      std::string vtkName = "cf_cpu";
+      if constexpr (use_materForge) {
+         vtkName += "_mftempdep";
+      } else {
+         vtkName += "_mfconst";
+      }
+      std::string vtkOutputDir = vtkName + "_out";
+      auto vtkOutput = vtk::createVTKOutput_BlockData(*blocks, vtkName, vtkWriteFrequency, 0, false, vtkOutputDir,
+                                                       "simulation_step", false, true, true, false, 0);
+
       auto densityWriter = make_shared<field::VTKWriter<ScalarField_T>>(rhoId, "density");
       vtkOutput->addCellDataWriter(densityWriter);
       
