### 2025-11-18 4784dfd
Initialize velocity field to zero instead of analytical profile for CouetteFlowGPUScaling, dynamic vtk output
