### 2025-11-18 60b36e0
Update lumi.sh
