### 2025-11-18 e4e0d11
Correct arguments for temperature depenedent case
