### 2025-11-18 185f540
Use calculated block grid instead of config file
