### 2025-11-19 51b56cb
Unify cpu/gpu code
