### 2025-11-19 f7c33e2
Unify cpu/gpu code
