+++ apps/CouetteFlowScaling.cpp
@@ -1,0 +1,445 @@
+//======================================================================================================================
+//! \file CouetteFlowScaling.cpp
+//! \author Rahil Doshi
+//! \brief Unified scaling benchmark for Couette flow (CPU/GPU)
+//!
+//! Usage: ./CouetteFlowScaling config.prm
+//!
+//! Bash scripts override parameters via command line:
+//!   -DomainSetup.blocks <x,y,z>
+//!   -DomainSetup.cellsPerBlock <x,y,z>
+//!   -Parameters.timesteps <N>
+//======================================================================================================================
+
+#include <string>
+#include <cmath>
+#include "core/all.h"
+#include "blockforest/all.h"
+#include "stencil/all.h"
+#include "field/all.h"
+#include "timeloop/all.h"
+#include "blockforest/communication/UniformBufferedScheme.h"
+#include "field/communication/StencilRestrictedPackInfo.h"
+#include "domain_decomposition/SharedSweep.h"
+#include "walberla/experimental/Sweep.hpp"
+#include "core/timing/TimingPool.h"
+
+// GPU support (if available)
+#ifdef WALBERLA_BUILD_WITH_GPU_SUPPORT
+#include "gpu/AddGPUFieldToStorage.h"
+#include "gpu/DeviceSelectMPI.h"
+#include "gpu/FieldCopy.h"
+#include "gpu/HostFieldAllocator.h"
+#include "gpu/communication/UniformGPUScheme.h"
+#include "gpu/communication/MemcpyPackInfo.h"
+#include "gen/CouetteFlowSweepsGPU.hpp"
+#else
+#include "gen/CouetteFlowSweeps.hpp"
+#endif
+
+namespace CouetteFlow
+{
+
+constexpr bool use_materForge = true;
+
+using namespace walberla;
+
+// Type definitions
+using ScalarField_T = field::GhostLayerField<real_t, 1>;
+using VectorField_T = field::GhostLayerField<real_t, 3>;
+
+using LbStencil = stencil::D3Q19;
+using PdfField_T = field::GhostLayerField<real_t, LbStencil::Q>;
+
+#ifdef WALBERLA_BUILD_WITH_GPU_SUPPORT
+using CommScheme = gpu::communication::UniformGPUScheme<LbStencil>;
+using GPUField = gpu::GPUField<real_t>;
+#else
+using CommScheme = blockforest::communication::UniformBufferedScheme<LbStencil>;
+#endif
+using PdfsPackInfo = field::communication::StencilRestrictedPackInfo<PdfField_T, LbStencil>;
+
+//======================================================================
+// MAIN FUNCTION
+//======================================================================
+void run(int argc, char **argv)
+{
+    Environment env{argc, argv};
+    auto cfgFile = env.config();
+    
+    if (!cfgFile) {
+        WALBERLA_ABORT("Usage: " << argv << " path-to-configuration-file\n");
+    }
+    
+#ifdef WALBERLA_BUILD_WITH_GPU_SUPPORT
+    gpu::selectDeviceBasedOnMpiRank();
+    WALBERLA_LOG_INFO_ON_ROOT("GPU support enabled");
+#else
+    WALBERLA_LOG_INFO_ON_ROOT("CPU-only build");
+#endif
+    
+    // Log build information
+    WALBERLA_LOG_INFO_ON_ROOT(*cfgFile);
+    
+    const uint_t numProcesses = uint_c(MPIManager::instance()->numProcesses());
+    
+    //======================================================================
+    // READ CONFIGURATION - All parameters from .prm file
+    // Bash scripts override via: -DomainSetup.blocks <x,y,z>
+    //======================================================================
+    
[…]
