### 2025-11-19 0ada729
Match StreamCollide arguments for gpu with and without materforge usage
