### 2025-11-20 dcf4b0c
WIP: reorganize apps structure
