### 2025-11-25 5e7b004
Disable GPU-aware MPI
