### 2025-11-26 cbed1ac
Replace if constexpr with preprocessor #if directives
