### 2025-12-08 6179eed
Replace if constexpr with preprocessor #if directives
