### 2025-12-08 2ea6d50
Unify cpu/gpu code for couette flow
