+++ apps/CouetteFlowSweepsGPU.py
@@ -1,234 +1,0 @@
-#!/usr/bin/env python3
-
-"""
-CouetteFlowSweepsGPU.py - 3D Thermal Couette Flow with integrated analytical solution
-"""
-
-import logging
-from pathlib import Path
-from argparse import ArgumentParser
-
-from lbmpy import (
-    LBStencil,
-    Stencil,
-    LBMConfig,
-    LBMOptimisation,
-    Method,
-    relaxation_rate_from_lattice_viscosity,
-    create_lb_method,
-    create_lb_update_rule
-)
-from lbmpy.boundaries import NoSlip, UBB
-from lbmpy.macroscopic_value_kernels import macroscopic_values_setter
-
-import sympy as sp
-import pystencils as ps
-import numpy as np
-from scipy.integrate import cumulative_trapezoid
-
-from pystencilssfg import SourceFileGenerator
-from sweepgen import Sweep, get_build_config
-from sweepgen.boundaries import GenericBoundary
-from sweepgen.symbolic import cell, domain
-from sweepgen.prefabs import LbmBulk
-from sweepgen.build_config import DEBUG
-
-DEBUG.use_cpu_default()
-logging.basicConfig(level=logging.WARNING, format="%(asctime)s %(levelname)s %(name)s -> %(message)s")
-
-print(f"Starting code generation at {Path(__file__).resolve()}")
-
-use_materforge = True
-
-with SourceFileGenerator(keep_unknown_argv=True) as sfg:
-    sfg.namespace("CouetteFlow::gen")
-    
-    parser = ArgumentParser()
-    parser.add_argument("-t", "--target", choices=["cpu", "gpu"], default="cpu")
-    args = parser.parse_args(sfg.context.argv)
-
-    build_cfg = get_build_config(sfg)
-    build_cfg.target = ps.Target[args.target.upper()]
-    
-    data_type = "float64"
-    stencil = LBStencil(Stencil.D3Q19)
-    D, Q = stencil.D, stencil.Q
-
-    # Symbolic parameters
-    nu = sp.Symbol("nu")
-    u_max = sp.Symbol("u_max")
-    T_top, T_bottom = sp.symbols("T_top, T_bottom")
-
-    # Define fields
-    f_pdfs, f_pdfs_tmp, f_density, f_temperature, f_viscosity, f_velocity = ps.fields(
-        f"f_pdfs({Q}), f_pdfs_tmp({Q}), f_density, f_temperature, f_viscosity, f_velocity({D}): {data_type}[{D}D]",
-        layout='fzyx'
-    )
-    
-    const_nu = 1./6.  # Constant viscosity if MaterForge is not used
-
-    analytical_velocity_expr = None
-    
-    if use_materforge:
-        print("Code generation: Using temperature-dependent viscosity from MaterForge")
-        try:
-            from materforge.parsing.api import create_material
-            
-            yaml_path = Path(__file__).parent / 'CouetteFlowMaterial_Updated.yaml'
-            
-            if yaml_path.exists():
-                # Create MaterForge material with temperature field dependency
-                mat = create_material(yaml_path=yaml_path, dependency=f_temperature.center(), enable_plotting=False)
-                
-                if hasattr(mat, 'dynamic_viscosity'):
-                    nu_expr = mat.dynamic_viscosity
-                    print(f"MaterForge viscosity expression loaded: {nu_expr}")
-                    # ========================================================
-                    # COMPUTE ANALYTICAL SOLUTION
-                    # ========================================================
-                    print("Computing analytical solution for error calculation...")
-                    
-                    # Extract the symbolic temperature variable from nu_expr
-                    temp_symbols = [s for s in nu_expr.free_symbols if 'temperature' in str(s)]
-                    if temp_symbols:
[…]
