### 2026-01-22 b92a100
lumi.sh: LUMI/25.03, ROCm 6.3.4, CCE 19 (post-Jan 2026 maintenance)
