+++ lumi.sh
@@ -1,87 +1,116 @@
 #!/bin/bash
 
-# LUMI setup for materforge project
+################################################################################
+# LUMI Setup Script for MaterForge
+# Post-January 2026 Maintenance (ROCm 6.3.4, CCE 19.0.0, LUMI/25.03)
+#
+# USAGE:
+#   source ./lumi.sh cpu      # Load CPU environment (default)
+#   source ./lumi.sh gpu      # Load GPU environment
+#
+# This script must be SOURCED, not executed, to persist environment variables
+################################################################################
 
-# CPU Setup - Default execution
+set -o pipefail
+
 lumi_cpu() {
-    module purge
+    # CPU partition with GNU compiler
     module --force purge
-    module load LUMI/24.03 partition/C PrgEnv-gnu buildtools/24.03 cray-python/3.11.7
+    module load LUMI/25.03 partition/C PrgEnv-gnu
+    module load cray-python
     
-    # Activate virtual environment
+    # Disable GPU MPI for CPU builds
+    unset MPICH_GPU_SUPPORT_ENABLED
+
+    # Activate MaterForge virtual environment
     if [[ -f /projappl/project_465002382/venvs/materforge/bin/activate ]]; then
         source /projappl/project_465002382/venvs/materforge/bin/activate
     fi
     
+    # Print environment info
     echo "✅ LUMI CPU environment loaded!"
-    echo "Partition: CPU (partition/C)"
-    echo "Compiler: $(cc --version | head -1)"
-    echo "CMake: $(cmake --version | head -1)"
+    echo "   Partition: CPU (partition/C)"
+    echo "   Python: $(python3 --version)"
+    echo "   Compiler: $(cc --version | head -1)"
+    echo "   CMake: $(cmake --version | head -1)"
 }
 
-# GPU Setup (ROCm)
 lumi_gpu() {
-    module purge  
+    # GPU partition with Cray compiler and ROCm
     module --force purge
     
-    # Clear ALL potentially problematic environment variables
-    unset HIP_FLAGS
-    unset HIPCXX_FLAGS
-    unset CMAKE_HIP_FLAGS
-    unset CRAY_ROCM_INCLUDE_OPTS
-    unset HIP_COMPILE_FLAGS
-    unset HIP_LINK_FLAGS
+    # Clear any lingering HIP/ROCm variables that might conflict
+    unset HIP_FLAGS HIP_COMPILE_FLAGS HIP_LINK_FLAGS CMAKE_HIP_FLAGS
+    unset CRAY_ROCM_INCLUDE_OPTS HIPCXX_FLAGS
     
-    module load LUMI/24.03 partition/G PrgEnv-cray buildtools/24.03 craype-accel-amd-gfx90a rocm/6.0.3
-    module load cray-mpich
+    # Load GPU stack
+    module load LUMI/25.03 partition/G PrgEnv-cray
+    module load rocm/6.3.4        # Explicitly use 6.3.4 (actual version post-maintenance)
+    module load cray-mpich/8.1.32 # Explicit version for reproducibility
+    module load cray-python
     
-    # CRITICAL: Enable GPU support for MPI GTL
+    # Enable GPU support for MPI
     export MPICH_GPU_SUPPORT_ENABLED=1
-
[…]
