### 2026-01-22 0f2b7a3
Update CMakeLists and CMakePresets post-Jan 2026 maintenance
