+++ apps/CMakeLists.txt
@@ -27,6 +27,7 @@ if(TARGET_PLATFORM STREQUAL "GPU-HIP")
     enable_language(HIP)
     set(HIP_ARCHITECTURES "gfx90a")
     set(CMAKE_HIP_ARCHITECTURES "gfx90a")
+    set(AMDGPU_TARGETS "gfx90a")
     message(STATUS "✅ HIP language enabled for gfx90a")
 elseif(TARGET_PLATFORM STREQUAL "GPU-CUDA")
     enable_language(CUDA)
