### 2026-01-27 #9909 @editorialbot
[REVIEW]: MaterForge: Materials Formulation Engine with Python
**Submitting author:** @rahildoshi97 (<a href="http://orcid.org/0009-0008-3570-9841">Rahil Miten Doshi</a>)
**Repository:** https://github.com/rahildoshi97/materforge
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.6.0
