### 2026-02-04 c4dce2c
Add print statements for StreamCollide kernel operation count
