### 2026-02-04 e56dc59
Add WALBERLA_MPI_BARRIER before LBM communication
