### 2026-02-13 d7694fd
Add individual bash scripts for LUMI-C
