### 2026-02-13 eb37690
Remove benchmarking scripts for LUMI-G
