### 2026-02-16 #9909 @jan-janssen
How does MaterForge compare to open source CALPHAD packages like `pycalphad` https://github.com/pycalphad/pycalphad ? At the moment the the table in the **Comparison with Existing Tools** section just mentions CALPHAD Tools as a general category, while the other items you compare to like `coolprop` and `NIST WebBook` are specific solutions. So I would recommend to do a direct comparison with `pycalphad`  as that is currently the most commonly used open source library for thermodynamic modeling at least in the Python community.
