### 2026-02-16 #9909 @jan-janssen
@rahildoshi97 I was able to install materforge from pypi and execute the examples in https://github.com/rahildoshi97/materforge/tree/master/examples . Still this primarily returns the results in text form, while the manuscript highlights a number of plots for commonly used material properties. I was not able to find the example code to reproduce the result from the manuscript including the plots. Can you include those in the examples?
