### 2026-02-16 #9909 @jan-janssen
@rahildoshi97 Questions regarding the manuscript:
* The manuscript claims ` The library supports various property definition methods such as constant values, step functions, file-based data (.xlsx, .csv, .txt), tabular data, piecewise equations, and computed properties`. Still in the demonstration only `xlsx` files are tested, can you provide examples for `csv` files and `txt` files?
* The manuscript claims ` Extensible Material Support: The framework supports any material type through its extensible architecture. Currently implemented for pure metals and alloys, its modular design allows straightforward extension to materials such as ceramics, polymers, or composites.` Can you provide examples for ceramics, polymers, and composites?
*  It is unclear how `Regression and Data Reduction` differs from `Intelligent Simplification Timing`, to me both refer to the same section in the configuration file - the `regression` section in the YAML files contains `simplify: post`, `degree: 1` and `segments: 3` . Can you explain this in more detail, or merge both sections?
