### 2026-02-16 #9909 @jan-janssen
Validation of the YAML files: Currently the validation of the YAML files is done manually in https://github.com/rahildoshi97/materforge/blob/master/src/materforge/parsing/config/material_yaml_parser.py rather than using https://docs.pydantic.dev/ models which is commonly best practice for validating data structures in Python. Can you comment on why you implemented your own tool for YAML file validation? This should also be mentioned in the manuscript in the `Built-in Validation Framework` section.
