### 2026-02-19 #9909 @JosePizarro3
Hi @rahildoshi97,

I am reviewing your JOSS submission, and already in the first step, after pip installing the package, I cannot run the `getting_started.md` part. When trying to create a material:
```python
...
material = create_material("simple_steel.yaml", T)
```

I got back a ValueError:
```
...
ValueError: Missing required field: material_type
```

I am reviewing other parts, but to go in depth, I would like this to work. **Would you mind patching it quickly and realising the patched version?**

Thanks 🙂
