### 2026-02-19 c983134
Update walberla, add missing material_type to getting_started.md
