### 2026-02-19 #9909 @rahildoshi97
Hi @jan-janssen,

Thanks a lot for your detailed questions and suggestions. I will answer them one by one by the end of this week :)

Hi @JosePizarro3,

Yes, I have patched it and this should now reflect in the master branch. Thanks for pointing it out :)
