### 2026-02-19 #9909 @JosePizarro3
Thanks @rahildoshi97. While waiting for a new release so that I can check the functionality of MaterForge, I decided to leave some general comments, ask you questions, and give you minor suggestions.

First, I have some positive points 🙂

- Although the documentation is lacking content (see below), I think its structure (following [diataxis](https://diataxis.fr/)) is very nice.
- Testing is very extensive, covers unit and integration, and even a couple of extra implementations. Congratulations.
- The structure of the YAML is very clean. I clearly see this could evolve into a standard that multiple people in this community follows.

Now, I have the feedback:

1. I had the feeling that big parts of the code were completely coded by an AI code assistant. I can see that on the overall architecture and design, but later, when coming to the details, it is sometimes unclear the direction the code is taking. **Do you mind disclosing how much and where you used AI to support this work?** Btw, you do not need to tell me about the documentation, that's fine, but I would like to know what happened during coding.
2. As mentioned, the documentation is good, although due to the implementation in src/ of this package, I think its usage is quite involved. This means that I was expecting more how-tos or explanations.

Besides these, I completely agree with all comments from @jan-janssen and won't ellaborate more on those.

And to finish, a few minor suggestions for code health (I took notes in my initial pass of the code):

- `[all]` in pyproject.toml is not needed
- links in pypi for readme are broken: use full paths (not relative ones)
- suggestion: move .sh scripts to a folder scripts/
- suggestion: add ruff as dependency and in ci.yml to check code formatting quality
- where is the sphynx documentation? is not hosted in a Github page?
- quick start in index.rst does not work (missing steel.yaml)
- explain why the `-e` flag during the pip installation: not clear this is only for development
- two dependencies for docs: in pyproject.toml and in docs/requirements.txt
- tutorial/getting_started.md shows pip installation from git, not from pypi
- maybe add the steel_example.yaml in the examples/ folder?
- why create_material() takes a long time? Is this because of the internal validations? I would like to explore this myself, but if you have a quick answer, that would be enough for me

---

_After the fix, I will evaluate the functionality and rest of open points in my todo list above_
