### 2026-02-19 #9909 @rahildoshi97
> [@rahildoshi97](https://github.com/rahildoshi97) I have a couple of questions regarding the materials data provided in MaterForge:
>
>     * **Physical Constants**  In neither your paper nor your documentation, I was able to find any reference for the physical constants. Only in the corresponding python files you mention `CODATA 2018 values` for the fundamental constants:
>       https://github.com/rahildoshi97/materforge/blob/master/src/materforge/data/constants/physical_constants.py . Given that `scipy` is already a dependency of your software, could you explain how your physical constants differ from the ones defined in `scipy`? https://docs.scipy.org/doc/scipy/reference/constants.html Alternatively I would strongly encourage to use the contestants from `scipy` to maintain consistency within the Python ecosystem.
>
>     * **Element Data** The same applies for the references to the element data. You only provide the references in the code as National Institute of Standards and Technology, Royal Society of Chemistry, and CRC Handbook of Chemistry and Physics without any more details https://github.com/rahildoshi97/materforge/blob/master/src/materforge/data/elements/element_data.py . Furthermore, you currently only provide data for 14 elements which drastically restricts the design space of MaterForge. Still this restriction is neither mentioned in the documentation or the manuscript. Why do not you use an established library like mendeleev which already contains the same elemental data for the majority of the elements in the periodic table? https://mendeleev.readthedocs.io/en/stable/ Or at least you could allow the user to use mendeleev as optional dependency, otherwise the application of MaterForge seems to be very restrictive to me.
>
>     * **Alloy Data** MaterForge provides data for a single alloy - referenced as steel 1.4301 : https://github.com/rahildoshi97/materforge/tree/master/src/materforge/data/materials/alloys/1.4301 . The data is stored in the proprietary `xlsx` format. Could you explain why you do not you an open format for storing the data like column separated values (CSV)?  The manuscript mentions this data is computed with the commercial software Jmatpro https://www.sentesoftware.co.uk/jmatpro/full-details but such a reference is missing in the corresponding files. Furthermore, it is unclear to me if you are legally allowed to publish this data or if the copyright belongs to the developers of the jmatpro software. It would be more consistent to provide references to the public data used in the Jmatpro software, in that case it would be fine for using the Jmatpro software for the processing of this data. Finally, is the data format used for the `xlsx` files in Materforge standardized? Are data files for other alloys publicly available? Or is Materforge primarily a post processing tool for output files of the jmatpro software? Both cases are fine, but as they drastically change the scope of application of Materforge, I think this should be clearly lined out in both the documentation and the manuscript.

HI @jan-janssen,

Thanks a lot for the detailed feedback.

1. I agree that using `scipy.constants` is preferable for consistency in the Python ecosystem. I will test this out and add it by the end of this week.

2. I also checked mendeleev for the element data. The earlier MaterForge data was very limited and I am currently thinking about the integration of `mendeleev.element` in MaterForge. However some elements like carbon are missing certain properties like `melting_point`,  `fusion_heat` and `evaporation_heat`. I am just thinking how to tackle such elements. An idea would be to throw an error for pure metals when such a property is missing as it makes no sense without complete phase-change data. For alloys, partial thermophysical data can be allowed, because alloys in MaterForge use user‑supplied property definitions rather than relying on elemental fusion/evaporation heats.

3. You are right that the current distribution contains data for a single alloy (steel 1.4301) and that it is stored in an .xlsx file. That file is meant purely as an example/demo dataset to exercise the YAML/property pipeline. Data can also be used in the form of a CSV or txt file, as preferred by the user. MaterForge is not intended to be a general‑purpose material database or a Jmatpro post-processor. The core design is to let users define arbitrary properties as YAML (constants, equations, tabular data) and turn those into symbolic, validated models. Jmatpro is just one such possible source. The 1.4301 data were generated with JMatPro by a collaborator at FAU (acknowledged in the manuscript).

I will however make things more explicit in the repository/manuscript to avoid any confusion.
