### 2026-02-19 #9909 @rahildoshi97
> [@rahildoshi97](https://github.com/rahildoshi97) I was able to install materforge from pypi and execute the examples in https://github.com/rahildoshi97/materforge/tree/master/examples . Still this primarily returns the results in text form, while the manuscript highlights a number of plots for commonly used material properties. I was not able to find the example code to reproduce the result from the manuscript including the plots. Can you include those in the examples?

The existing example uses the following YAML files:
https://i10git.cs.fau.de/rahil.doshi/materforge/-/blob/master/src/materforge/data/materials/alloys/1.4301/1.4301.yaml?ref_type=heads
https://i10git.cs.fau.de/rahil.doshi/materforge/-/blob/master/src/materforge/data/materials/pure_metals/Al/Al.yaml?ref_type=heads
Although the examples return results in text form, they also create a `materforge_plots` folder in the same directory as your YAML file. So you should see the generated plots here:
src/materforge/data/materials/alloys/1.4301/materforge_plots/1.4301_properties.png
src/materforge/data/materials/pure_metals/Al/materforge_plots/Aluminum_properties.png

The plots are always generated in the same directory as the YAML file.

However you're right that the example code does not reproduces the exact figures from the manuscript. I will add a dedicated script that explicitly generates the exact figures. I will make it a point to also mention that the plots are generated in the same directory as the YAML files.
