### 2026-02-19 #9909 @rahildoshi97
> Validation of the YAML files: Currently the validation of the YAML files is done manually in https://github.com/rahildoshi97/materforge/blob/master/src/materforge/parsing/config/material_yaml_parser.py rather than using https://docs.pydantic.dev/ models which is commonly best practice for validating data structures in Python. Can you comment on why you implemented your own tool for YAML file validation? This should also be mentioned in the manuscript in the `Built-in Validation Framework` section.

Thank you for the suggestion. Pydantic works well when we know the shape of the data upfront, but that's not really the case here. The `properties` block is intentionally flexible. An entry could be a plain scalar, a file reference with regression config, a computed expression derived from other properties, etc. The type is inferred at runtime by looking at which keys are present, so there's no single field we can point Pydantic to for deciding which model to apply. The physical checks like composition fractions summing to 1.0, validating names, required fields, etc are also domain specific rules that would need custom validator methods. Using Pydantic wouldn't remove that logic, it would just add a dependency and wrap the same code in a different decorator.
