### 2026-02-20 7b96c09
Integrate mendeleev for element data
