+++ src/materforge/data/elements/__init__.py
@@ -5,13 +5,8 @@
 
 from .element_data import (
     element_map, get_element,
-    CARBON, NITROGEN, ALUMINIUM, SILICON, PHOSPHORUS, SULFUR,
-    TITANIUM, VANADIUM, CHROMIUM, MANGANESE, IRON, NICKEL, MOLYBDENUM
 )
 
 __all__ = [
     "element_map", "get_element",
-    # Individual elements for direct access
-    "CARBON", "NITROGEN", "ALUMINIUM", "SILICON", "PHOSPHORUS", "SULFUR",
-    "TITANIUM", "VANADIUM", "CHROMIUM", "MANGANESE", "IRON", "NICKEL", "MOLYBDENUM"
 ]
