+++ src/materforge/data/__init__.py
@@ -11,9 +11,7 @@ chemical element data, and material property databases used throughout PyMatLib.
 from .constants.physical_constants import PhysicalConstants
 from .constants.processing_constants import ProcessingConstants, ErrorMessages, FileConstants
 from .elements.element_data import (
-    element_map,
-    CARBON, NITROGEN, ALUMINIUM, SILICON, PHOSPHORUS, SULFUR,
-    TITANIUM, VANADIUM, CHROMIUM, MANGANESE, IRON, NICKEL, MOLYBDENUM
+    element_map, get_element,
 )
 
 __all__ = [
@@ -22,7 +20,5 @@ __all__ = [
     "ErrorMessages",
     "FileConstants",
     "element_map",
-    # Individual elements for direct access
-    "CARBON", "NITROGEN", "ALUMINIUM", "SILICON", "PHOSPHORUS", "SULFUR",
-    "TITANIUM", "VANADIUM", "CHROMIUM", "MANGANESE", "IRON", "NICKEL", "MOLYBDENUM"
+    "get_element",
 ]
