+++ pyproject.toml
@@ -39,6 +39,7 @@ dependencies = [
     "pandas>=1.4.0,<3.0.0",
     "openpyxl>=3.1.5",
     "pwlf>=2.5.1,<3.0.0",
+    "mendeleev>=1.1.0",
 ]
 
 [project.optional-dependencies]

[…]
