### 2026-02-20 #9909 @rahildoshi97
> How does MaterForge compare to open source CALPHAD packages like `pycalphad` https://github.com/pycalphad/pycalphad ? At the moment the the table in the **Comparison with Existing Tools** section just mentions CALPHAD Tools as a general category, while the other items you compare to like `coolprop` and `NIST WebBook` are specific solutions. So I would recommend to do a direct comparison with `pycalphad` as that is currently the most commonly used open source library for thermodynamic modeling at least in the Python community.

Thank you for this suggestion. It prompted me to reconsider how I position MaterForge. After reflection, I agree the comparison table is misleading.

pycalphad is a thermodynamic modeling library: it reads databases, solves multi-component Gibbs energy minimization, and computes phase equilibria. MaterForge does not do any of this. It operates at a fundamentally different layer of the workflow: it takes property data already generated by tools like pycalphad, CoolProp, or NIST WebBook as input, and post-processes it into optimized symbolic representations suitable for numerical simulations.​

I will therefore removed the comparison table and replaced it with an explicit description of the three-stage workflow in which MaterForge fits. pycalphad will be mentioned as a complementary upstream tool rather than a competing one.

Stage 1: material property data generation with tools like pycalphad, NIST Webbook, etc.
Stage 2: MaterForge for postprocessing and optimization
Stage 3: simulation integration

I hope that makes sense :)
