+++ src/materforge/data/constants/__init__.py
@@ -1,7 +1,7 @@
 # SPDX-FileCopyrightText: 2025 Rahil Miten Doshi, Friedrich-Alexander-Universität Erlangen-Nürnberg
 # SPDX-License-Identifier: BSD-3-Clause
 
-"""Physical and processing constants for PyMatLib."""
+"""Physical and processing constants for MaterForge."""
 
 from .physical_constants import PhysicalConstants
 from .processing_constants import ProcessingConstants, ErrorMessages, FileConstants
