+++ src/materforge/parsing/validation/__init__.py
@@ -1,7 +1,7 @@
 # SPDX-FileCopyrightText: 2025 Rahil Miten Doshi, Friedrich-Alexander-Universität Erlangen-Nürnberg
 # SPDX-License-Identifier: BSD-3-Clause
 
-"""Validation utilities for PyMatLib."""
+"""Validation utilities for MaterForge."""
 
 from .array_validator import is_monotonic
 from .errors import PropertyError, DependencyError, CircularDependencyError
