+++ src/materforge/data/materials/alloys/1.4301/1.4301.yaml
@@ -1,8 +1,8 @@
 # ====================================================================================================
-# PYMATLIB MATERIAL CONFIGURATION FILE
+# MaterForge MATERIAL CONFIGURATION FILE
 # ====================================================================================================
 # This file defines material properties for 1.4301 (Steel) using the new materforge format.
-# Pymatlib supports 6 property types: CONSTANT_VALUE, STEP_FUNCTION, FILE_IMPORT, TABULAR_DATA, PIECEWISE_EQUATION, and COMPUTED_PROPERTY
+# MaterForge supports 6 property types: CONSTANT_VALUE, STEP_FUNCTION, FILE_IMPORT, TABULAR_DATA, PIECEWISE_EQUATION, and COMPUTED_PROPERTY
 #
 # IMPORTANT: All property configurations must include 'bounds' parameter (except CONSTANT properties)
 #
