### 2026-02-20 474ea13
Add user example for hypothetical alloy
