+++ examples/myAlloy.py
@@ -1,0 +1,239 @@
+# SPDX-FileCopyrightText: 2025 Rahil Miten Doshi, Friedrich-Alexander-Universität Erlangen-Nürnberg
+# SPDX-License-Identifier: BSD-3-Clause
+
+"""Demonstration script for material property evaluation."""
+import logging
+from pathlib import Path
+import sympy as sp
+
+from materforge.parsing.api import (
+    create_material,
+    get_supported_properties,
+    evaluate_material_properties,
+    get_material_property_names,
+    validate_yaml_file,
+    get_material_info
+)
+
+
+def setup_logging():
+    """Setup logging configuration."""
+    logging.basicConfig(
+        level=logging.WARNING,  # DEBUG/INFO/WARNING/ERROR/CRITICAL
+        format="%(asctime)s %(levelname)s %(name)s -> %(message)s"
+    )
+    # Silence noisy libraries
+    logging.getLogger('matplotlib').setLevel(logging.WARNING)
+    logging.getLogger('PIL').setLevel(logging.WARNING)
+    logging.getLogger('fontTools').setLevel(logging.WARNING)
+
+
+def demonstrate_material_properties():
+    """Comprehensive demonstration of material property evaluation."""
+    setup_logging()
+
+    # Create symbolic temperature variables for different materials
+    T = sp.Symbol('T')  # Temperature symbol for myAlloy
+
+    # Set up file paths
+    current_file = Path(__file__)
+    yaml_path = current_file.parent / "myAlloy.yaml"
+
+    materials = []
+    material_symbols = {}  # Dictionary to track which symbol goes with which material
+
+    print(f"\n{'=' * 80}")
+    print("MaterForge MATERIAL PROPERTY DEMONSTRATION")
+    print(f"{'=' * 80}")
+
+    # ===================================================================
+    # 1. YAML FILE VALIDATION
+    # ===================================================================
+    print(f"\n{'1. YAML FILE VALIDATION':<50}")
+    print(f"{'-' * 50}")
+
+    for yaml_path, name in [(yaml_path, "myAlloy")]:
+        if yaml_path.exists():
+            try:
+                is_valid = validate_yaml_file(yaml_path)
+                print(f"✓ {name} YAML validation: {'PASSED' if is_valid else 'FAILED'}")
+            except Exception as e:
+                print(f"✗ {name} YAML validation: FAILED - {e}")
+        else:
+            print(f"✗ {name} YAML file not found: {yaml_path}")
+
+    # ===================================================================
+    # 2. MATERIAL INFO EXTRACTION
+    # ===================================================================
+    print(f"\n{'2. MATERIAL INFO EXTRACTION':<50}")
+    print(f"{'-' * 50}")
+
+    for yaml_path, name in [(yaml_path, "myAlloy")]:
+        if yaml_path.exists():
+            try:
+                info = get_material_info(yaml_path)
+                print(f"\n{name} Information:")
+                print(f"  Name: {info['name']}")
+                print(f"  Type: {info['material_type']}")
+                print(f"  Total Properties: {info['total_properties']}")
+                print(f"  Composition: {info['composition']}")
+
+                # Temperature properties
+                if info['material_type'] == 'pure_metal':
+                    print(f"  Melting Temperature: {info.get('melting_temperature', 'N/A')} K")
+                    print(f"  Boiling Temperature: {info.get('boiling_temperature', 'N/A')} K")
+                else:
+                    print(f"  Solidus Temperature: {info.get('solidus_temperature', 'N/A')} K")
+                    print(f"  Liquidus Temperature: {info.get('liquidus_temperature', 'N/A')} K")
+
+            except Exception as e:
+                print(f"✗ Failed to get {name} info: {e}")
+
[…]
