+++ examples/myAlloy.yaml
@@ -1,0 +1,120 @@
+# ====================================================================================================
+# MaterForge MATERIAL CONFIGURATION FILE
+# ====================================================================================================
+# This is an example file that defines material properties for a hypothetical alloy using the new materforge format.
+# MaterForge supports 6 property types: CONSTANT_VALUE, STEP_FUNCTION, FILE_IMPORT, TABULAR_DATA, PIECEWISE_EQUATION, and COMPUTED_PROPERTY
+#
+# IMPORTANT: All property configurations must include 'bounds' parameter (except CONSTANT properties)
+#
+# Note: T is just a placeholder in equations.
+# The final equation will depend on the symbol passed via the API.
+# ====================================================================================================
+
+name: "myAlloy"
+material_type: alloy  # Must be 'alloy' or 'pure_metal'
+
+# Composition fractions must sum to 1.0
+composition:
+  Fe: 0.675
+  Cr: 0.170
+  Ni: 0.120
+  Mo: 0.025
+  Mn: 0.01
+
+# Required temperature properties for alloys
+solidus_temperature: 1605.          # Temperature where melting begins (K)
+liquidus_temperature: 1735.         # Temperature where material is completely melted (K)
+initial_boiling_temperature: 3090.  # Temperature where boiling begins (K)
+final_boiling_temperature: 3200.    # Temperature where material is completely vaporized (K)
+
+properties:
+
+  # ====================================================================================================
+  # PROPERTY TYPE 1: CONSTANT_VALUE
+  # Format: property_name: numeric_value
+  # Note: Must be float, not integer (use 1.0 instead of 1)
+  # ====================================================================================================
+
+  density: 7000.  # kg/m^3
+
+  # ====================================================================================================
+  # PROPERTY TYPE 2: STEP_FUNCTION
+  # Format: Uses 'dependency' (single reference) and 'value' (list of 2 values)
+  # Supports temperature references and arithmetic expressions
+  # ====================================================================================================
+
+  latent_heat_of_fusion:
+    dependency: solidus_temperature
+    value: [0.0, 171401.0]
+    bounds: [constant, constant]
+
+  # ====================================================================================================
+  # PROPERTY TYPE 3: TABULAR_DATA
+  # Format: Uses 'dependency' (list/tuple) and 'value' (list) with same length
+  # Supports temperature references, explicit lists, and tuple notation
+  # ====================================================================================================
+
+  heat_conductivity:
+    dependency: [500, 1000, 1600, 1700, 1750, 2000, 2500]
+    value: [19.25, 25.47, 32.94, 33.52, 31.53, 35.33, 42.95]
+    bounds: [extrapolate, extrapolate]
+
+  # ====================================================================================================
+  # PROPERTY TYPE 4: FILE_IMPORT
+  # Format: Uses 'file_path', 'dependency_column', 'property_column'
+  # Supports .txt, .csv, .xlsx files with automatic missing value handling
+  # ====================================================================================================
+
+  heat_capacity:
+    file_path: ./myAlloy.csv
+    dependency_column: T (K)
+    property_column: Specific heat (J/(Kg K))
+    bounds: [constant, constant]
+    regression:
+      simplify: pre
+      degree: 3
+      segments: 6
+
+  # ====================================================================================================
+  # PROPERTY TYPE 5: PIECEWISE_EQUATION
+  # Format: Uses 'dependency' (breakpoints) and 'equation' (list of expressions)
+  # Number of equations = number of breakpoints - 1
+  # ====================================================================================================
+
+  viscosity:
+    dependency: [300, 1660, 1736, 3000]
[…]
