### 2026-02-22 9237e11
Remove [all] optional dependency group from pyproject.toml
