### 2026-02-22 f643c1f
Remove f prefix for non-f-string
