### 2026-02-22 85bc14e
Replace == True with is True in piecewise inverter
