### 2026-02-22 8a15650
Consolidate dependencies into pyproject.toml, remove docs/requirements.txt
