+++ .readthedocs.yml
@@ -27,4 +27,3 @@ python:
       path: .
       extra_requirements:
         - docs
-    - requirements: docs/requirements.txt
