### 2026-02-22 12d6bc5
Use sp.true identity check for SymPy Piecewise final piece detection
SymPy converts Python's True to sp.true when building Piecewise.
Identity check `is True` fails because sp.true and
Python's True are different objects. `== True` worked previously
but triggered ruff E712. `is sp.true` is both ruff-clean and runtime-correct.
