### 2026-02-22 5a7121b
Migrate setuptools-scm write_to to version_file
