+++ pyproject.toml
@@ -110,7 +110,7 @@ materforge = [
 ]
 
 [tool.setuptools_scm]
-write_to = "src/materforge/_version.py"
+version_file = "src/materforge/_version.py"
 version_scheme = "python-simplified-semver"
 local_scheme = "node-and-date"
 fallback_version = "0.6.3"
