### 2026-02-22 61a4128
Update getting_started doc
