### 2026-02-22 31765a0
Replace hardcoded property lists with metadata-tagged fields
