+++ src/materforge/core/materials.py
@@ -1,6 +1,7 @@
 # SPDX-FileCopyrightText: 2025 Rahil Miten Doshi, Friedrich-Alexander-Universität Erlangen-Nürnberg
 # SPDX-License-Identifier: BSD-3-Clause
 
+import dataclasses
 import logging
 from dataclasses import dataclass, field
 from typing import Dict, List, Optional, Tuple, Union
@@ -15,70 +16,81 @@ from materforge.core.exceptions import MaterialCompositionError, MaterialTempera
 
 logger = logging.getLogger(__name__)
 
+def _prop(**kwargs) -> dataclasses.Field:
+    """Declare a material property field, discoverable via property_field_names()."""
+    return field(default=None, metadata={"is_material_property": True}, **kwargs)
 
 @dataclass
 class Material:
     """
-    Represents a material with its composition and properties.
+    Represents a material with its composition and thermophysical properties.
 
-    This class supports both pure metals and alloys with comprehensive
-    temperature validation and property calculation.
+    Supports pure metals and alloys. All physical property fields are tagged
+    with metadata so property_field_names() can discover them automatically
+    without a separate registry.
     """
     # Basic material information
     name: str
     material_type: str  # 'pure_metal' or 'alloy'
     elements: List[ChemicalElement]
     composition: Union[np.ndarray, List[float], Tuple]  # List of fractions summing to 1.0
-    # Temperature properties for pure metals
+    # Temperature bounds - pure metals
     melting_temperature: Optional[sp.Float] = None
     boiling_temperature: Optional[sp.Float] = None
-    # Temperature properties for alloys
+    # Temperature bounds - alloys
     solidus_temperature: Optional[sp.Float] = None
     liquidus_temperature: Optional[sp.Float] = None
     initial_boiling_temperature: Optional[sp.Float] = None
     final_boiling_temperature: Optional[sp.Float] = None
-    # Optional properties with default values
-    bulk_modulus: Optional[sp.Float] = None
-    density: sp.Expr = None
-    dynamic_viscosity: sp.Expr = None
-    elastic_modulus: Optional[sp.Expr] = None  # Young's modulus
-    electrical_conductivity: Optional[sp.Expr] = None
-    electrical_resistivity: Optional[sp.Expr] = None
-    energy_density: Optional[Union[sp.Piecewise, sp.Expr]] = None
-    energy_density_solidus: sp.Float = None
-    energy_density_liquidus: sp.Float = None
-    fracture_toughness: Optional[sp.Expr] = None
-    hardness: Optional[sp.Expr] = None
-    heat_capacity: sp.Expr = None
-    heat_conductivity: sp.Expr = None
-    kinematic_viscosity: sp.Expr = None
-    latent_heat_of_fusion: sp.Expr = None
-    latent_heat_of_vaporization: sp.Expr = None
-    magnetic_permeability: Optional[sp.Expr] = None
-    melting_point_pressure: Optional[sp.Float] = None
-    poisson_ratio: Optional[sp.Float] = None
-    shear_modulus: Optional[sp.Expr] = None
-    specific_enthalpy: sp.Expr = None
-    surface_tension: sp.Expr = None
-    thermal_diffusivity: sp.Expr = None
-    thermal_expansion_coefficient: sp.Expr = None
-    ultimate_tensile_strength: Optional[sp.Expr] = None
-    viscosity: sp.Expr = None
-    yield_strength: Optional[sp.Expr] = None
-    # Calculated properties (set during initialization)
+    # Thermophysical properties
+    bulk_modulus: Optional[sp.Float] = _prop()
+    density: Optional[sp.Expr] = _prop()
[…]
