### 2026-02-22 d0b70c1
Clean PyPI package (no plots, SPDX license)
