+++ pyproject.toml
@@ -1,5 +1,5 @@
 [build-system]
-requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
+requires = ["setuptools>=77", "setuptools_scm[toml]>=8"]
 build-backend = "setuptools.build_meta"
 
 [project]
@@ -7,7 +7,8 @@ name = "materforge"
 dynamic = ["version"]
 description = "A high-performance Python library for material simulation and analysis"
 readme = "README.md"
-license = {file = "LICENSE"}
+license = "BSD-3-Clause"
+license-files = ["LICENSE"]
 authors = [
     {name = "Rahil Doshi", email = "rahil.doshi@fau.de"},
 ]
@@ -19,7 +20,6 @@ classifiers = [
     "Development Status :: 4 - Beta",
     "Intended Audience :: Science/Research",
     "Intended Audience :: Developers",
-    "License :: OSI Approved :: BSD License",
     "Operating System :: OS Independent",
     "Programming Language :: Python :: 3",
     "Programming Language :: Python :: 3.10",
@@ -27,7 +27,7 @@ classifiers = [
     "Programming Language :: Python :: 3.12",
     "Programming Language :: Python :: 3.13",
     "Topic :: Scientific/Engineering :: Physics",
-    "Topic :: Scientific/Engineering"
+    "Topic :: Scientific/Engineering",
 ]
 requires-python = ">=3.10"
 dependencies = [
@@ -47,9 +47,6 @@ dev = [
     "pytest-xdist>=3.0.0",
     "pytest-mock>=3.10.0",
     "hypothesis>=6.0.0",
-    "black>=23.0.0",
-    "isort>=5.12.0",
-    "flake8>=6.0.0",
     "mypy>=1.0.0",
     "pre-commit>=3.0.0",
     "ruff>=0.1.0",
@@ -96,16 +93,12 @@ exclude = ["tests*", "apps*", "examples*", "*materforge_plots*", "benchmarking-s
 
 [tool.setuptools.package-data]
 materforge = [
-    "data/**/*",
-    "data/alloys/**/*",
-    "data/pure_metals/**/*",
-    "!**/materforge_plots/**",
-    "*.yaml",
-    "*.yml",
-    "*.json",
-    "*.xlsx",
-    "*.csv",
-    "*.txt",
+    "data/**/*.yaml",
+    "data/**/*.yml",
+    "data/**/*.json",
+    "data/**/*.xlsx",
+    "data/**/*.csv",
+    "data/**/*.txt",
     "py.typed",
 ]
 
@@ -114,4 +107,3 @@ version_file = "src/materforge/_version.py"
 version_scheme = "python-simplified-semver"
 local_scheme = "node-and-date"
 fallback_version = "0.6.3"
-tag_regex = "^(?:v)?(?P<version>\\d+\\.\\d+\\.\\d+)$"
