### 2026-02-22 d8d2609
Update CHANGELOG and CITATION.cff for v0.6.4 release
