### 2026-02-23 #9909 @rahildoshi97
> Thanks [@rahildoshi97](https://github.com/rahildoshi97). While waiting for a new release so that I can check the functionality of MaterForge, I decided to leave some general comments, ask you questions, and give you minor suggestions.
>
> First, I have some positive points 🙂
>
> * Although the documentation is lacking content (see below), I think its structure (following [diataxis](https://diataxis.fr/)) is very nice.
> * Testing is very extensive, covers unit and integration, and even a couple of extra implementations. Congratulations.
> * The structure of the YAML is very clean. I clearly see this could evolve into a standard that multiple people in this community follows.
>
> Now, I have the feedback:
>
> 1. I had the feeling that big parts of the code were completely coded by an AI code assistant. I can see that on the overall architecture and design, but later, when coming to the details, it is sometimes unclear the direction the code is taking. **Do you mind disclosing how much and where you used AI to support this work?** Btw, you do not need to tell me about the documentation, that's fine, but I would like to know what happened during coding.
> 2. As mentioned, the documentation is good, although due to the implementation in src/ of this package, I think its usage is quite involved. This means that I was expecting more how-tos or explanations.
>
> Besides these, I completely agree with all comments from [@jan-janssen](https://github.com/jan-janssen) and won't ellaborate more on those.
>
> And to finish, a few minor suggestions for code health (I took notes in my initial pass of the code):
>
> * `[all]` in pyproject.toml is not needed
> * links in pypi for readme are broken: use full paths (not relative ones)
> * suggestion: move .sh scripts to a folder scripts/
> * suggestion: add ruff as dependency and in ci.yml to check code formatting quality
> * where is the sphynx documentation? is not hosted in a Github page?
> * quick start in index.rst does not work (missing steel.yaml)
> * explain why the `-e` flag during the pip installation: not clear this is only for development
> * two dependencies for docs: in pyproject.toml and in docs/requirements.txt
> * tutorial/getting_started.md shows pip installation from git, not from pypi
> * maybe add the steel_example.yaml in the examples/ folder?
> * why create_material() takes a long time? Is this because of the internal validations? I would like to explore this myself, but if you have a quick answer, that would be enough for me
>
> _After the fix, I will evaluate the functionality and rest of open points in my todo list above_

Hi @JosePizarro3, thank you for the detailed review and the kind words about the Diataxis structure, the test coverage, and the YAML schema. The YAML schema actually went through many iterations with my supervisors before we were happy with it, and I do hope it becomes something others in this community can build on.

Let me address each of your points below.

1. AI usage in development
Thanks for asking this directly. It's a fair question and I'm happy to be transparent about it.
During development of MaterForge, a coding assistant (mainly GitHub Copilot in VS Code) was mainly used for:
- Boilerplate completions - repeating established patterns across functions (argument validation, __init__ assignments, control flow), generating class scaffolding, and occasionally drafting parametrized test cases where the structure was already clear
- Exception handling and logging setup - suggesting try/except blocks and logging patterns
- Occasionally proposing refactorings or alternative implementations for functions whose behavior and interface I had already designed

Copilot's suggestions were treated as first drafts. All AI-assisted code was reviewed, edited, and extended by me, and validated through the test suite and example workflows before being committed. In several places the final implementation diverges substantially from the initial suggestion after repeated manual refinement.

The core functionality of MaterForge, including the overall architecture, YAML schema and validation logic, property type detection, dependency resolution pipeline, and piecewise regression approach were designed and implemented by me, then iteratively refined based on feedback from my supervisors until we were satisfied with both the design and behavior.

Some of that inconsistency in the code comes from Copilot-assisted boilerplate sitting alongside hand-written domain logic that evolved over many iterations. I'm actively cleaning this up as part of addressing the review. I will also add a formal 'AI usage disclosure' section to the manuscript in accordance with the JOSS AI policy.

2. Documentation
I agree that more how-to guides and explanations are needed, especially given that the src/ layout makes the usage path less obvious. I'll be adding more practical how-tos in the coming updates.

Code health suggestions:
I've addressed all of your code health suggestions in v0.6.4 (now live on [PyPI](https://pypi.org/project/materforge/0.6.4/))

Why does create_material() take a long time?

The real reason is that create_material() does all its symbolic math upfront, at construction time. For every property with tabular data, from files or computed expressions, MaterForge converts the raw data points into a SymPy Piecewise function with one linear segment per data point. A property with 100 temperature measurements produces a SymPy expression with 101 conditions. SymPy is a symbolic library, not a numerical one, so creating and internally processing that expression is inherently slow, and it gets worse the more data points a property has. Interestingly, properties with regression enabled are actually faster to construct, because regression reduces those raw points down to a handful of polynomial segments, a much smaller object for SymPy to handle.

On top of that, properties flagged for post-processing regression go through the pipeline twice: once to build an initial representation, and again to fit a cleaner regression model. That second pass lambdifies the already-built SymPy expression just to sample it numerically, then rebuilds from scratch.

This is a deliberate design tradeoff: create_material() is called once per simulation run, while the resulting property callables are evaluated millions of times inside the solver's time loop. The slow work happens once at the beginning so that every property lookup during the actual simulation is essentially instant.

Hope this helps 🙂
