### 2026-02-24 f62bd9c
Replace fixed property list with dynamic tracker; remove get_supported_properties
