+++ src/materforge/core/materials.py
@@ -1,39 +1,41 @@
 # SPDX-FileCopyrightText: 2025 Rahil Miten Doshi, Friedrich-Alexander-Universität Erlangen-Nürnberg
 # SPDX-License-Identifier: BSD-3-Clause
 
-import dataclasses
 import logging
 from dataclasses import dataclass, field
-from typing import Dict, List, Optional, Tuple, Union
+from typing import Dict, List, Optional, Set, Tuple, Union
 
 import numpy as np
 import sympy as sp
 
-from materforge.core.elements import (ChemicalElement,
-                                    interpolate_atomic_mass,
-                                    interpolate_atomic_number)
+from materforge.core.elements import ChemicalElement, interpolate_atomic_mass, interpolate_atomic_number
 from materforge.core.exceptions import MaterialCompositionError, MaterialTemperatureError
 
 logger = logging.getLogger(__name__)
 
-def _prop(**kwargs) -> dataclasses.Field:
-    """Declare a material property field, discoverable via property_field_names()."""
-    return field(default=None, metadata={"is_material_property": True}, **kwargs)
+# Structural identity fields (never tracked as material properties)
+_CORE_FIELDS = frozenset({'name', 'material_type', 'elements', 'composition',
+    'melting_temperature', 'boiling_temperature',
+    'solidus_temperature', 'liquidus_temperature',
+    'initial_boiling_temperature', 'final_boiling_temperature',
+    'atomic_number', 'atomic_mass', '_dynamic_properties',
+})
+
 
 @dataclass
 class Material:
     """
     Represents a material with its composition and thermophysical properties.
 
-    Supports pure metals and alloys. All physical property fields are tagged
-    with metadata so property_field_names() can discover them automatically
-    without a separate registry.
+    Supports pure metals, alloys, and future material types (ceramics, polymers, etc.).
+    Thermophysical properties are assigned dynamically via setattr and tracked
+    automatically - no predefined property fields required.
     """
     # Basic material information
     name: str
-    material_type: str  # 'pure_metal' or 'alloy'
+    material_type: str
     elements: List[ChemicalElement]
-    composition: Union[np.ndarray, List[float], Tuple]  # List of fractions summing to 1.0
+    composition: Union[np.ndarray, List[float], Tuple]
     # Temperature bounds - pure metals
     melting_temperature: Optional[sp.Float] = None
     boiling_temperature: Optional[sp.Float] = None
@@ -42,69 +44,57 @@ class Material:
     liquidus_temperature: Optional[sp.Float] = None
     initial_boiling_temperature: Optional[sp.Float] = None
     final_boiling_temperature: Optional[sp.Float] = None
-    # Thermophysical properties
-    bulk_modulus: Optional[sp.Float] = _prop()
-    density: Optional[sp.Expr] = _prop()
-    dynamic_viscosity: Optional[sp.Expr] = _prop()
-    elastic_modulus: Optional[sp.Expr] = _prop()
-    electrical_conductivity: Optional[sp.Expr] = _prop()
-    electrical_resistivity: Optional[sp.Expr] = _prop()
-    energy_density: Optional[Union[sp.Piecewise, sp.Expr]] = _prop()
-    energy_density_solidus: Optional[sp.Float] = _prop()
-    energy_density_liquidus: Optional[sp.Float] = _prop()
-    fracture_toughness: Optional[sp.Expr] = _prop()
-    hardness: Optional[sp.Expr] = _prop()
-    heat_capacity: Optional[sp.Expr] = _prop()
-    heat_conductivity: Optional[sp.Expr] = _prop()
[…]
