+++ src/materforge/parsing/api.py
@@ -1,13 +1,7 @@
 # SPDX-FileCopyrightText: 2025 Rahil Miten Doshi, Friedrich-Alexander-Universität Erlangen-Nürnberg
 # SPDX-License-Identifier: BSD-3-Clause
 
-"""
-Main API module for MaterForge material property library.
-
-This module provides the primary interface for creating and working with material objects
-from YAML configuration files. It includes functions for material creation, validation,
-property evaluation, and information extraction.
-"""
+"""Main API module for MaterForge material property library."""
 
 import logging
 from pathlib import Path
@@ -19,10 +13,6 @@ from materforge.core.materials import Material
 from materforge.parsing.config.material_yaml_parser import MaterialYAMLParser
 from materforge.parsing.config.yaml_keys import (
     NAME_KEY, MATERIAL_TYPE_KEY, COMPOSITION_KEY, PROPERTIES_KEY,
-    PURE_METAL_KEY, MELTING_TEMPERATURE_KEY, BOILING_TEMPERATURE_KEY,
-    SOLIDUS_TEMPERATURE_KEY, LIQUIDUS_TEMPERATURE_KEY,
-    INITIAL_BOILING_TEMPERATURE_KEY, FINAL_BOILING_TEMPERATURE_KEY,
-    ALLOY_KEY
 )
 
 logger = logging.getLogger(__name__)
@@ -34,66 +24,32 @@ logger = logging.getLogger(__name__)
 
 def create_material(yaml_path: Union[str, Path], dependency: sp.Symbol,
                     enable_plotting: bool = True) -> Material:
-    """Create material instance from YAML configuration file.
-
-    This function serves as the main entry point for creating material
-    objects from YAML configuration files. It handles the parsing of the configuration
-    and creation of the material with the specified temperature.
-
-    Parameters
-    ----------
-    yaml_path : Union[str, Path]
-        Path to the YAML configuration file
-    dependency : sp.Symbol
-        Sympy symbol for property evaluation. Use a symbolic variable
-        (e.g., sp.Symbol('T') or sp.Symbol('u_C')) for symbolic temperature expressions
-    enable_plotting : bool, optional
-        Whether to generate visualization plots (default: True)
-
-    Returns
-    -------
-    Material
-        The material instance with all properties initialized
-
-    Raises
-    ------
-    FileNotFoundError
-        If the YAML file doesn't exist
-    ValueError
-        If the YAML content is invalid or material creation fails
-    TypeError
-        If temperature parameter has invalid type
-
-    Notes
-    -----
-    In YAML files, always use 'T' as the temperature variable in equations.
-    The system will automatically substitute this with your provided symbol.
-
-    Examples
-    --------
-    Create a material with symbolic temperature expressions:
-
-    >>> import sympy as sp
-    >>> T = sp.Symbol('T')
-    >>> material = create_material('steel.yaml', T)
-    >>> print(material.name)
-    Steel
-
-    Create a material with a custom temperature symbol:
-
-    >>> u_C = sp.Symbol('u_C')
-    >>> material_copper = create_material('copper.yaml', u_C)
+    """Create a Material from a YAML configuration file.
+    Args:
+        yaml_path: Path to the YAML configuration file.
+        dependency: SymPy symbol for property evaluation (e.g. sp.Symbol('T')).
+            In YAML equations always use 'T'; it is substituted automatically.
+        enable_plotting: Generate visualisation plots (default: True).
+    Returns:
+        Material: Fully initialised material instance.
+    Raises:
+        FileNotFoundError: YAML file does not exist.
+        TypeError: dependency is not a sp.Symbol.
+        ValueError: YAML content is invalid or material creation fails.
+    Example:
[…]
