### 2026-02-24 2b9b38c
Simplify YAML, remove keys(material_type, composition, etc.), remove additional validation logic, update tests
