+++ tests/unit/test_visualization/test_plotters_error_handling.py
@@ -5,17 +5,15 @@ import tempfile
 import sympy as sp
 from pathlib import Path
 from unittest.mock import Mock, patch
-from materforge.visualization.plotters import PropertyVisualizer
 from materforge.core.materials import Material
-from materforge.core.elements import ChemicalElement
+from materforge.visualization.plotters import PropertyVisualizer
 
 
 class TestPropertyVisualizerErrorHandling:
-    """Test error handling in property visualization."""
+    """Error handling tests for PropertyVisualizer."""
 
     @pytest.fixture
     def mock_parser(self):
-        """Create a mock parser for testing."""
         parser = Mock()
         parser.base_dir = Path(tempfile.gettempdir())
         parser.config_path = "test.yaml"
@@ -25,129 +23,97 @@ class TestPropertyVisualizerErrorHandling:
             'FILE': [],
             'STEP_FUNCTION': [],
             'PIECEWISE_EQUATION': [],
-            'COMPUTE': []
+            'COMPUTE': [],
         }
-        parser.config = {
-            'name': 'TestMaterial',
-            'material_type': 'pure_metal'
-        }
+        parser.config = {'name': 'TestMaterial'}
         return parser
 
     @pytest.fixture
     def sample_material(self):
-        """Create a sample material for testing."""
-        element = ChemicalElement(
-            name="Iron", atomic_number=26, atomic_mass=55.845,
-            melting_temperature=1811, boiling_temperature=3134,
-            latent_heat_of_fusion=13800, latent_heat_of_vaporization=340000
-        )
-
-        material = Material(
-            name="TestMaterial",
-            material_type="pure_metal",
-            elements=[element],
-            composition=[1.0],
-            melting_temperature=sp.Float(1811),
-            boiling_temperature=sp.Float(3134)
-        )
-        material.density = sp.Float(7850)
-        return material
+        mat = Material(name="TestMaterial")
+        mat.melting_temperature = 1811.0
+        mat.boiling_temperature = 3134.0
+        mat.density = sp.Float(7850)
+        return mat
 
     def test_visualizer_no_figure_available(self, mock_parser, sample_material):
-        """Test visualization when no figure is available."""
+        """Property skipped silently when plots are not initialised."""
         visualizer = PropertyVisualizer(mock_parser)
-        # Don't initialize plots
-        T = sp.Symbol('T')
-        # Should not raise error, just log warning
         visualizer.visualize_property(
             material=sample_material,
             prop_name='density',
-            T=T,
-            prop_type='CONSTANT'
+            T=sp.Symbol('T'),
+            prop_type='CONSTANT',
         )
-        # Property should not be added to visualized set
         assert 'density' not in visualizer.visualized_properties
 
     def test_visualizer_numeric_temperature_skip(self, mock_parser, sample_material):
-        """Test that visualization is skipped for numeric temperature."""
+        """Numeric temperature causes visualizer to skip the property."""
         visualizer = PropertyVisualizer(mock_parser)
         visualizer.initialize_plots()
-        # Should skip visualization for numeric temperature
         visualizer.visualize_property(
             material=sample_material,
             prop_name='density',
-            T=500.0,  # Numeric temperature
-            prop_type='CONSTANT'
+            T=500.0,
+            prop_type='CONSTANT',
         )
         assert 'density' not in visualizer.visualized_properties
 
-    def test_visualizer_disabled(self, mock_parser, sample_material):
-        """Test visualization when disabled."""
+    def test_visualizer_disabled(self, mock_parser):
+        """Disabled visualizer does not initialise figure."""
         visualizer = PropertyVisualizer(mock_parser)
         visualizer.is_enabled = False
-        # Should not initialize plots
         visualizer.initialize_plots()
         assert visualizer.fig is None
 
[…]
