+++ tests/integration/test_end_to_end.py
@@ -1,194 +1,138 @@
 """End-to-end integration tests."""
+import tempfile
+from pathlib import Path
 
 import pytest
-import tempfile
 from ruamel.yaml import YAML
-from pathlib import Path
 
 from materforge.parsing.api import create_material
 
 class TestEndToEnd:
     """End-to-end integration tests."""
-    def test_create_aluminum_material_from_existing_yaml(self, temp_symbol):
-        """Test material creation from existing aluminum YAML file."""
-        # Try multiple possible paths for the aluminum YAML file
-        possible_paths = [
-            Path("src/materforge/data/materials/pure_metals/Al/Al.yaml"),
-            Path("../src/materforge/data/materials/pure_metals/Al/Al.yaml"),
-            Path("materforge/data/materials/pure_metals/Al/Al.yaml"),
-        ]
-        aluminum_yaml_path = None
-        for path in possible_paths:
-            if path.exists():
-                aluminum_yaml_path = path
-                break
-        # Skip test if file doesn't exist
-        if aluminum_yaml_path is None or not aluminum_yaml_path.exists():
-            pytest.skip(f"Aluminum YAML file not found. Tried paths: {[str(p) for p in possible_paths]}")
-        # Create material
+    def test_create_aluminum_material_from_existing_yaml(self, aluminum_yaml_path, temp_symbol):
+        """Material creation from existing aluminum YAML file."""
+        if not aluminum_yaml_path.exists():
+            pytest.skip(f"Aluminum YAML not found: {aluminum_yaml_path}")
         material = create_material(aluminum_yaml_path, temp_symbol, enable_plotting=False)
-        # Verify material properties
-        assert material.name == "Aluminum"
-        assert material.material_type == "pure_metal"
-        assert len(material.elements) == 1
-        assert material.elements[0].name in ["Aluminum", "Aluminium"]  # Handle both spellings
-        # Verify processed properties
-        assert hasattr(material, 'density')
-        # Handle SymPy expressions properly
-        if hasattr(material.density, 'evalf'):
-            try:
-                density_value = float(material.density.evalf())
-            except (TypeError, ValueError):
-                # If evalf() fails, try substituting a temperature value
-                density_value = float(material.density.subs(temp_symbol, 300))
-        elif hasattr(material.density, 'subs'):
-            # If it's a symbolic expression, substitute a temperature value
-            density_value = float(material.density.subs(temp_symbol, 300))
-        else:
-            density_value = float(material.density)
-        assert density_value > 0
+        assert "Aluminum" in material.name or "Al" in material.name
+        assert 'density' in material.property_names()
+        density_val = float(material.density) if isinstance(
+            material.density, (int, float)) else float(
+            material.density.subs(temp_symbol, 300))
+        assert density_val > 0
 
-    def test_create_steel_material_from_existing_yaml(self, temp_symbol):
-        """Test material creation from existing steel YAML file."""
-        # Try multiple possible paths for the steel YAML file
-        possible_paths = [
-            Path("src/materforge/data/materials/alloys/1.4301/1.4301.yaml"),
-            Path("../src/materforge/data/materials/alloys/1.4301/1.4301.yaml"),
-            Path("materforge/data/materials/alloys/1.4301/1.4301.yaml"),
-        ]
-        steel_yaml_path = None
-        for path in possible_paths:
-            if path.exists():
-                steel_yaml_path = path
-                break
-        # Skip test if file doesn't exist
-        if steel_yaml_path is None or not steel_yaml_path.exists():
-            pytest.skip(f"Steel YAML file not found. Tried paths: {[str(p) for p in possible_paths]}")
-        # Create material
+    def test_create_steel_material_from_existing_yaml(
+        self, steel_yaml_path, temp_symbol
+    ):
[…]
