### 2026-02-25 df4b720
Add __getattr__ to suppress Pylint E1101 false positives
Dynamic properties assigned via setattr at runtime are invisible to
static analysers. Defining __getattr__ signals to Pylint and mypy that
the class handles arbitrary attribute access, eliminating no-member
warnings at all callsites without inline suppressions.
