### 2026-02-25 d55a112
Require explicit symbol in property evaluation, remove include_constants and properties subset params
