+++ examples/material_properties_demo.py
@@ -37,6 +37,7 @@ def _scalar_properties(mat) -> dict:
         and isinstance(getattr(mat, n), (int, float, sp.Float, sp.Integer))
     }
 
+
 def demonstrate_material_properties() -> None:
     """Comprehensive demonstration of material property evaluation."""
     setup_logging()
@@ -103,7 +104,7 @@ def demonstrate_material_properties() -> None:
     print(f"{'-' * 50}")
 
     for path, symbol, label in [
-        (yaml_path_Al,    T1, "Aluminum"),
+        (yaml_path_Al,     T1, "Aluminum"),
         (yaml_path_SS304L, T2, "Steel 1.4301"),
     ]:
         if not path.exists():
@@ -147,7 +148,6 @@ def demonstrate_material_properties() -> None:
         print(f"{'=' * 80}")
         print(f"Name: {mat.name}")
 
-        # Show scalar constants (temperature references, fixed densities, etc.)
         scalars = _scalar_properties(mat)
         if scalars:
             print("Scalar constants:")
@@ -168,7 +168,7 @@ def demonstrate_material_properties() -> None:
             print(f"  {prop_name:<30}: {prop_value}")
             print(f"  {'':30}  Type: {type(prop_value).__name__}")
             allowed_types = (sp.Piecewise, sp.Float, sp.Expr, float, int, type(None))
-            marker = ("Valid type" if isinstance(prop_value, allowed_types) else "! WARNING: Unexpected type")
+            marker = "Valid type" if isinstance(prop_value, allowed_types) else "! WARNING: Unexpected type"
             print(f"  {'':30}  {marker}")
 
         # -------------------------------------------------------------------
@@ -196,47 +196,25 @@ def demonstrate_material_properties() -> None:
         print(f"\n{'API METHODS:':<50}")
         print(f"{'-' * 50}")
 
-        print("Method 2: material.evaluate_properties_at_temperature()")
+        print("Method 2: material.evaluate_properties_at_temperature(symbol, value)")
+        all_values = None
         try:
-            all_values = mat.evaluate_properties_at_temperature(test_temp)
+            all_values = mat.evaluate_properties_at_temperature(T_mat, test_temp)
             print(f"All properties at {test_temp} K:")
             for prop, value in sorted(all_values.items()):
                 print(f"  {prop:<30}: {value:.6e}")
         except Exception as e:
             print(f"Error in Method 2: {e}")
 
-        print("\nMethod 3: evaluate_material_properties()")
+        print("\nMethod 3: evaluate_material_properties(material, symbol, value)")
         try:
-            all_values_func = evaluate_material_properties(mat, test_temp)
-            print(f"  Results match Method 2: {all_values == all_values_func}")
+            all_values_func = evaluate_material_properties(mat, T_mat, test_temp)
+            if all_values is not None:
+                print(f"  Results match Method 2: {all_values == all_values_func}")
+            else:
+                print("  Functional API evaluation completed successfully")
         except Exception as e:
             print(f"Error in Method 3: {e}")
-
-        print("\nMethod 4: Specific Properties Only")
-        try:
-            specific_props = sorted(available_props)[:3]
-            if specific_props:
-                specific_values = mat.evaluate_properties_at_temperature(
-                    test_temp, properties=specific_props)
-                print(f"  Requested: {specific_props}")
-                for prop, value in sorted(specific_values.items()):
-                    print(f"  {prop:<30}: {value:.6e}")
-            else:
-                print("  No properties available")
-        except Exception as e:
-            print(f"Error in Method 4: {e}")
-
-        print("\nMethod 5: Temperature-Dependent Properties Only")
-        try:
-            temp_dependent = mat.evaluate_properties_at_temperature(
[…]
