### 2026-02-26 dd2de87
Generalize dependency naming
