+++ src/materforge/parsing/validation/property_type_detector.py
@@ -5,7 +5,6 @@ import logging
 import re
 from enum import auto, Enum
 from typing import Any, Dict, Set
-
 import sympy as sp
 
 from materforge.parsing.config.yaml_keys import (
@@ -17,9 +16,8 @@ from materforge.data.constants import ProcessingConstants
 
 logger = logging.getLogger(__name__)
 
-# Valid Python identifier - used to accept any user-defined scalar reference
+# Valid Python identifier - accepts any user-defined scalar reference.
 _IDENTIFIER_RE = re.compile(r'^[a-zA-Z_][a-zA-Z0-9_]*$')
-
 
 class PropertyType(Enum):
     CONSTANT_VALUE = auto()
@@ -30,9 +28,8 @@ class PropertyType(Enum):
     COMPUTED_PROPERTY = auto()
     INVALID = auto()
 
-
 class PropertyTypeDetector:
-    """Utility class for detecting and validating property types from YAML config values."""
+    """Detects and validates property types from YAML config values."""
 
     # Detection rules - order is critical: most specific first.
     DETECTION_RULES = [
@@ -45,15 +42,18 @@ class PropertyTypeDetector:
         (lambda c: DEPENDENCY_KEY in c and EQUATION_KEY in c and isinstance(c.get(EQUATION_KEY), str),
          PropertyType.COMPUTED_PROPERTY),
     ]
+    # Validator dispatch table - built once at class definition time.
+    _VALIDATOR_MAP = None  # populated lazily via _get_validator_map()
 
     # --- Public API ---
+
     @staticmethod
     def determine_property_type(prop_name: str, config: Any) -> PropertyType:
-        """Determines the property type using a rule-based approach.
+        """Determines the PropertyType for a config entry using rule-based detection.
 
         Args:
             prop_name: Name of the property (used in error messages only).
-            config: Raw config value from the YAML properties block.
+            config:    Raw config value from the YAML properties block.
         Returns:
             The detected PropertyType.
         Raises:
@@ -64,50 +64,45 @@ class PropertyTypeDetector:
             return PropertyType.CONSTANT_VALUE
         if not isinstance(config, dict):
             raise ValueError(f"Property '{prop_name}' has an invalid format. "
-                             f"Expected a dictionary or a numeric constant, got {type(config).__name__}.")
+                f"Expected a dictionary or a numeric constant, got {type(config).__name__}.")
         for detector, prop_type in PropertyTypeDetector.DETECTION_RULES:
             if detector(config):
                 logger.debug("Detected '%s' as %s", prop_name, prop_type.name)
                 return prop_type
-        raise ValueError(f"Property '{prop_name}' doesn't match any known configuration pattern. "
+        raise ValueError(f"Property '{prop_name}' does not match any known configuration pattern. "
             f"Present keys: {sorted(config.keys())}.")
 
     @staticmethod
     def validate_property_config(prop_name: str, config: Any,
                                   prop_type: PropertyType) -> None:
         """Performs strict structural validation for the detected property type.
+
         Args:
             prop_name: Name of the property.
-            config: Raw config value from the YAML properties block.
+            config:    Raw config value from the YAML properties block.
             prop_type: The already-detected PropertyType.
         Raises:
-            ValueError: If config structure is invalid for the given type.
+            ValueError:          If config structure is invalid for the given type.
             NotImplementedError: If no validator exists for prop_type.
         """
         logger.debug("Validating '%s' as %s", prop_name, prop_type.name)
-        validator_map = {
[…]
