+++ src/materforge/core/__init__.py
@@ -5,18 +5,13 @@
 Core data structures and material definitions.
 
 This module contains the fundamental classes and interfaces that define
-materials, chemical elements, and the core abstractions used throughout
-the MaterForge library.
+materials, and the core abstractions used throughout the MaterForge library.
 """
 
 from .materials import Material
 from .symbol_registry import SymbolRegistry
-from .exceptions import MaterialError, MaterialCompositionError, MaterialTemperatureError
 
 __all__ = [
     "Material",
     "SymbolRegistry",
-    "MaterialError",
-    "MaterialCompositionError",
-    "MaterialTemperatureError"
 ]

[…]
