### 2026-02-27 2392c75
modifications due to reviewer comments from JOSS manuscript
- major change in core/materials.py; all properties are stored in a dict, where all properties are added via the __setattr__ function; this change avoids the problem during initialization of the class, where the try / except part is now obsolete
- simplified evaluation function in core/materials.py; additionally, this function now returns a copy of the material with the evaluated properties
- removed numeric mode, i.e. the dependency is always a sp.Symbol; In case constant material properties are intended, use the material evaluate fu
[…]
