+++ src/materforge/data/constants/__init__.py
@@ -3,11 +3,9 @@
 
 """Physical and processing constants for MaterForge."""
 
-from .physical_constants import PhysicalConstants
 from .processing_constants import ProcessingConstants, ErrorMessages, FileConstants
 
 __all__ = [
-    "PhysicalConstants",
     "ProcessingConstants",
     "ErrorMessages",
     "FileConstants"
