### 2026-02-27 7fc8cbf
fix im apps/1.4301_HeatEquationKernelWithMaterial.yaml
