### 2026-03-03 678b222
Simpify repository structure in materials directory, update tests
