+++ examples/myAlloy.py
@@ -14,6 +14,11 @@ from materforge.parsing.api import (
     get_material_info,
 )
 
+MATERIAL_CONFIGS = [
+    #("src/materforge/data/materials/Al.yaml", "T"),
+    #("src/materforge/data/materials/1.4301.yaml", "T"),
+    ("examples/myAlloy.yaml", "u_C"),
+]
 
 def setup_logging() -> None:
     logging.basicConfig(
@@ -30,20 +35,22 @@ def _scalar_properties(mat) -> dict:
     return {
         n: getattr(mat, n)
         for n in mat.property_names()
-        if not (hasattr(getattr(mat, n), 'free_symbols')
-                and getattr(mat, n).free_symbols)
+        if not (hasattr(getattr(mat, n), 'free_symbols') and getattr(mat, n).free_symbols)
         and isinstance(getattr(mat, n), (int, float, sp.Float, sp.Integer))
     }
 
 def demonstrate_material_properties() -> None:
     setup_logging()
 
-    T = sp.Symbol('T')
-
+    repo_root = Path(__file__).parent.parent
     current_file = Path(__file__)
-    yaml_path = current_file.parent / "myAlloy.yaml"
-    # yaml_path = (current_file.parent.parent / "src" / "materforge" / "data" / "materials" / "pure_metals" / "Al" / "Al.yaml")
-    # yaml_path = (current_file.parent.parent / "src" / "materforge" / "data" / "materials" / "alloys" / "1.4301" / "1.4301.yaml")
+
+    targets = []
+    for rel_path, sym_name in MATERIAL_CONFIGS:
+        yaml_path = repo_root / rel_path
+        if not yaml_path.exists():
+            raise FileNotFoundError(f"YAML file not found: {yaml_path} (configured in {current_file})")
+        targets.append((yaml_path, sp.Symbol(sym_name)))
 
     materials = []
     material_symbols = {}
@@ -58,14 +65,15 @@ def demonstrate_material_properties() -> None:
     print(f"\n{'1. YAML FILE VALIDATION':<50}")
     print(f"{'-' * 50}")
 
-    if yaml_path.exists():
-        try:
-            is_valid = validate_yaml_file(yaml_path)
-            print(f"myAlloy YAML validation: {'PASSED' if is_valid else 'FAILED'}")
-        except Exception as e:
-            raise ValueError(f"myAlloy YAML validation: FAILED - {e}")
-    else:
-        raise ValueError(f"myAlloy YAML file not found: {yaml_path}")
+    for yaml_path, _ in targets:
+        if yaml_path.exists():
+            try:
+                is_valid = validate_yaml_file(yaml_path)
+                print(f"{yaml_path.stem} YAML validation: {'PASSED' if is_valid else 'FAILED'}")
+            except Exception as e:
+                raise ValueError(f"{yaml_path.stem} YAML validation FAILED: {e}") from e
+        else:
+            raise ValueError(f"YAML file not found: {yaml_path}")
 
     # ===================================================================
     # 2. MATERIAL INFO EXTRACTION
@@ -73,10 +81,12 @@ def demonstrate_material_properties() -> None:
     print(f"\n{'2. MATERIAL INFO EXTRACTION':<50}")
     print(f"{'-' * 50}")
 
-    if yaml_path.exists():
+    for yaml_path, _ in targets:
+        if not yaml_path.exists():
+            continue
         try:
             info = get_material_info(yaml_path)
-            print("\nmyAlloy Information:")
+            print(f"\n{info['name']} Information:")
             print(f"  Name:             {info['name']}")
             print(f"  Total Properties: {info['total_properties']}")
             print(f"  Properties:       {info['properties']}")
[…]
