+++ tests/performance/test_benchmarks.py
@@ -1,153 +1,0 @@
-"""Performance benchmarks for MaterForge."""
-
-import pytest
-import time
-import numpy as np
-import sympy as sp
-import os
-
-from materforge.algorithms.interpolation import interpolate_value
-from materforge.algorithms.piecewise_builder import PiecewiseBuilder
-
-# Try to import psutil, skip memory tests if not available
-try:
-    import psutil
-    PSUTIL_AVAILABLE = True
-except ImportError:
-    PSUTIL_AVAILABLE = False
-
-def get_performance_threshold(base_threshold: float) -> float:
-    """Adjust thresholds based on environment"""
-    if os.getenv('CI'):
-        # Allow 2x slower performance in CI
-        return base_threshold * 2
-    return base_threshold
-
-class TestPerformance:
-    """Performance benchmark tests."""
-    @pytest.mark.slow
-    def test_interpolation_performance(self):
-        """Test interpolation performance with large datasets."""
-        # Large dataset
-        x_array = np.linspace(300, 1000, 10000)
-        y_array = np.sin(x_array / 100) * 1000 + 1000
-        # Time multiple interpolations
-        start_time = time.time()
-        for _ in range(1000):
-            _ = interpolate_value(500.0, x_array, y_array, 'constant', 'constant')
-        end_time = time.time()
-        avg_time = (end_time - start_time) / 1000
-        # Adjust threshold for CI environment
-        threshold = get_performance_threshold(0.005)  # 10ms in CI vs 5ms locally
-        assert avg_time < threshold, f"Interpolation too slow: {avg_time:.6f}s per call (threshold: {threshold:.6f}s)"
-
-    @pytest.mark.slow
-    def test_piecewise_creation_performance(self):
-        """Test piecewise function creation performance."""
-        T = sp.Symbol('T')
-        temp_array = np.linspace(300, 1000, 1000)
-        prop_array = np.sin(temp_array / 100) * 1000 + 1000
-        config = {'bounds': ['constant', 'constant']}
-        start_time = time.time()
-        piecewise_func = PiecewiseBuilder.build_from_data(
-            temp_array, prop_array, T, config, "test_property"
-        )
-        end_time = time.time()
-        creation_time = end_time - start_time
-        # Adjust threshold for CI environment
-        threshold = get_performance_threshold(2.0)  # 4s in CI vs 2s locally
-        assert creation_time < threshold, f"Piecewise creation too slow: {creation_time:.3f}s (threshold: {threshold:.1f}s)"
-        # Test evaluation performance
-        start_time = time.time()
-        for temp in [350, 450, 550, 650, 750]:
-            _ = float(piecewise_func.subs(T, temp))
-        end_time = time.time()
-        eval_time = (end_time - start_time) / 5
-        # Adjust threshold for CI environment
-        eval_threshold = get_performance_threshold(0.06)  # 120ms in CI vs 60ms locally
-        assert eval_time < eval_threshold, f"Piecewise evaluation too slow: {eval_time:.6f}s per call (threshold: {eval_threshold:.6f}s)"
-
-    @pytest.mark.slow
-    def test_large_dataset_performance(self):
-        """Test performance with very large datasets."""
-        T = sp.Symbol('T')
-        # Very large dataset
-        temp_array = np.linspace(300, 2000, 5000)
-        prop_array = np.random.normal(1000, 100, 5000)  # Random data
-        config = {'bounds': ['constant', 'constant']}
-        start_time = time.time()
-        piecewise_func = PiecewiseBuilder.build_from_data(
-            temp_array, prop_array, T, config, "large_property"
-        )
-        end_time = time.time()
-        creation_time = end_time - start_time
-        # Adjust threshold for CI environment
-        threshold = get_performance_threshold(10.0)  # 20s in CI vs 10s locally
[…]
