+++ src/materforge/data/materials/pure_metals/__init__.py
@@ -1,16 +1,0 @@
-# SPDX-FileCopyrightText: 2025 Rahil Miten Doshi, Friedrich-Alexander-Universität Erlangen-Nürnberg
-# SPDX-License-Identifier: BSD-3-Clause
-
-"""Pure metal material configuration files."""
-
-# This directory contains YAML configuration files for pure metal materials
-# Each YAML file defines a complete material specification including:
-# - Single element composition (fraction = 1.0)
-# - Melting and boiling temperatures
-# - Material properties with various definition formats
-
-# Available pure metal configurations:
-# - Al/Al.yaml: Pure Aluminum with examples of STEP_FUNCTION, KEY_VAL,
-#   PIECEWISE_EQUATION, and COMPUTE property types
-
-__all__ = []
