+++ src/materforge/data/materials/alloys/__init__.py
@@ -1,17 +1,0 @@
-# SPDX-FileCopyrightText: 2025 Rahil Miten Doshi, Friedrich-Alexander-Universität Erlangen-Nürnberg
-# SPDX-License-Identifier: BSD-3-Clause
-
-"""Alloy material configuration files."""
-
-# This directory contains YAML configuration files for alloy materials
-# Each YAML file defines a complete material specification including:
-# - Multi-element composition with fractions
-# - Solidus and liquidus temperatures
-# - Material properties with various definition formats
-
-# Available alloy configurations:
-# - 1.4301.yaml: Stainless Steel 304L with comprehensive property definitions
-#   Includes examples of all 6 property types: CONSTANT, STEP_FUNCTION,
-#   FILE, KEY_VAL, PIECEWISE_EQUATION, and COMPUTE
-
-__all__ = []

[…]
