### 2026-03-04 b5890f7
Restore T as fixed YAML placeholder
