### 2026-03-04 ea64ad4
Enforce strict top-level YAML schema validation
