### 2026-03-04 f3c7d1d
Add yaml config for ceramic Al2O3, minor cleanup
