+++ examples/myAlloy.py
@@ -15,9 +15,10 @@ from materforge.parsing.api import (
 )
 
 MATERIAL_CONFIGS = [
-    #("src/materforge/data/materials/Al.yaml", "T"),
-    #("src/materforge/data/materials/1.4301.yaml", "T"),
-    ("examples/myAlloy.yaml", "u_C"),
+    #("src/materforge/data/materials/Al.yaml", "T_Al"),
+    #("src/materforge/data/materials/1.4301.yaml", "T_SS"),
+    #("examples/myAlloy.yaml", "u_C"),
+    ("src/materforge/data/materials/Al2O3.yaml", "T_Al2O3"),
 ]
 
 def setup_logging() -> None:
