### 2026-03-04 b0d94d6
Update HeatEquationKernelWithMaterial app
