### 2026-03-04 a10d96c
Rename "extrapolate" keyword to "linear" due to linear extrapolation
