+++ tests/unit/test_algorithms/test_inverse_1.py
@@ -5,8 +5,6 @@ import logging
 from pathlib import Path
 
 from materforge.parsing.api import create_material
-from materforge.parsing.config.yaml_keys import CONSTANT_KEY, LINEAR_KEY
-
 
 logger = logging.getLogger(__name__)
 
