### 2026-03-05 112509b
Eliminate duplicate error logs, update docs, nminor cleanup
