### 2026-03-09 65397c3
Introduce typed config exceptions
