+++ src/materforge/parsing/config/material_yaml_parser.py
@@ -8,6 +8,7 @@ import sympy as sp
 from ruamel.yaml import YAML, constructor, scanner
 from materforge.core.materials import Material
 from materforge.parsing.processors.property_processor import PropertyProcessor
+from materforge.parsing.validation.errors import MaterialConfigError, PropertyConfigError
 from materforge.parsing.validation.property_type_detector import PropertyType, PropertyTypeDetector
 from materforge.visualization.plotters import PropertyVisualizer
 from materforge.parsing.config.yaml_keys import NAME_KEY, PROPERTIES_KEY
@@ -37,7 +38,7 @@ class YAMLFileParser(BaseFileParser):
         yaml.allow_duplicate_keys = False
         try:
             logger.debug("Loading YAML file: %s", self.config_path)
-            with open(self.config_path, 'r') as f:
+            with open(self.config_path, 'r', encoding='utf-8') as f:
                 config = yaml.load(f)
             logger.debug("YAML loaded - %d top-level keys", len(config) if config else 0)
             return config
@@ -57,6 +58,7 @@ class YAMLFileParser(BaseFileParser):
 
 class MaterialYAMLParser(YAMLFileParser):
     """Parser for material YAML configuration files.
+
     Validates and processes a schema-agnostic YAML file containing a name
     and a properties block. All thermophysical properties live in the properties
     block and are assigned dynamically to the Material instance.
@@ -75,7 +77,7 @@ class MaterialYAMLParser(YAMLFileParser):
     # --- Public API ---
     def create_material(self, dependency: sp.Symbol, enable_plotting: bool = True) -> Material:
         """Creates a Material instance from the parsed configuration.
-    
+
         Args:
             dependency:      SymPy symbol used as the independent variable in
                              property expressions (e.g. sp.Symbol('T')).
@@ -83,7 +85,7 @@ class MaterialYAMLParser(YAMLFileParser):
         Returns:
             Fully initialised Material with all properties assigned.
         Raises:
-            ValueError: If configuration is invalid or property processing fails.
+            MaterialConfigError: If configuration is invalid or property processing fails.
         """
         name = self.config.get(NAME_KEY, "Unnamed Material")
         logger.info("Creating material: '%s'", name)
@@ -98,8 +100,8 @@ class MaterialYAMLParser(YAMLFileParser):
                 logger.info("Visualization enabled")
             else:
                 logger.debug("Visualization disabled - %s",
-                             "numeric dependency" if not isinstance(dependency, sp.Symbol)
-                             else "plotting not requested")
+                    "numeric dependency" if not isinstance(dependency, sp.Symbol)
+                    else "plotting not requested")
             logger.info("Processing properties for '%s'", name)
             self.property_processor.process_properties(
                 material=material,
@@ -115,51 +117,53 @@ class MaterialYAMLParser(YAMLFileParser):
             logger.info("Successfully created material: '%s'", name)
             return material
         except Exception as e:
-            # Intermediate layer — do not log here, bubble up to api.py
-            raise ValueError(f"Failed to create material \n -> {str(e)}") from e
+            # Intermediate layer - do not log here, bubble up to api.py
+            raise MaterialConfigError(f"Failed to create material \n -> {str(e)}") from e
 
     # --- Validation ---
     def _validate_config(self) -> None:
         """Validates the top-level YAML structure.
+
         Raises:
-            ValueError: If the config is not a dict, contains unknown top-level keys,
-                or if name/properties fields are missing or invalid.
+            MaterialConfigError: If the config is not a dict, contains unknown
[…]
