### 2026-03-09 2d60941
Correct evaluate() docstring return type description
