### 2026-03-09 077b512
Fix broken evaluate() example and remove duplicate paragraph
