### 2026-03-09 0203989
Correct create_inverse() args and evaluate() usage in how-to guides
