### 2026-03-09 60cb926
Correct evaluate(), create_inverse(), and u.center() usage
