+++ CHANGELOG.md
@@ -5,6 +5,87 @@ All notable changes to MaterForge will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.6.5] - 2026-03-09
+
+### Added
+- `MaterialConfigError` and `PropertyConfigError` typed exception classes in `errors.py`;
+  both subclass `ValueError` so existing callers are unaffected
+- `UnicodeDecodeError` handler in `YAMLFileParser._load_config` with a clear actionable
+  message ("Save as UTF-8 and try again") instead of a generic crash
+- Al2O3 ceramic YAML material configuration
+- `__getattr__` on `Material` class to suppress Pylint E1101 false positives for
+  dynamically assigned properties
+- `docs/_static/.gitkeep` to ensure the directory is tracked in git and Sphinx
+  `html_static_path` resolves correctly in CI
+
+### Changed
+- **Breaking**: renamed `extrapolate` boundary keyword to `linear` in all YAML files
+  to accurately reflect linear extrapolation behaviour; update any custom YAML files
+- Enforced strict top-level YAML schema: only `name` and `properties` keys are allowed;
+  unknown keys raise `MaterialConfigError` immediately
+- Removed `material_type`, `composition`, and other non-essential top-level YAML keys;
+  schema is now intentionally minimal
+- Replaced fixed property list with dynamic tracker across all processors;
+  removed `get_supported_properties` from public API
+- All properties on `Material` are now stored in an internal dict via `__setattr__`;
+  `evaluate()` returns a new `Material` copy with all expressions substituted at the
+  given value - not a dict
+- Dependency is now always a `sp.Symbol`; numeric mode removed - use
+  `material.evaluate(symbol, value)` for constant properties instead
+- `T` restored as the fixed YAML placeholder symbol; equations in YAML files must
+  always use `T`, which is substituted at runtime with the caller-provided symbol
+- `_validate_config` now raises `MaterialConfigError` (was `ValueError`)
+- `_analyze_and_categorize_properties` now raises `PropertyConfigError` (was `ValueError`)
+- `create_material` in `material_yaml_parser.py` now wraps failures in
+  `MaterialConfigError` (was plain `ValueError`)
+- `validate_yaml_file` and `get_material_info` in `api.py` now re-raise typed
+  exceptions directly instead of wrapping everything in a new `ValueError`
+- `get_material_property_names` and `evaluate_material_properties` now raise
+  `TypeError` for wrong argument types (was `ValueError`)
+- `open()` call in `YAMLFileParser` now explicitly specifies `encoding='utf-8'`
+  for platform-independent behaviour (fixes Pylint W1514)
+- `DependencyError.__init__`: `available_props: List[str] = None` →
+  `Optional[List[str]] = None` (fixes Pylint type hint warning)
+- Simplified materials directory structure
+
+### Fixed
+- Removed unused exception variable bindings `as e` in bare `raise` handlers
+  (Ruff F841); affected `validate_yaml_file` and `get_material_info` in `api.py`
+- Eliminated duplicate error log entries that occurred when exceptions bubbled
+  through multiple layers
+- Removed unused imports across `api.py`, `material_yaml_parser.py`, and app files
+
+### Testing
+- Updated `test_api.py` to match refactored exception types:
+  - `test_get_material_property_names_rejects_non_material`: `ValueError` → `TypeError`
+  - `test_validate_incomplete_yaml_raises_value_error`: `ValueError` → `MaterialConfigError`
+- Moved error class imports to module level in test files to prevent `pytest.raises`
[…]
