+++ src/materforge/parsing/validation/array_validator.py
@@ -9,7 +9,6 @@ from materforge.data.constants import ProcessingConstants
 
 logger = logging.getLogger(__name__)
 
-
 def is_monotonic(arr: np.ndarray, name: str = "Array",
                  mode: str = "strictly_increasing",
                  threshold: float = ProcessingConstants.MONOTONICITY_THRESHOLD,
