### 2026-03-09 aa46dd0
Add Matthias Markl copyright headers to co-authored files
