### 2026-03-09 ae95733
Bump version to 0.6.5, update LICENSE and CITATION.cff
