+++ CITATION.cff
@@ -6,9 +6,13 @@ authors:
     email: "rahil.doshi@fau.de"
     affiliation: "Friedrich-Alexander-Universität Erlangen-Nürnberg"
     orcid: "0009-0008-3570-9841"
+contributors:
+  - family-names: "Markl"
+    given-names: "Matthias"
+    affiliation: "Friedrich-Alexander-Universität Erlangen-Nürnberg"
 title: "MaterForge: Materials Formulation Engine with Python"
-version: "0.6.4"
-date-released: "2026-02-22"
+version: "0.6.5"
+date-released: "2026-03-09"
 url: "https://i10git.cs.fau.de/rahil.doshi/materforge"
 repository-code: "https://i10git.cs.fau.de/rahil.doshi/materforge"
 license: "BSD-3-Clause"
