+++ pyproject.toml
@@ -106,4 +106,4 @@ materforge = [
 version_file = "src/materforge/_version.py"
 version_scheme = "python-simplified-semver"
 local_scheme = "node-and-date"
-fallback_version = "0.6.4"
+fallback_version = "0.6.5"
