### 2026-03-09 93ae2df
Update paper.md as per JOSS reviewer feedback
