+++ paper/paper.md
@@ -24,7 +24,7 @@ affiliations:
     index: 2
   - name: Erlangen National High Performance Computing Center (NHR@FAU), Erlangen, Germany
     index: 3
-date: 20 August 2025
+date: 09 March 2026
 bibliography: paper.bib
 ---
 
@@ -35,7 +35,7 @@ material properties in numerical simulations.
 The library supports complex material behaviors, from simple constants to experimental data
 in user-friendly YAML configurations.
 These are internally converted into symbolic mathematical expressions for scientific computing frameworks.
-MaterForge supports various material types,
+MaterForge supports any material type through a schema-agnostic design,
 provides flexible property definitions,
 and automatically resolves dependency order for derived properties while detecting cycles.
 It is designed for high-performance computing (HPC) applications
@@ -53,14 +53,41 @@ creating significant integration hurdles for researchers.
 To manage this complexity, researchers often resort to manual interpolation, custom scripting, or proprietary software,
 which compromises reproducibility and standardization [@ashby2013materials].
 While valuable resources like the NIST WebBook [@linstrom2001nist] and CoolProp [@coolprop] provide valuable raw data,
-they lack integrated processing to unify these varied formats.
-CALPHAD databases [@calphad] are powerful but often require proprietary software
-and do not easily integrate with general-purpose simulation codes.
+they lack integrated processing to unify these varied formats into simulation-ready symbolic expressions.
+Thermodynamic modeling tools such as pycalphad [@pycalphad] and CALPHAD databases [@calphad]
+are powerful for phase equilibria calculations but operate at a different layer of the workflow:
+they generate property data, not simulation-ready expressions.
 
-This leads to ad hoc solutions, hindering workflow efficiency and FAIR data adoption [@wilkinson2016fair].
-MaterForge bridges this gap by providing a unified framework that leverages
-symbolic mathematics, automatic regression, and dependency resolution
-to standardize and simplify the integration of realistic material behavior into scientific simulations.
+This creates a gap between property data generation and simulation integration,
+leading to ad hoc solutions that hinder workflow efficiency and FAIR data adoption [@wilkinson2016fair].
+MaterForge fills this gap by occupying a dedicated post-processing layer in the materials simulation workflow.
+
+# Position in the Simulation Workflow
+
+MaterForge is designed to operate as the intermediate layer in a three-stage workflow:
+
+**Stage 1 - Property Data Generation**: Raw material property data is produced by
+upstream tools such as pycalphad [@pycalphad] for thermodynamic phase equilibria,
+CoolProp [@coolprop] or the NIST WebBook [@linstrom2001nist] for fluid properties,
+commercial tools such as JMatPro for alloy properties,
+or direct experimental measurement.
+These tools generate tabular data, equations, or database entries.
+
+**Stage 2 - MaterForge Post-Processing**: MaterForge ingests this data via YAML configuration files
+and converts it into optimized symbolic mathematical expressions using SymPy [@sympy].
+It performs automatic regression and data reduction, resolves inter-property dependencies,
+validates configurations, and generates visualization plots for verification.
+The output is a fully configured `Material` object with all properties expressed as
+SymPy expressions in a user-chosen independent variable (temperature, pressure, composition, or any symbol).
+
+**Stage 3 - Simulation Integration**: The symbolic expressions are passed directly into
+simulation frameworks such as pystencils [@pystencils] or waLBerla [@walberla]
+for code generation, or into any Python-based finite element or CFD solver.
+Because properties are SymPy expressions, they plug into symbolic assignment collections
+without any additional conversion.
+
+This layered design means MaterForge is complementary to, not competing with,
[…]
