+++ paper/paper.bib
@@ -92,6 +92,18 @@
     isbn = {978-0521868112}
 }
 
+@article{pycalphad,
+    title = {pycalphad: {CALPHAD}-based Computational Thermodynamics in Python},
+    author = {Otis, Richard and Liu, Zi-Kui},
+    journal = {Journal of Open Research Software},
+    volume = {5},
+    number = {1},
+    pages = {1},
+    year = {2017},
+    doi = {10.5334/jors.140},
+    url = {https://doi.org/10.5334/jors.140}
+}
+
 @article{voller1987fixed,
     title = {A fixed grid numerical modelling methodology for convection-diffusion mushy region phase-change problems},
     author = {Voller, Vaughan R. and Prakash, C.},
