### 2026-03-11 7d661fb
Update for JOSS plots
