### 2026-03-11 49632c5
Update JOSS figures
