+++ paper/paper.md
@@ -24,7 +24,7 @@ affiliations:
     index: 2
   - name: Erlangen National High Performance Computing Center (NHR@FAU), Erlangen, Germany
     index: 3
-date: 09 March 2026
+date: 12 March 2026
 bibliography: paper.bib
 ---
 
@@ -32,7 +32,7 @@ bibliography: paper.bib
 
 MaterForge is an extensible, open-source Python library that streamlines the definition and use of
 material properties in numerical simulations.
-The library supports complex material behaviors, from simple constants to experimental data
+The library supports complex material behaviors, from simple constants to experimental data,
 in user-friendly YAML configurations.
 These are internally converted into symbolic mathematical expressions for scientific computing frameworks.
 MaterForge supports any material type through a schema-agnostic design,
@@ -43,8 +43,8 @@ and serves as a bridge between experimental data and numerical simulation.
 
 # Statement of Need
 
-Accurate numerical simulation requires material properties such as thermal conductivity, density, and viscosity
-that depend on variables like temperature, pressure, or strain rate [@lewis1996finite].
+Accurate numerical simulations of physical processes rely on well-characterized material properties -
+quantities such as thermal conductivity, density, and viscosity that depend on state variables like temperature, pressure, or strain rate [@lewis1996finite].
 This challenge is compounded by the wide variation in data availability,
 from well-characterized models for established materials to sparse experimental points for novel materials.
 Property definitions consequently range from simple constants to complex tabular datasets or sophisticated equations,
@@ -52,7 +52,7 @@ creating significant integration hurdles for researchers.
 
 To manage this complexity, researchers often resort to manual interpolation, custom scripting, or proprietary software,
 which compromises reproducibility and standardization [@ashby2013materials].
-While valuable resources like the NIST WebBook [@linstrom2001nist] and CoolProp [@coolprop] provide valuable raw data,
+While valuable resources like the NIST WebBook [@linstrom2001nist] and CoolProp [@coolprop] provide raw data,
 they lack integrated processing to unify these varied formats into simulation-ready symbolic expressions.
 Thermodynamic modeling tools such as pycalphad [@pycalphad] and CALPHAD databases [@calphad]
 are powerful for phase equilibria calculations but operate at a different layer of the workflow:
@@ -60,7 +60,9 @@ they generate property data, not simulation-ready expressions.
 
 This creates a gap between property data generation and simulation integration,
 leading to ad hoc solutions that hinder workflow efficiency and FAIR data adoption [@wilkinson2016fair].
-MaterForge fills this gap by occupying a dedicated post-processing layer in the materials simulation workflow.
+MaterForge fills this gap by occupying a dedicated post-processing layer in the materials simulation workflow,
+complementing rather than competing with thermodynamic modeling tools:
+pycalphad generates the data; MaterForge prepares it for simulation.
 
 # Position in the Simulation Workflow
 
@@ -69,7 +71,7 @@ MaterForge is designed to operate as the intermediate layer in a three-stage wor
 **Stage 1 - Property Data Generation**: Raw material property data is produced by
 upstream tools such as pycalphad [@pycalphad] for thermodynamic phase equilibria,
 CoolProp [@coolprop] or the NIST WebBook [@linstrom2001nist] for fluid properties,
-commercial tools such as JMatPro for alloy properties,
+or commercial tools such as JMatPro and Thermo-Calc for alloy properties,
 or direct experimental measurement.
 These tools generate tabular data, equations, or database entries.
 
@@ -77,17 +79,13 @@ These tools generate tabular data, equations, or database entries.
 and converts it into optimized symbolic mathematical expressions using SymPy [@sympy].
 It performs automatic regression and data reduction, resolves inter-property dependencies,
 validates configurations, and generates visualization plots for verification.
[…]
