+++ paper/paper.bib
@@ -40,27 +40,26 @@
 @article{coolprop,
     author = {Bell, Ian H. and Wronski, Jorrit and Quoilin, Sylvain and Lemort, Vincent},
     title = {Pure and Pseudo-pure Fluid Thermophysical Property Evaluation and
-             the Open-Source Thermophysical Property Library CoolProp},
+             the Open-Source Thermophysical Property Library {CoolProp}},
     journal = {Industrial \& Engineering Chemistry Research},
     volume = {53},
     number = {6},
     pages = {2498--2508},
     year = {2014},
     doi = {10.1021/ie4033999},
-    URL = {http://pubs.acs.org/doi/abs/10.1021/ie4033999},
-    eprint = {http://pubs.acs.org/doi/pdf/10.1021/ie4033999}
+    url = {https://doi.org/10.1021/ie4033999}
 }
 
 @article{linstrom2001nist,
     author = {Linstrom, Peter J. and Mallard, William G.},
-    title = {The NIST Chemistry WebBook:  A Chemical Data Resource on the Internet},
+    title = {The {NIST} Chemistry {WebBook}: A Chemical Data Resource on the Internet},
     journal = {Journal of Chemical \& Engineering Data},
     volume = {46},
     number = {5},
     pages = {1059-1063},
     year = {2001},
     doi = {10.1021/je000236i},
-    URL = {https://doi.org/10.1021/je000236i},
+    url = {https://doi.org/10.1021/je000236i}
 }
 
 @book{lewis1996finite,
@@ -85,7 +84,7 @@
 }
 
 @book{calphad,
-    title = {Computational Thermodynamics: The Calphad Method},
+    title = {Computational Thermodynamics: The {Calphad} Method},
     author = {Lukas, Hans Leo and Fries, Suzana G. and Sundman, Bo},
     year = 2007,
     publisher = {Cambridge University Press},
@@ -150,7 +149,7 @@
 }
 
 @article{wilkinson2016fair,
-    title = {The FAIR Guiding Principles for scientific data management and stewardship},
+    title = {The {FAIR} Guiding Principles for scientific data management and stewardship},
     author = {Wilkinson, Mark D. and Dumontier, Michel and Aalbersberg, IJsbrand Jan and Appleton, Gabrielle and Axton, Myles and Baak, Arie and Blomberg, Niklas and Boiten, Jan-Willem and da Silva Santos, Luiz Bonino and Bourne, Philip E. and others},
     journal = {Scientific Data},
     volume = 3,
