### 2026-03-12 72c955e
Accept integer constants in YAML config
