### 2026-03-12 #9909 @rahildoshi97
Hi @jan-janssen @JosePizarro3,

Thank you both for your thorough and constructive reviews. We have addressed all open points and published a new release **MaterForge v0.6.5** (now live on [PyPI](https://pypi.org/project/materforge/0.6.5/)).

Below is a summary of what changed.

**Source code**

- The YAML schema is now fully schema-agnostic: the only required top-level fields are `name` and `properties`. Any property name and any material kind are valid. `material_type`, `composition`, and all other previously required top-level keys have been removed.
- Element data tables and physical constants have been removed from the package entirely. These were needed to validate `composition` fields, which no longer exist in the simplified schema.
- Typed exception classes (`MaterialConfigError`, `PropertyConfigError`) replace bare `ValueError` throughout, making error handling more precise and easier to catch selectively.
- `evaluate()` now returns a new `Material` instance with all expressions substituted at the given value, rather than a plain dict.
- A worked example for a hypothetical alloy has been added to `examples/` as suggested.
- An Al2O3 ceramic configuration has been added, validating the schema-agnostic design for non-metal materials.

**Documentation**

- All how-to guides, tutorials, and API reference pages have been corrected to reflect the updated `evaluate()` return type, the `create_material()` symbol requirement, and the new exception types.

**Manuscript**

- The comparison table has been removed and replaced with an explicit three-stage workflow description (data generation -> MaterForge post-processing -> simulation integration), positioning pycalphad as a complementary upstream tool rather than a competitor, as suggested by @jan-janssen.
- An **AI Usage Disclosure** section has been added in accordance with JOSS policy.
- The YAML usage example in the paper has been updated to reflect the new minimal schema.
- The Python code example has been corrected to use `material.evaluate()` with the new `Material` return type.

We believe all checklist items and reviewer comments have now been addressed. Please let us know if anything requires further clarification or revision.

Best regards,
Rahil Doshi
