### 2026-03-12 #9909 @jan-janssen
@rahildoshi97 I again tested the examples in the paper and they do not work for me. So far I tested one and two with version 0.6.5.

## Example 1:
```
properties:
    density: 6950
```
It still requires the name and the density has to be a float
```
name: myAlloy
properties:
    density: 6950.0
```

## Example 2:
```
properties:
    latent_heat_of_fusion:
        dependency: solidus_temperature
        value: [0, 171401]
```
Again the name is missing and it complains about the dependency reference.

I try to run both examples using:
```
import sympy as sp
from materforge.parsing.api import create_material

solidus_temperature = sp.Symbol('T')
mat = create_material('test.yaml', dependency=T, enable_plotting=True)
```

Can you please fix the examples in the paper and provide sufficient documentation how the user can reproduce the results you show? In addition, can you include these examples in the repository that makes it easier to validate them.
