### 2026-03-12 9761a57
Update test to cover integer constant inputs
