### 2026-03-12 23b5e42
Fix figure 1 caption
