### 2026-03-12 #9909 @rahildoshi97
> [@rahildoshi97](https://github.com/rahildoshi97) I again tested the examples in the paper and they do not work for me. So far I tested one and two with version 0.6.5.
>
> ## Example 1:
> ```
> properties:
>     density: 6950
> ```
>
> It still requires the name and the density has to be a float
>
> ```
> name: myAlloy
> properties:
>     density: 6950.0
> ```
>
> ## Example 2:
> ```
> properties:
>     latent_heat_of_fusion:
>         dependency: solidus_temperature
>         value: [0, 171401]
> ```
>
> Again the name is missing and it complains about the dependency reference.
>
> I try to run both examples using:
>
> ```
> import sympy as sp
> from materforge.parsing.api import create_material
>
> solidus_temperature = sp.Symbol('T')
> mat = create_material('test.yaml', dependency=T, enable_plotting=True)
> ```
>
> Can you please fix the examples in the paper and provide sufficient documentation how the user can reproduce the results you show? In addition, can you include these examples in the repository that makes it easier to validate them.

Hi @jan-janssen ,

Thank you for testing these - the confusion here is legitimate and I apologize for it.

On figure 1: The YAML snippets in figure 1 are illustrative fragments showing the syntax for each property type, not complete runnable configurations. However I agree the caption is misleading - "corresponding YAML examples" implies they should work as-is. I will update the caption to make the illustrative nature explicit in the next paper revision.

On the int/float issue: You are correct that density: 6950 (integer) fails in v0.6.5. The restriction has been relaxed in master - both integer and float notation are now accepted. I will release v0.6.6 shortly so that the published PyPI version reflects this.

Working minimal example based on the above example for immediate use with v0.6.5:
test.yaml:
```
name: test_material
properties:
  density: 6950.0
  solidus_temperature: 1605.0
  latent_heat_of_fusion:
    dependency: solidus_temperature
    value: [0.0, 171401.0]
    bounds: [constant, constant]
```
```
import sympy as sp
from materforge.parsing.api import create_material
T = sp.Symbol('T')
mat = create_material('test.yaml', dependency=T, enable_plotting=False)
```

A similar file has already been added to [examples/](https://github.com/rahildoshi97/materforge/tree/master/examples) in the repository for reference along with the corresponding documentation for [Defining Custom Material Properties](https://materforge.readthedocs.io/en/stable/how-to/define_materials.html#complete-example).

The full `Usage` section in the paper and the [getting started](https://materforge.readthedocs.io/en/stable/tutorials/getting_started.html#create-simple-steel-yaml) tutorial contain additional complete examples.

Hope this helps :)
