### 2026-03-13 #9909 @jan-janssen
> On figure 1: The YAML snippets in figure 1 are illustrative fragments showing the syntax for each property type, not complete runnable configurations. However I agree the caption is misleading - "corresponding YAML examples" implies they should work as-is. I will update the caption to make the illustrative nature explicit in the next paper revision.

Maybe it is easier to have one larger example and then discuss the individual aspects, meaning you could still have the different plots just have one combined code block.
