### 2026-03-16 9ebbccd
Prepare release v0.6.6
