### 2026-03-16 1b1dbbd
Update JOSS paper with unified YAML example and new property plots
