+++ examples/myAlloy.yaml
@@ -14,42 +14,19 @@ name: myAlloy
 
 properties:
 
-  # ====================================================================================================
-  # PROPERTY TYPE 1: CONSTANT_VALUE
-  # Format: property_name: numeric_value
-  # ====================================================================================================
+  density: 6950
 
-  density1: 6950  # kg/m^3
-
-  # ====================================================================================================
-  # PROPERTY TYPE 2: STEP_FUNCTION
-  # Format: Uses 'dependency' (single reference) and 'value' (list of 2 values)
-  # Supports temperature references and arithmetic expressions
-  # ====================================================================================================
-
-  latent_heat_of_fusion1:
-    dependency: solidus_temp1 + 5
+  latent_heat_of_fusion:
+    dependency: density / 4.33
     value: [0, 171401]
     bounds: [constant, constant]
 
-  # ====================================================================================================
-  # PROPERTY TYPE 3: TABULAR_DATA
-  # Format: Uses 'dependency' (list/tuple) and 'value' (list) with same length
-  # Supports temperature references, explicit lists, and tuple notation
-  # ====================================================================================================
-
-  heat_conductivity1:
+  heat_conductivity:
     dependency: [500, 1000, 1600, 1700, 1750, 2000, 2500]
     value: [19.25, 25.47, 32.94, 33.52, 31.53, 35.33, 42.95]
     bounds: [linear, linear]
 
-  # ====================================================================================================
-  # PROPERTY TYPE 4: FILE_IMPORT
-  # Format: Uses 'file_path', 'dependency_column', 'property_column'
-  # Supports .txt, .csv, .xlsx files with automatic missing value handling
-  # ====================================================================================================
-
-  heat_capacity1:
+  heat_capacity:
     file_path: ./myAlloy.csv
     dependency_column: T (K)
     property_column: Specific heat (J/(Kg K))
@@ -59,45 +36,30 @@ properties:
       degree: 3
       segments: 6
 
-  # ====================================================================================================
-  # PROPERTY TYPE 5: PIECEWISE_EQUATION
-  # Format: Uses 'dependency' (breakpoints) and 'equation' (list of expressions)
-  # Number of equations = number of breakpoints - 1
-  # ====================================================================================================
-
-  viscosity1:
+  viscosity:
     dependency: [300, 1660, 1736, 3000]
     equation: [7877.39-0.37*T, 11816.63-2.74*T, 8596.40-0.88*T]
     bounds: [constant, constant]
 
-  # ====================================================================================================
-  # PROPERTY TYPE 6: COMPUTED_PROPERTY
-  # Format: Uses 'equation' (single expression) and 'temperature' (range definition)
-  # Supports dependency resolution and automatic ordering
-  # ====================================================================================================
-
-  thermal_diffusivity1:
+  thermal_diffusivity:
     dependency: (3000, 300, -5.0)
-    equation: heat_conductivity1 /(density1 * heat_capacity1)
-    bounds: [linear, linear]
+    equation: heat_conductivity /(density * heat_capacity)
+    bounds: [constant, linear]
     regression:
       simplify: post
       degree: 3
       segments: 7
 
-  specific_enthalpy1:
+  specific_enthalpy:
     dependency: (3000, 300, 541)
-    equation: Integral(heat_capacity1, T)
+    equation: Integral(heat_capacity, T)
     bounds: [ constant, constant ]
     regression:
       simplify: post
       degree: 2
       segments: 4
 
-  solidus_temp1: 1605          # Temperature where melting begins (K)
-  liquidus_temp1: 1735         # Temperature where material is completely melted (K)
-  initial_boiling_temp1: 3090  # Temperature where boiling begins (K)
[…]
