+++ paper/paper.md
@@ -30,115 +30,75 @@ bibliography: paper.bib
 
 # Summary
 
-MaterForge is an extensible, open-source Python library that streamlines the definition and use of
-material properties in numerical simulations.
-The library supports complex material behaviors, from simple constants to experimental data,
-in user-friendly YAML configurations.
+MaterForge is an extensible, open-source Python library that streamlines the definition and use of material properties in numerical simulations.
+The library supports complex material behaviors, from simple constants to experimental data, in user-friendly YAML configurations.
 These are internally converted into symbolic mathematical expressions for scientific computing frameworks.
-MaterForge supports any material type through a schema-agnostic design,
-provides flexible property definitions,
-and automatically resolves dependency order for derived properties while detecting cycles.
-It is designed for high-performance computing (HPC) applications
-and serves as a bridge between experimental data and numerical simulation.
+MaterForge supports any material type through a schema-agnostic design, provides flexible property definitions, and automatically resolves dependency order for derived properties while detecting cycles.
+It is designed for high-performance computing (HPC) applications and serves as a bridge between experimental data and numerical simulation.
 
 # Statement of Need
 
-Accurate numerical simulations of physical processes rely on well-characterized material properties -
-quantities such as thermal conductivity, density, and viscosity that depend on state variables like temperature, pressure, or strain rate [@lewis1996finite].
-This challenge is compounded by the wide variation in data availability,
-from well-characterized models for established materials to sparse experimental points for novel materials.
-Property definitions consequently range from simple constants to complex tabular datasets or sophisticated equations,
-creating significant integration hurdles for researchers.
+Accurate numerical simulations of physical processes rely on well-characterized material properties - quantities such as thermal conductivity, density, and viscosity that depend on state variables like temperature, pressure, or strain rate [@lewis1996finite].
+This challenge is compounded by the wide variation in data availability, from well-characterized models for established materials to sparse experimental points for novel materials.
+Property definitions consequently range from simple constants to complex tabular datasets or sophisticated equations, creating significant integration hurdles for researchers.
 
-To manage this complexity, researchers often resort to manual interpolation, custom scripting, or proprietary software,
-which compromises reproducibility and standardization [@ashby2013materials].
-While valuable resources like the NIST WebBook [@linstrom2001nist] and CoolProp [@coolprop] provide raw data,
-they lack integrated processing to unify these varied formats into simulation-ready symbolic expressions.
-Thermodynamic modeling tools such as pycalphad [@pycalphad] and CALPHAD databases [@calphad]
-are powerful for phase equilibria calculations but operate at a different layer of the workflow:
-they generate property data, not simulation-ready expressions.
+To manage this complexity, researchers often resort to manual interpolation, custom scripting, or proprietary software, which compromises reproducibility and standardization [@ashby2013materials].
+While valuable resources like the NIST WebBook [@linstrom2001nist] and CoolProp [@coolprop] provide raw data, they lack integrated processing to unify these varied formats into simulation-ready symbolic expressions.
+Thermodynamic modeling tools such as pycalphad [@pycalphad] and CALPHAD databases [@calphad] are powerful for phase equilibria calculations but operate at a different layer of the workflow: they generate property data, not simulation-ready expressions.
 
-This creates a gap between property data generation and simulation integration,
-leading to ad hoc solutions that hinder workflow efficiency and FAIR data adoption [@wilkinson2016fair].
[…]
