### 2026-03-16 19707ca
Update myAlloy_properties.png
