### 2026-03-16 28c35d4
Update paper.md
