### 2026-03-17 40b5b4a
Change x-axis label for figure 2 subplots, update paper.md
