+++ paper/paper.md
@@ -108,41 +108,41 @@ name: myAlloy
 
 properties:
 
-  density: 6950  # Fig. 2a
+    density: 6950  # Fig. 2a
 
-  latent_heat_of_fusion:  # Fig. 2b
-    dependency: density / 4.33
-    value: [0, 171401]
-    bounds: [constant, constant]
+    latent_heat_of_fusion:  # Fig. 2b
+        dependency: density / 4.33
+        value: [0, 171401]
+        bounds: [constant, constant]
 
-  heat_conductivity:  # Fig. 2c
-    dependency: [500, 1000, 1600, 1700, 1750, 2000, 2500]
-    value: [19.25, 25.47, 32.94, 33.52, 31.53, 35.33, 42.95]
-    bounds: [linear, linear]
+    heat_conductivity:  # Fig. 2c
+        dependency: [500, 1000, 1600, 1700, 1750, 2000, 2500]
+        value: [19.25, 25.47, 32.94, 33.52, 31.53, 35.33, 42.95]
+        bounds: [linear, linear]
 
-  heat_capacity:  # Fig. 2d
-    file_path: ./myAlloy.csv
-    dependency_column: T (K)
-    property_column: Specific heat (J/(Kg K))
-    bounds: [constant, constant]
-    regression:
-      simplify: pre
-      degree: 3
-      segments: 6
+    heat_capacity:  # Fig. 2d
+        file_path: ./myAlloy.csv
+        dependency_column: T (K)
+        property_column: Specific heat (J/(Kg K))
+        bounds: [constant, constant]
+        regression:
+            simplify: pre
+            degree: 3
+            segments: 6
 
-  viscosity:  # Fig. 2e
-    dependency: [300, 1660, 1736, 3000]
-    equation: [7877.39-0.37*T, 11816.63-2.74*T, 8596.40-0.88*T]
-    bounds: [constant, constant]
+    viscosity:  # Fig. 2e
+        dependency: [300, 1660, 1736, 3000]
+        equation: [7877.39-0.37*T, 11816.63-2.74*T, 8596.40-0.88*T]
+        bounds: [constant, constant]
 
-  thermal_diffusivity:  # Fig. 2f
-    dependency: (3000, 300, -5)
-    equation: heat_conductivity / (density * heat_capacity)
-    bounds: [constant, linear]
-    regression:
-      simplify: post
-      degree: 3
-      segments: 7
+    thermal_diffusivity:  # Fig. 2f
+        dependency: (3000, 300, -5.0)
+        equation: heat_conductivity / (density * heat_capacity)
+        bounds: [constant, linear]
+        regression:
+            simplify: post
+            degree: 3
+            segments: 7
 ```
 
 ## Python Integration
