### 2026-03-17 #9909 @rahildoshi97
Thanks for the suggestion. I have made the necessary changes in the manuscript. The new v0.6.6 on pypi now also supports integers along with floats in the YAML config :)
