### 2026-03-17 #9909 @jan-janssen
Looks good to me, just one quick question from a users perspective, would it be possible add `create_material()` to the `__init__.py` ? This would shorten the import for the users. Rather than having to type:
```python
from materforge.parsing.api import create_material
```
The user could:
```python
from materforge import create_material
```
From my perspective this would simplify the user interface. Or is there any specific reason why you do not want the import to live in the root of the package?
