### 2026-03-17 146a9d0
Expose evaluation helpers in top-level API
