### 2026-03-17 ea5fca2
Refactor to use top-level materforge API for create_material
