+++ tests/unit/test_algorithms/test_inverse_1.py
@@ -4,7 +4,7 @@ from typing import List, Tuple
 import logging
 from pathlib import Path
 
-from materforge.parsing.api import create_material
+from materforge import create_material
 
 logger = logging.getLogger(__name__)
 
