### 2026-03-17 #9909 @rahildoshi97
> Looks good to me, just one quick question from a users perspective, would it be possible add `create_material()` to the `__init__.py` ? This would shorten the import for the users. Rather than having to type:
>
> from materforge.parsing.api import create_material
> The user could:
>
> from materforge import create_material
> From my perspective this would simplify the user interface. Or is there any specific reason why you do not want the import to live in the root of the package?

Thanks a lot for the suggestion. I agree this is nicer from a user’s perspective.

`create_material` was already exposed at the package root via `materforge/__init__.py`, but all examples and internal code were still importing it from `materforge.parsing.api`, so users didn’t really see the flat API. I’ve now:

- kept `create_material` exported at the top level and
- updated the examples and internal imports to use

```python
from materforge import create_material
```

as the primary entry point. So users can now consistently use the short form you suggested.
