### 2026-03-17 7455323
Use top-level materforge API for create_material
