### 2026-03-17 2bda9f5
Update YAML config in paper.md
