### 2026-03-17 e4dec2a
Update YAML config in paper.md
