### 2026-03-17 4b22a22
Update YAML config in paper.md
