### 2026-03-17 e5684b8
Split YAML config in paper.md
