### 2026-03-17 2a91e5e
Update YAML config in paper.md
