### 2026-03-17 8d769fc
Update YAML config in paper.md
