### 2026-03-17 e9f0b79
Update YAML config in paper.md
