### 2026-03-17 b6e05ed
Update paper.md
