+++ paper/paper.md
@@ -17,7 +17,7 @@ authors:
   - name: Matthias Markl
     orcid: 0000-0002-3000-7081
     affiliation: 2
-affiliations: 
+affiliations:
   - name: Chair for System Simulation, Friedrich-Alexander-Universität Erlangen-Nürnberg, Germany
     index: 1
   - name: Chair of Materials Science and Engineering for Metals, Friedrich-Alexander-Universität Erlangen-Nürnberg, Germany
@@ -28,7 +28,9 @@ date: 17 March 2026
 bibliography: paper.bib
 ---
 
+
 # Summary
+
 
 MaterForge is an extensible, open-source Python library that streamlines the definition and use of material properties in numerical simulations.
 The library supports complex material behaviors, from simple constants to experimental data, in user-friendly YAML configurations.
@@ -36,63 +38,85 @@ These are internally converted into symbolic mathematical expressions for scient
 MaterForge supports any material type through a schema-agnostic design, provides flexible property definitions, and automatically resolves dependency order for derived properties while detecting cycles.
 It is designed for high-performance computing (HPC) applications and serves as a bridge between experimental data and numerical simulation.
 
+
 # Statement of Need
+
 
 Accurate numerical simulations of physical processes rely on well-characterized material properties - quantities such as thermal conductivity, density, and viscosity that depend on state variables like temperature, pressure, or strain rate [@lewis1996finite].
 This challenge is compounded by the wide variation in data availability, from well-characterized models for established materials to sparse experimental points for novel materials.
 Property definitions consequently range from simple constants to complex tabular datasets or sophisticated equations, creating significant integration hurdles for researchers.
 
+
 To manage this complexity, researchers often resort to manual interpolation, custom scripting, or proprietary software, which compromises reproducibility and standardization [@ashby2013materials].
 While valuable resources like the NIST WebBook [@linstrom2001nist] and CoolProp [@coolprop] provide raw data, they lack integrated processing to unify these varied formats into simulation-ready symbolic expressions.
 Thermodynamic modeling tools such as pycalphad [@pycalphad] and CALPHAD databases [@calphad] are powerful for phase equilibria calculations but operate at a different layer of the workflow: they generate property data, not simulation-ready expressions.
 
+
 This creates a gap between property data generation and simulation integration, leading to ad hoc solutions that hinder workflow efficiency and FAIR data adoption [@wilkinson2016fair].
 MaterForge fills this gap by occupying a dedicated post-processing layer in the materials simulation workflow, complementing rather than competing with thermodynamic modeling tools: pycalphad generates the data; MaterForge prepares it for simulation.
 
+
 # Simulation Workflow
+
 
 MaterForge is designed to operate as the intermediate layer in a three-stage workflow:
 
+
 **Stage 1 - Property Data Generation**: Raw material property data is produced by upstream tools such as pycalphad [@pycalphad] for thermodynamic phase equilibria, CoolProp [@coolprop] or the NIST WebBook [@linstrom2001nist] for fluid properties, or commercial tools such as JMatPro and Thermo-Calc for alloy properties, or direct experimental measurement.
 These tools generate tabular data, equations, or database entries.
+
 
 **Stage 2 - MaterForge Post-Processing**: MaterForge ingests this data via YAML configuration files and converts it into optimized symbolic mathematical expressions using SymPy [@sympy].
 It performs automatic regression and data reduction, resolves inter-property dependencies, validates configurations, and generates visualization plots for verification.
[…]
