+++ paper/paper.bib
@@ -104,7 +104,7 @@
 }
 
 @article{voller1987fixed,
-    title = {A fixed grid numerical modelling methodology for convection-diffusion mushy region phase-change problems},
+    title = {A fixed grid numerical modeling methodology for convection-diffusion mushy region phase-change problems},
     author = {Voller, Vaughan R. and Prakash, C.},
     journal = {International Journal of Heat and Mass Transfer},
     volume = 30,
